### 2024-04-22 #6658 @editorialbot
[PRE REVIEW]: Singularity-EOS: Performance Portable Equations of State and Mixed Cell Closures
**Submitting author:** @Yurlungur (<a href="http://orcid.org/0000-0001-6432-7860">Jonah Miller</a>)
**Repository:** https://github.com/lanl/singularity-eos
**Branch with paper.md** (empty if default branch):
**Version:** v1.8.0
