+++ singularity-eos/base/variadic_utils.hpp
@@ -24,17 +24,26 @@ namespace variadic_utils {
 // Some generic variatic utilities
 // ======================================================================
 
+// C++14 implementation of std::remove_cvref (available since C++20)
+// credit to CJ + Diego
+template <typename T>
+struct remove_cvref {
+  typedef std::remove_cv_t<std::remove_reference_t<T>> type;
+};
+
+// Helper types
+template <typename T>
+using remove_cvref_t = typename remove_cvref<T>::type;
+
 // SFINAE to check if a value is a null ptr
 template <typename T, typename = typename std::enable_if<
-                          std::is_pointer<typename std::remove_reference<
-                              typename std::remove_cv<T>::type>::type>::value>::type>
-constexpr inline bool is_nullptr(T &&t) {
+                          std::is_pointer<remove_cvref_t<T>>::value>::type>
+inline bool is_nullptr(T &&t) {
   return std::forward<T>(t) == nullptr;
 }
-template <typename T,
-          typename std::enable_if<!std::is_pointer<typename std::remove_reference<
-              typename std::remove_cv<T>::type>::type>::value>::type * = nullptr>
-constexpr inline bool is_nullptr(T &&) {
+template <typename T, typename std::enable_if<
+                          !std::is_pointer<remove_cvref_t<T>>::value>::type * = nullptr>
+inline bool is_nullptr(T &&) {
   return false;
 }
 
