### 2024-04-23 #6658 @Yurlungur
> Hi @Yurlungur, it looks like the paper compilation error is coming from the dash before `orcid` field in Daniel Hollday's author block.

Oops. Hopefully fixed now. Thanks @kyleniemeyer !
