### 2024-04-23 c75b6ca
Missing index for affiliation 11
