+++ joss-paper/paper.md
@@ -68,6 +68,7 @@ affiliations:
   - name: Department of Physics, Florida State University, USA
     index: 10
   - name: XCP-4, Continuum Models and Numerical Methods, Los Alamos National Laboratory, USA
+    index: 11
 date: 13 April 2024
 bibliography: refs.bib
 
