+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -566,7 +566,7 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     vtemp = AssignIncrement(scratch, nmat);
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
-      if (variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
+      if (!variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
     }*/
   }
 
