### 2024-04-23 88f616b
static casts for everybody
