### 2024-04-23 921d1c0
one more static cast for those at home
