+++ joss-paper/paper.md
@@ -32,6 +32,7 @@ authors:
     orcid: 0000-0002-8825-0893
     affiliation: "1, 2, 6, 7"
   - name: Alexander Holas
+    orcid: 0000-0001-5184-6928
     affiliation: "2, 3, 8"
   - name: Ann E. Mattsson
     affiliation: 9
@@ -51,7 +52,7 @@ affiliations:
     index: 1
   - name: Center for Theoretical Astrophysics, Los Alamos National Laboratory, Los Alamos, NM
     index: 2
-  - name: CCS-7, Applied Computer Scienc, Los Alamos National Laboratory, USA
+  - name: CCS-7, Applied Computer Science, Los Alamos National Laboratory, USA
     index: 3
   - name: XCP-2, Eulerian Codes, Los Alamos National Laboratory, USA
     index: 4
