### 2024-04-24 234d258
convert everything up to jwl and ideal
