+++ singularity-eos/eos/eos_helmholtz.hpp
@@ -45,6 +45,7 @@
 #include <singularity-eos/base/math_utils.hpp>
 #include <singularity-eos/base/robust_utils.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
+#include <singularity-eos/base/variadic_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
 // spiner
@@ -149,6 +150,8 @@
 
 namespace singularity {
 using namespace eos_base;
+using variadic_utils::ptrt_nullptr;
+using variadic_utils::DoIfNotNull;
 
 // TODO(JMM): Maybe want to move these utility functions into something like an
 // ASCII-utils file. Worth considering at some later date.
@@ -463,7 +466,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TemperatureFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+      Indexer_t &&lambda = ptrt_nullptr()) const {
     Real rl = rho;
     Real el = sie;
     Real temperature, p, cv, bmod;
@@ -473,7 +476,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
       const Real rho, const Real temperature,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+      Indexer_t &&lambda = ptrt_nullptr()) const {
     Real rl = rho;
     Real tl = temperature;
     Real sie, p, cv, bmod;
@@ -483,7 +486,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PressureFromDensityTemperature(
       const Real rho, const Real temperature,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+      Indexer_t &&lambda = ptrt_nullptr()) const {
     Real rl = rho;
     Real tl = temperature;
     Real sie, p, cv, bmod;
@@ -493,7 +496,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+      Indexer_t &&lambda = ptrt_nullptr()) const {
     Real rl = rho;
     Real el = sie;
     Real temperature, p, cv, bmod;
@@ -503,7 +506,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   }
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real MinInternalEnergyFromDensity(
-      const Real rho, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+      const Real rho, Indexer_t &&lambda = ptrt_nullptr()) const {
     MinInternalEnergyIsNotEnabled("Helmholtz");
     return 0.0;
   }
@@ -511,7 +514,7 @@ class Helmholtz : public EosBase<Helmholtz> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real
   EntropyFromDensityTemperature(const Real rho, const Real temperature,
[…]
