+++ singularity-eos/eos/eos_jwl.hpp
@@ -28,11 +28,14 @@
 #include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/eos_error.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
+#include <singularity-eos/base/variadic_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
 namespace singularity {
 
 using namespace eos_base;
+using variadic_utils::ptrt_nullptr;
+using variadic_utils::DoIfNotNull;
 
 // COMMENT: This is meant to be an implementation of the "standard" JWL as
 // implemented in xRAGE for eostype(1).  It does not include any energy shifting
@@ -53,59 +56,59 @@ class JWL : public EosBase<JWL> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TemperatureFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
       const Real rho, const Real temperature,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real
   PressureFromDensityTemperature(const Real rho, const Real temperature,
-                                 Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+                                 Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real MinInternalEnergyFromDensity(
-      const Real rho, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      const Real rho, Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real
   EntropyFromDensityTemperature(const Real rho, const Real temperature,
-                                Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+                                Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real EntropyFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityTemperature(
       const Real rho, const Real temperature,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
+      Indexer_t &&lambda = ptrt_nullptr()) const;
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
       const Real rho, const Real sie,
-      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const;
[…]
