### 2024-04-24 1eeed8a
add constexpr
