### 2024-04-25 32bcf9c
use array for host test of lambda, just to make the compiler stop complaining
