### 2024-04-25 f594ee1
add constexpr function for generating nullptr of specific pointer type
