### 2024-04-25 c158ff3
First clean up for better structure of tests.
