+++ test/pte_test_2phaseVinetSn.hpp
@@ -21,6 +21,10 @@
 #include <ports-of-call/portable_arrays.hpp>
 #include <singularity-eos/eos/eos.hpp>
 
+// This data is taken from the .ult file for a test case in Flag, MPMat_KPT_PTsolv.ult.std
+// Note that at this point in time (April 2024) the global and phase times are not alinged
+// in .ult but they will be in the near future.
+
 namespace pte_test_2phaseVinetSn {
 
 constexpr int NMAT = 2;
@@ -29,20 +33,24 @@ constexpr int NPTS = NTRIAL * NMAT;
 constexpr int HIST_SIZE = 10;
 
 constexpr Real in_rho_tot[NTRIAL] = {7.27800000, 7.27981950, 7.28163945, 7.28345986,
-                                7.28528073};
+                                     7.28528073};
 constexpr Real in_sie_tot[NTRIAL] = {8.41323509e08, 8.41325565e08, 8.41331734e08,
-                                8.41342022e08, 8.41356432e08};
-constexpr Real in_lambda[NMAT] = {0.500480901,0.499519099};
-constexpr Real trial_vfrac[NMAT] = {47.0 / 100.0, 53.0/100.0};
+                                     8.41342022e08, 8.41356432e08};
+constexpr Real in_lambda[NMAT] = {0.500480901, 0.499519099};
+constexpr Real trial_vfrac[NMAT] = {47.0 / 100.0, 53.0 / 100.0};
 
 constexpr Real out_press[NTRIAL] = {-3.29164604e-6, 1.19722694e8, 2.3968114450000003e8,
-                               3.598077865e8, 4.80104422e8};
+                                    3.598077865e8, 4.80104422e8};
 constexpr Real out_temp[NTRIAL] = {298., 298.192952, 298.38594450000005, 298.5790125,
-                              298.7721535};
-constexpr Real out_rho0[NTRIAL] = {7.28500000, 7.28653963, 7.28808705, 7.28963516, 7.29118416};
-constexpr Real out_rho1[NTRIAL] = {7.27100000, 7.27309885, 7.27519088, 7.27728316, 7.27937551};
-constexpr Real out_vfrac0[NTRIAL] = {0.5, 0.500019325, 0.500038138, 0.500056927, 0.500075679};
-constexpr Real out_vfrac1[NTRIAL] = {0.5, 0.499980675, 0.499961862, 0.499943073, 0.499924321};
+                                   298.7721535};
+constexpr Real out_rho0[NTRIAL] = {7.28500000, 7.28653963, 7.28808705, 7.28963516,
+                                   7.29118416};
+constexpr Real out_rho1[NTRIAL] = {7.27100000, 7.27309885, 7.27519088, 7.27728316,
+                                   7.27937551};
+constexpr Real out_vfrac0[NTRIAL] = {0.5, 0.500019325, 0.500038138, 0.500056927,
+                                     0.500075679};
+constexpr Real out_vfrac1[NTRIAL] = {0.5, 0.499980675, 0.499961862, 0.499943073,
+                                     0.499924321};
 
 template <typename T>
 inline void set_eos(T *eos) {
@@ -62,14 +70,14 @@ inline void set_eos(T *eos) {
 }
 
 template <typename RealIndexer, typename EOSIndexer>
-inline void set_trial_state(int n, RealIndexer &&rho, RealIndexer &&vfrac, RealIndexer &&sie,
-                      EOSIndexer &&eos) {
+inline void set_trial_state(int n, RealIndexer &&rho, RealIndexer &&vfrac,
+                            RealIndexer &&sie, EOSIndexer &&eos) {
 
[…]

[…]
