### 2024-05-01 3fbafd9
Fix bmod calculation in spiner EOS
