### 2024-05-03 e4e046e
Specify fixed var in therm derivs
