### 2024-05-10 fa9d603
Allow E0 to adjust isentrope energy
