### 2024-05-10 96d71fc
Add place-holder for CJ state calculation
