### 2024-05-10 ed042bd
Add positive temperature check for debug builds
