### 2024-05-10 8c0efef
Revert Davis EOS changes
