### 2024-05-10 4a8e824
Change to non-negative temperature check rather than strictly positive
