### 2024-05-10 733d12f
Remove Davis EOS change
