### 2024-05-11 83a658b
Add robustness to calculations
