+++ singularity-eos/eos/eos_davis.hpp
@@ -18,6 +18,7 @@
 #include <cmath>
 #include <cstdio>
 
+#include <ports-of-call/portable_errors.hpp>
 #include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/math_utils.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
@@ -42,16 +43,21 @@ class DavisReactants : public EosBase<DavisReactants> {
       const Real rho, const Real sie,
       Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     const Real es = Es(rho);
-    const Real tmp = std::pow((1.0 + _alpha) / (Ts(rho) * _Cv0) * (sie - es) + 1.0,
-                              1.0 / (1.0 + _alpha));
-    if (tmp > 0) return Ts(rho) * tmp;
-    return Ts(rho) + (sie - es) / _Cv0; // This branch is a negative temperature
+    const Real power_base = (1.0 + _alpha) / (Ts(rho) * _Cv0) * (sie - es) + 1.0;
+    if (power_base <= 0) {
+      // This case would result in an imaginary temperature (i.e. negative), but we won't
+      // allow that so return zero
+      return 0.;
+    }
+    const Real tmp = std::pow(power_base, 1.0 / (1.0 + _alpha));
+    return Ts(rho) * tmp;
   }
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
       const Real rho, const Real temp,
       Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     const Real t_s = Ts(rho);
+    PORTABLE_REQUIRE(temp >= 0, "Negative temperature provided");
     return Es(rho) +
            _Cv0 * t_s / (1.0 + _alpha) * (std::pow(temp / t_s, 1.0 + _alpha) - 1.0);
   }
@@ -72,8 +78,11 @@ class DavisReactants : public EosBase<DavisReactants> {
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real MinInternalEnergyFromDensity(
       const Real rho, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
-    MinInternalEnergyIsNotEnabled("DavisReactants");
-    return 0.0;
+    // Minimum enegy is when the returned temperature is zero. This only happens
+    // when the base to the exponent is zero (see T(rho, e) equation)
+    const Real es = Es(rho);
+    const Real ts = Ts(rho);
+    return es - (1 + _alpha) / (_Cv0 * ts);
   }
   template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real
@@ -100,8 +109,12 @@ class DavisReactants : public EosBase<DavisReactants> {
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityInternalEnergy(
       const Real rho, const Real sie,
       Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
-    return _Cv0 / std::pow((1 + _alpha) / (Ts(rho) * _Cv0) * (sie - Es(rho)) + 1,
-                           -_alpha / (1 + _alpha));
+    const Real power_base = (1 + _alpha) / (Ts(rho) * _Cv0) * (sie - Es(rho)) + 1;
+    if (power_base <= 0) {
+      // Return zero heat capacity instead of an imaginary value
+      return 0.;
+    }
[…]
