### 2024-05-11 355cf71
Create default MinInternalEnergyFromDensity behavior
