+++ singularity-eos/eos/eos_base.hpp
@@ -642,7 +642,16 @@ class EosBase {
     PORTABLE_ALWAYS_THROW_OR_ABORT(msg);
   }
 
-  // Default MinInternalEnergyFromDensity behavior is to cause an error
+  // Default MinInternalEnergyFromDensity behavior is to just return zero. This is really
+  // only appropriate for ideal-gas like materials where the reference state is at zero
+  // density
+  PORTABLE_FORCEINLINE_FUNCTION
+  Real MinInternalEnergyFromDensity(const Real rho) const {
+    return 0.;
+  }
+
+  // For EOS where the minimum energy is not trivially zero and where the real function
+  // hasn't been implemented, an error should be thrown
   PORTABLE_FORCEINLINE_FUNCTION
   void MinInternalEnergyIsNotEnabled(const char *eosname) const {
     // Construct the error message using char* so it works on device
