### 2024-05-11 2eda134
Add minimum energy modifier
