+++ singularity-eos/eos/modifiers/minimum_energy.hpp
@@ -1,0 +1,379 @@
+//------------------------------------------------------------------------------
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
+// program was produced under U.S. Government contract 89233218CNA000001
+// for Los Alamos National Laboratory (LANL), which is operated by Triad
+// National Security, LLC for the U.S.  Department of Energy/National
+// Nuclear Security Administration. All rights in the program are
+// reserved by Triad National Security, LLC, and the U.S. Department of
+// Energy/National Nuclear Security Administration. The Government is
+// granted for itself and others acting on its behalf a nonexclusive,
+// paid-up, irrevocable worldwide license in this material to reproduce,
+// prepare derivative works, distribute copies to the public, perform
+// publicly and display publicly, and to permit others to do so.
+//------------------------------------------------------------------------------
+
+#ifndef _SINGULARITY_EOS_EOS_MINIMUM_ENERGY_
+#define _SINGULARITY_EOS_EOS_MINIMUM_ENERGY_
+
+#include "stdio.h"
+#include <cstdlib>
+#include <iostream>
+
+#include <ports-of-call/portability.hpp>
+#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/eos/eos_base.hpp>
+
+namespace singularity {
+
+using namespace eos_base;
+
+template <typename T>
+class MinimumEnergy : public EosBase<MinimumEnergy<T>> {
+ public:
+  // Generic functions provided by the base class. These contain
+  // e.g. the vector overloads that use the scalar versions declared
+  // here We explicitly list, rather than using the macro because we
+  // overload some methods.
+
+  // TODO(JMM): The modifier EOS's should probably call the specific
+  // sub-functions of the class they modify so that they can leverage,
+  // e.g., an especially performant or special version of these
+  using EosBase<MinimumEnergy<T>>::TemperatureFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::InternalEnergyFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::PressureFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::PressureFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::MinInternalEnergyFromDensity;
+  using EosBase<MinimumEnergy<T>>::EntropyFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::EntropyFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::SpecificHeatFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::SpecificHeatFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::BulkModulusFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::BulkModulusFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::GruneisenParamFromDensityTemperature;
+  using EosBase<MinimumEnergy<T>>::GruneisenParamFromDensityInternalEnergy;
+  using EosBase<MinimumEnergy<T>>::FillEos;
+
+  using BaseType = T;
+
+  // give me std::format or fmt::format...
+  static std::string EosType() {
+    return std::string("MinimumEnergy<") + T::EosType() + std::string(">");
+  }
+
+  static std::string EosPyType() { return std::string("MinimumEnergy") + T::EosPyType(); }
+
+  MinimumEnergy() = default;
+
+  auto GetOnDevice() { return MinimumEnergy<T>(t_.GetOnDevice()); }
+  inline void Finalize() { t_.Finalize(); }
+
+  template <typename Indexer_t = Real *>
+  PORTABLE_FUNCTION Real TemperatureFromDensityInternalEnergy(
[…]
