### 2024-05-11 1c29d51
Add minimum energy modifier
