### 2024-05-11 c6ed8dc
Rely on default minimum energy functionality in base class (if this works)
