### 2024-05-11 99cfb4b
Reference shift modifier for Davis Products EOS
