### 2024-05-11 8370048
Remove E0 from input
