### 2024-05-11 0032e67
Remove E0 from Davis Products EOS
