+++ singularity-eos/eos/singularity_eos.cpp
@@ -154,8 +154,8 @@ int init_sg_JWL(const int matindex, EOS *eos, const double A, const double B,
 
 int init_sg_DavisProducts(const int matindex, EOS *eos, const double a, const double b,
                           const double k, const double n, const double vc,
-                          const double pc, const double Cv,
-                          int const *const enabled, double *const vals) {
+                          const double pc, const double Cv, int const *const enabled,
+                          double *const vals) {
   assert(matindex >= 0);
   EOS eosi = SGAPPLYMODSIMPLE(DavisProducts(a, b, k, n, vc, pc, Cv));
   if (enabled[3] == 1) {
