### 2024-05-11 c519d29
Remove E0 from print parameters function
