### 2024-05-13 bbf23e1
Update changelong to reflect that Davis Products E0 is now gone instead of 'fixed'
