### 2024-05-13 c265ca4
Remove E0 from Davis Products EOS
