### 2024-05-13 a6f09ba
Remove Davis Products EOS E0 parameter
