### 2024-05-14 768f629
Finally remove E0 from DavisProducts EOS pybinding init
