+++ python/module.cpp
@@ -60,7 +60,7 @@ PYBIND11_MODULE(singularity_eos, m) {
   eos_class<DavisProducts>(m, "DavisProducts")
     .def(py::init())
     .def(
-      py::init<Real, Real, Real, Real, Real, Real, Real, Real>(),
+      py::init<Real, Real, Real, Real, Real, Real, Real>(),
       py::arg("a"), py::arg("b"), py::arg("k"), py::arg("n"), py::arg("vc"),
       py::arg("pc"), py::arg("Cv")
     );
