### 2024-05-14 61edd15
Fix missing maxes and remove last negative density branch
