### 2024-05-14 4ec06fe
Add missing maxes and do a clang format
