### 2024-05-14 4e2c4b4
Break out dimensionless energy into its own function
