### 2024-05-14 43fb3ad
Correct real to Real and switch to error instead of maxing density
