### 2024-05-14 e30a855
Add missing density to DimlessEDiff function
