+++ singularity-eos/eos/eos_davis.hpp
@@ -354,7 +354,7 @@ class DavisProducts : public EosBase<DavisProducts> {
 
 PORTABLE_FORCEINLINE_FUNCTION Real DavisReactants::DimlessEdiff(const Real rho,
                                                                 const Real sie) const {
-    return (1.0 + _alpha) / (Ts(rho) * _Cv0) * (sie - Es(rho)) + 1.0;
+  return (1.0 + _alpha) / (Ts(rho) * _Cv0) * (sie - Es(rho)) + 1.0;
 }
 
 PORTABLE_INLINE_FUNCTION Real DavisReactants::Ps(const Real rho) const {
