+++ CMakeLists.txt
@@ -420,14 +420,7 @@ if(SINGULARITY_BUILD_CLOSURE)
 
   if(SINGULARITY_USE_KOKKOS)
     if(Kokkos_ENABLE_COMPILE_AS_CMAKE_LANGUAGE)
-      foreach(arg ${eos_srcs})
-        get_filename_component(FILE_EXT ${arg} LAST_EXT)
-        string(REPLACE "." "" ext ${FILE_EXT})
-        if(ext IN_LIST CMAKE_CXX_SOURCE_FILE_EXTENSIONS)
-          message(STATUS "Setting ${arg}")
-          set_source_files_properties(${arg} PROPERTIES LANGUAGE CUDA)
-        endif()
-      endforeach()
+      set_cpp_files_list_to_lang(eos_srcs Kokkos_COMPILE_LANGUGE)
     endif()
   endif()
 endif()
