### 2024-05-16 f7c5c43
Fixup cmake/singularity_utils.cmake
