### 2024-05-16 #6658 @kyleniemeyer
Hello @parikshitbajpai @corykinney @fwitte would any of you be able to review this submission for JOSS? It describes a library, Singularity-EOS, for equations of state.
