### 2024-05-16 #6658 @parikshitbajpai
> Hello @parikshitbajpai @corykinney @fwitte would any of you be able to review this submission for JOSS? It describes a library, Singularity-EOS, for equations of state.

Hi @kyleniemeyer, I can review this but I'm quite busy until May 30 so I won't be able to do it before then. I'll be open to review if it can wait till then.
