### 2024-05-17 #6658 @fwitte
Hi and thanks for the invitation. Unfortunately I am not experienced with C++, and more on the side of applying EOS for modeling instead of writing them myself. Therefore I would like to pass on this.

Other ideas for reviewers could be people from the CoolProp ecosystem (ibell or joker for example).
