### 2024-05-17 #6658 @kyleniemeyer
@esclapez would you be available to review this submission to JOSS? It describes a library, Singularity-EOS, for equations of state.
