### 2024-05-23 a6c80a0
Separate out potentially useful type lists and variant utility functions to separate header files that could be used downstream w/o needing the variant definition itself.
