+++ singularity-eos/eos/eos_type_lists.hpp
@@ -1,0 +1,130 @@
+//------------------------------------------------------------------------------
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
+// program was produced under U.S. Government contract 89233218CNA000001
+// for Los Alamos National Laboratory (LANL), which is operated by Triad
+// National Security, LLC for the U.S.  Department of Energy/National
+// Nuclear Security Administration. All rights in the program are
+// reserved by Triad National Security, LLC, and the U.S. Department of
+// Energy/National Nuclear Security Administration. The Government is
+// granted for itself and others acting on its behalf a nonexclusive,
+// paid-up, irrevocable worldwide license in this material to reproduce,
+// prepare derivative works, distribute copies to the public, perform
+// publicly and display publicly, and to permit others to do so.
+//------------------------------------------------------------------------------
+
+#ifndef _SINGULARITY_EOS_EOS_EOS_TYPE_LISTS_HPP_
+#define _SINGULARITY_EOS_EOS_EOS_TYPE_LISTS_HPP_
+
+#include <cassert>
+#include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#include <iostream>
+#include <limits>
+#include <utility>
+
+#include <ports-of-call/portability.hpp>
+#include <singularity-eos/eos/eos_base.hpp>
+#include <singularity-eos/eos/eos_variant.hpp>
+
+// Base stuff
+#include <singularity-eos/base/constants.hpp>
+#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/variadic_utils.hpp>
+
+// EOS models
+#include <singularity-eos/eos/eos_models.hpp>
+
+namespace singularity {
+
+// recreate variadic list
+template <typename... Ts>
+using tl = singularity::variadic_utils::type_list<Ts...>;
+
+template <template <typename> class... Ts>
+using al = singularity::variadic_utils::adapt_list<Ts...>;
+
+// transform variadic list: applies modifiers to eos's
+using singularity::variadic_utils::transform_variadic_list;
+
+// all eos's
+static constexpr const auto full_eos_list =
+    tl<IdealGas, Gruneisen, Vinet, MGUsup, PowerMG, JWL, DavisReactants, DavisProducts,
+       StiffGas, SAP_Polynomial, NobleAbel
+#ifdef SINGULARITY_USE_SPINER_WITH_HDF5
+#ifdef SINGULARITY_USE_HELMHOLTZ
+       ,
+       Helmholtz
+#endif // SINGULARITY_USE_HELMHOLTZ
+       ,
+       SpinerEOSDependsRhoT, SpinerEOSDependsRhoSie, StellarCollapse
+#endif // SINGULARITY_USE_SPINER_WITH_HDF5
+#ifdef SINGULARITY_USE_EOSPAC
+       ,
+       EOSPAC
+#endif // SINGULARITY_USE_EOSPAC
+       >{};
+// eos's that get relativistic modifier
+static constexpr const auto relativistic_eos_list =
+    tl<IdealGas
+#ifdef SINGULARITY_USE_SPINER_WITH_HDF5
+       ,
+       SpinerEOSDependsRhoT, SpinerEOSDependsRhoSie, StellarCollapse
+#endif // SINGULAIRTY_USE_SPINER_WITH_HDF5
+       >{};
+// eos's that get unit system modifier
+static constexpr const auto unit_system_eos_list =
+    tl<IdealGas
+#ifdef SPINER_USE_HDF
+       ,
+       SpinerEOSDependsRhoT, SpinerEOSDependsRhoSie, StellarCollapse
[…]
