+++ singularity-eos/CMakeLists.txt
@@ -46,6 +46,7 @@ register_headers(
     eos/eos_jwl.hpp
     eos/eos_helmholtz.hpp
     eos/eos_sap_polynomial.hpp
+    eos/eos_type_lists.hpp
     eos/modifiers/relativistic_eos.hpp
     eos/modifiers/scaled_eos.hpp
     eos/modifiers/ramps_eos.hpp
@@ -55,6 +56,7 @@ register_headers(
     eos/eos_eospac.hpp
     eos/eos_noble_abel.hpp
     eos/eos_stiff.hpp
+    eos/variant_utils.hpp
 )
 
 if (SINGULARITY_BUILD_CLOSURE)
