### 2024-05-29 #6658 @kyleniemeyer
Hello @athomps, are you available to review this submission for JOSS? It is about an equation of state library, Singularity-EOS. The repo is at https://github.com/lanl/singularity-eos and the draft paper is here: https://github.com/openjournals/joss-papers/blob/joss.06658/joss.06658/10.21105.joss.06658.pdf

If you aren't familiar with JOSS (the Journal of Open Source Software), the journal publishes articles about open source research software, where the reviews happen in the open on GitHub.
