### 2024-05-29 #6658 @athomps
@snikolov3 Will you review this JOSS contribution? It is about an equation of state library, Singularity-EOS. The repo is at https://github.com/lanl/singularity-eos and the draft paper is here: https://github.com/openjournals/joss-papers/blob/joss.06658/joss.06658/10.21105.joss.06658.pdf
