### 2024-05-30 #6658 @kyleniemeyer
@snikolov3 thanks! We ask reviewers to complete their reviews within two or three weeks if possible. However, JOSS's review process is a bit different, since it is checklist-based and we encourage interaction between reviewers and the authors via the review issue. So, you don't have to do it all at once.

A bit more about the reviewing for JOSS here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#
