### 2024-05-31 d59ba18
Add minimum energy
