+++ singularity-eos/eos/eos_models.hpp
@@ -24,6 +24,7 @@
 #include <singularity-eos/eos/eos_jwl.hpp>
 #include <singularity-eos/eos/eos_mgusup.hpp>
 #include <singularity-eos/eos/eos_noble_abel.hpp>
+#include <singularity-eos/eos/eos_carnahan_starling.hpp>
 #include <singularity-eos/eos/eos_powermg.hpp>
 #include <singularity-eos/eos/eos_sap_polynomial.hpp>
 #include <singularity-eos/eos/eos_spiner.hpp>
