+++ test/test_eos_carnahan_starling.cpp
@@ -17,8 +17,9 @@
 #include <cstdio>
 #include <cstdlib>
 #include <limits>
-#ifndef CATCH_CONFIG_RUNNER
-#include "catch2/catch.hpp"
+#ifndef CATCH_CONFIG_FAST_COMPILE
+#define CATCH_CONFIG_FAST_COMPILE
+#include <catch2/catch_test_macros.hpp>
 #endif
 
 #include <singularity-eos/base/constants.hpp>
