### 2024-05-31 70d341a
Fix bug in EOS variant specification
