### 2024-05-31 a81a857
Add to Carnahan-Staring documentation
