+++ singularity-eos/eos/eos_carnahan_starling.hpp
@@ -61,8 +61,9 @@ class CarnahanStarling : public EosBase<CarnahanStarling> {
     checkParams();
   }
   CarnahanStarling GetOnDevice() { return *this; }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real TemperatureFromDensityInternalEnergy(
-      const Real rho, const Real sie, Real *lambda = nullptr) const {
+      const Real rho, const Real sie, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return std::max(robust::SMALL(), (sie - _qq) / _Cv);
   }
   PORTABLE_INLINE_FUNCTION void checkParams() const {
@@ -70,21 +71,24 @@ class CarnahanStarling : public EosBase<CarnahanStarling> {
     PORTABLE_ALWAYS_REQUIRE(_gm1 >= 0, "Gruneisen parameter must be positive");
     PORTABLE_ALWAYS_REQUIRE(_bb >= 0, "Covolume must be positive");
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real ZedFromDensity(const Real rho,
-                                               Real *lambda = nullptr) const {
+                                               Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     const Real eta = _bb * rho;
     const Real zed =
         1.0 + robust::ratio(eta * (4.0 - 2.0 * eta), math_utils::pow<3>(1.0 - eta));
     return zed;
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PartialRhoZedFromDensity(const Real rho,
-                                                         Real *lambda = nullptr) const {
+                                                         Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     const Real eta = _bb * rho;
     return 1.0 + robust::ratio(eta * (8.0 - 2.0 * eta), math_utils::pow<4>(1.0 - eta));
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real DensityFromPressureTemperature(
       const Real press, const Real temperature, const Real guess = robust::SMALL(),
-      Real *lambda = nullptr) const {
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     Real real_root;
     auto poly = [=](Real dens) {
       return _Cv * temperature * _gm1 * dens * ZedFromDensity(dens);
@@ -104,22 +108,26 @@ class CarnahanStarling : public EosBase<CarnahanStarling> {
     }
     return std::max(robust::SMALL(), real_root);
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real InternalEnergyFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const {
+      const Real rho, const Real temperature, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     return std::max(_qq, _Cv * temperature + _qq);
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PressureFromDensityTemperature(
-      const Real rho, const Real temperature, Real *lambda = nullptr) const {
+      const Real rho, const Real temperature, Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
     const Real zed = ZedFromDensity(rho);
     return std::max(robust::SMALL(), zed * rho * temperature * _Cv * _gm1);
   }
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real PressureFromDensityInternalEnergy(
[…]
