### 2024-06-04 a918998
More changes to make creating non-default variants easier by moving tools and helper functions out of the header where variant is defined.
