+++ singularity-eos/eos/singularity_eos.cpp
@@ -1,5 +1,5 @@
 //------------------------------------------------------------------------------
-// © 2021-2023. Triad National Security, LLC. All rights reserved.  This
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
 // program was produced under U.S. Government contract 89233218CNA000001
 // for Los Alamos National Laboratory (LANL), which is operated by Triad
 // National Security, LLC for the U.S.  Department of Energy/National
@@ -17,80 +17,9 @@
 #include <singularity-eos/eos/eos.hpp>
 #include <singularity-eos/eos/eos_builder.hpp>
 #include <singularity-eos/eos/singularity_eos.hpp>
+#include <singularity-eos/eos/singularity_eos_init_utils.hpp>
 
 using namespace singularity;
-
-// TODO: Replace these with the new Modify method in EOSBuilder.
-// NOTE: The new EOSBuilder machinery will likely be slower than these.
-template <typename T>
-EOS applyShiftAndScale(T &&eos, bool scaled, bool shifted, Real scale, Real shift) {
-  if (shifted && scaled) {
-    ShiftedEOS<T> a(std::forward<T>(eos), shift);
-    ScaledEOS<ShiftedEOS<T>> b(std::move(a), scale);
-    return b;
-  }
-  if (shifted) {
-    return ShiftedEOS<T>(std::forward<T>(eos), shift);
-  }
-  if (scaled) {
-    return ScaledEOS<T>(std::forward<T>(eos), scale);
-  }
-  return eos;
-}
-
-template <typename T, template <typename> class W, typename... ARGS>
-EOS applyWrappedShiftAndScale(T &&eos, bool scaled, bool shifted, Real scale, Real shift,
-                              ARGS... args) {
-  if (shifted && scaled) {
-    ShiftedEOS<T> a(std::forward<T>(eos), shift);
-    ScaledEOS<ShiftedEOS<T>> b(std::move(a), scale);
-    W<ScaledEOS<ShiftedEOS<T>>> c(std::move(b), args...);
-    return c;
-  }
-  if (shifted) {
-    ShiftedEOS<T> sh_eos(std::forward<T>(eos), shift);
-    return W<ShiftedEOS<T>>(std::move(sh_eos), args...);
-  }
-  if (scaled) {
-    ScaledEOS<T> sc_eos(std::forward<T>(eos), scale);
-    return W<ScaledEOS<T>>(std::move(sc_eos), args...);
-  }
-  return W<T>(std::forward<T>(eos), args...);
-}
-
-template <typename T>
-EOS applyShiftAndScaleAndBilinearRamp(T &&eos, bool scaled, bool shifted, bool ramped,
-                                      Real scale, Real shift, Real r0, Real a, Real b,
-                                      Real c) {
-  if (ramped) {
-    return applyWrappedShiftAndScale<T, BilinearRampEOS>(
-        std::forward<T>(eos), scaled, shifted, scale, shift, r0, a, b, c);
-  } else {
-    return applyShiftAndScale(std::forward<T>(eos), scaled, shifted, scale, shift);
-  }
-}
-
-int init_sg_eos(const int nmat, EOS *&eos) {
-#ifdef PORTABILITY_STRATEGY_KOKKOS
-  if (!Kokkos::is_initialized()) Kokkos::initialize();
-#endif // PORTABILITY_STRATEGY_KOKKOS
-  EOS *eos_p = new EOS[nmat];
-  eos = eos_p;
-  return 0;
-}
-
-// apply everything but ramp in order to possibly calculate the
-// SAP ramp parameters from p-alhpa ramp parameters
-#define SGAPPLYMODSIMPLE(A)                                                              \
[…]
