### 2024-06-06 7fbf0c5
Add checks for bad values in inputs and outputs
