### 2024-06-11 4ce1cc7
Make checks stricter
