### 2024-06-18 c0f9b9c
Add zeroing out of material internal energy if its mass fraction is sufficiently low so as not to be participating in the solve. This fixes issues in the event the passed in internal energies are uninitialized.
