### 2024-06-19 cf9d202
GPU warning suppression of unused variables.
