+++ test/test_eos_powermg.cpp
@@ -575,7 +575,7 @@ SCENARIO("PowerMG EOS SetUp", "[VectorEOS][PowerMGEOS]") {
     Real rho0 = 7.285;
     Real T0 = 298.0;
     constexpr Real Cs = 2766.0e2;
-    //constexpr Real s = 1.5344;
+    // constexpr Real s = 1.5344;
     constexpr Real G0 = 2.4659;
     Real Cv0 = 0.2149e-05 * Mbcc_per_g;
     constexpr Real E0 = 0.658e-03 * Mbcc_per_g;
