### 2024-06-20 b1369d2
remove unneeded CI change that may have been needed during a recent upgrade hiccup.
