+++ .gitlab-ci.yml
@@ -40,8 +40,6 @@ default:
   id_tokens:
     SITE_ID_TOKEN:
       aud: https://asc-git.lanl.gov
-    CI_JOB_JWT:
-      aud: https://asc-git.lanl.gov
 
 ##########################
 # Build Matrix Variables #
