### 2024-06-22 7ef37e6
Lift StrCat out of the base class into a more accessible header
