+++ singularity-eos/eos/eos_sap_polynomial.hpp
@@ -25,7 +25,7 @@
 
 // Base stuff
 #include <singularity-eos/base/constants.hpp>
-#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 #include <singularity-eos/base/robust_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
