+++ singularity-eos/eos/modifiers/relativistic_eos.hpp
@@ -24,7 +24,7 @@
 #include <utility>
 
 #include <ports-of-call/portability.hpp>
-#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 #include <singularity-eos/base/robust_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
