+++ singularity-eos/eos/modifiers/ramps_eos.hpp
@@ -25,7 +25,7 @@
 
 #include <ports-of-call/portability.hpp>
 #include <singularity-eos/base/constants.hpp>
-#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 #include <singularity-eos/base/robust_utils.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
