+++ singularity-eos/eos/default_variant.hpp
@@ -29,7 +29,7 @@
 
 // Base stuff
 #include <singularity-eos/base/constants.hpp>
-#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 #include <singularity-eos/base/variadic_utils.hpp>
 
 // EOS models
