+++ singularity-eos/eos/modifiers/minimum_energy.hpp
@@ -20,7 +20,7 @@
 #include <iostream>
 
 #include <ports-of-call/portability.hpp>
-#include <singularity-eos/base/eos_error.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
 namespace singularity {
