### 2024-06-22 d933b8c
Rework value checking so entire state is printed when a bad value is detected
