+++ singularity-eos/eos/get_sg_eos_functors.hpp
@@ -17,6 +17,7 @@
 
 #include <ports-of-call/portability.hpp>
 #include <singularity-eos/eos/get_sg_eos.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 
 #include <cmath>
 
@@ -61,11 +62,6 @@ struct init_functor {
   PORTABLE_INLINE_FUNCTION
   void operator()(const int i, const int tid, double &mass_sum, int &npte,
                   const Real t_mult, const Real s_mult, const Real p_mult) const {
-    // Debug checks for inputs
-    check_val(spvol_v(i));
-    check_val(sie_v(i));
-    check_val(press_v(i));
-    // check_val(temp_v(i));
     /* normalize mass fractions */
     /* first find the mass sum */
     /* also set idxs as the decrement of the eos offsets */
@@ -77,8 +73,8 @@ struct init_functor {
     }
     for (int m = 0; m < nmat; ++m) {
       frac_mass_v(i, m) /= mass_sum;
-      check_val(frac_mass_v(i, m));
     }
+    check_all_vals(i);
     // count the number of participating materials and zero the inputs
     npte = 0;
     for (int m = 0; m < nmat; ++m) {
@@ -113,12 +109,50 @@ struct init_functor {
     return;
   }
  private:
-  // Debug check to make sure a value is normal or zero
+  // Debug check to make sure returned values are normal or zero. Extra NDEBUG
+  // check probably not required, but it's there just in case the compiler tries
+  // to evaluate the boolean
   template<typename valT>
   PORTABLE_FORCEINLINE_FUNCTION
-  void check_val(valT value) const {
-    PORTABLE_ALWAYS_REQUIRE(value == valT{0} || (std::isnormal(1e50 * value) && std::isnormal(value)),
-                     "Bad value input to singularity-eos interface");
+  int bad_val(valT const value, const char *const var_name) const {
+    const bool bad = (value == valT{0}) || (std::isnormal(1e10 * value) && std::isnormal(value));
+    if (bad) {
+      using PortsOfCall::printf;
+      printf("### ERROR: Bad singularity-eos input\n  Var: %s\n  Value: %e\n",
+             var_name, value);
+    }
+    return bad;
+  }
+  PORTABLE_FORCEINLINE_FUNCTION
+  void check_all_vals(int const i) const {
+    int n_bad_vals = 0;
+
+    for (int m = 0; m < nmat; ++m) {
+      n_bad_vals += bad_val(frac_mass_v(i, m), "frac_mass");
+    }
+    n_bad_vals += bad_val(press_v(i), "pres");
+    n_bad_vals += bad_val(sie_v(i), "sie");
+    n_bad_vals += bad_val(spvol_v(i), "bmod");
+
+    if (n_bad_vals > 0) {
+      using PortsOfCall::printf;
+      printf("### Bad Value Output state:\n");
+      printf("  - Bulk state -\n");
+      printf("    Pressure        : % #24.15g\n", press_v(i));
+      printf("    Specific IE     : % #24.15g\n", sie_v(i));
+      printf("    Specific Volume : % #24.15g\n", spvol_v(i));
+      printf("  - Material States -\n");
+      printf("     i:");
+      printf(" %23s", "Mass fraction");
+      printf("\n");
+      for (int m = 0; m < nmat; ++m) {
+        printf("   %3i:", m);
+        printf(" %24.15g", frac_mass_v(i, m));
+        printf("\n");
+      }
+      PORTABLE_ALWAYS_ABORT(
[…]
