### 2024-06-26 5678e51
Add introspection functions to functors to locate bad values
