+++ singularity-eos/eos/get_sg_eos_functors.hpp
@@ -18,6 +18,8 @@
 #include <ports-of-call/portability.hpp>
 #include <singularity-eos/eos/get_sg_eos.hpp>
 
+#include <cmath>
+
 #ifdef PORTABILITY_STRATEGY_KOKKOS
 
 namespace singularity {
@@ -71,31 +73,87 @@ struct init_functor {
     for (int m = 0; m < nmat; ++m) {
       frac_mass_v(i, m) /= mass_sum;
     }
-    /* set inputs */
+    check_all_vals(i);
+    // count the number of participating materials and zero the inputs
     npte = 0;
     for (int m = 0; m < nmat; ++m) {
       if (frac_mass_v(i, m) > 1.e-12) {
+        // participating materials are those with non-negligible mass fractions
         pte_idxs(tid, npte) = eos_offsets_v(m) - 1;
         pte_mats(tid, npte) = m;
         npte += 1;
       } else {
         frac_ie_v(i, m) = 0.0;
       }
+      // zero the inputs
       vfrac_pte(tid, m) = 0.0;
       sie_pte(tid, m) = 0.0;
       temp_pte(tid, m) = 0.0;
       press_pte(tid, m) = 0.0;
     }
+    // Populate the inputs with consistent values.
+    // NOTE: the volume fractions and densities need to be consistent with the
+    // total specific volume since they are used to calculate internal
+    // quantities for the PTE solver
     for (int mp = 0; mp < npte; ++mp) {
       const int m = pte_mats(tid, mp);
-      rho_pte(tid, mp) = npte / spvol_v(i) * frac_mass_v(i, m);
+      // Need to guess volume fractions
       vfrac_pte(tid, mp) = 1.0 / npte;
+      // Calculate densities to be consistent with these volume fractions
+      rho_pte(tid, mp) = frac_mass_v(i, m) / spvol_v(i) / vfrac_pte(tid, mp);
       temp_pte(tid, mp) = temp_v(i) * ev2k * t_mult;
       press_pte(tid, mp) = press_v(i) * p_mult;
       sie_pte(tid, mp) = sie_v(i) * frac_mass_v(i, m) * s_mult;
     }
     return;
   }
+ private:
+  // Check to make sure returned values are normal or zero
+  template<typename valT>
+  PORTABLE_FORCEINLINE_FUNCTION
+  int bad_val(valT const value, const char *const var_name) const {
+    const bool good = (value == valT{0}) or (std::isnormal(1e10 * value) and std::isnormal(value));
+    if (not good) {
+      using PortsOfCall::printf;
+      printf("### ERROR: Bad singularity-eos input\n  Var: %s\n  Value: %e\n",
+             var_name, value);
+    }
+    return not good;
+  }
+  // Only run these checks when built with debug flags
+  PORTABLE_FORCEINLINE_FUNCTION
+  void check_all_vals(int const i) const {
+#ifndef NDEBUG
+    int n_bad_vals = 0;
+
+    for (int m = 0; m < nmat; ++m) {
+      n_bad_vals += bad_val(frac_mass_v(i, m), "frac_mass");
+    }
+    n_bad_vals += bad_val(press_v(i), "pres");
+    n_bad_vals += bad_val(sie_v(i), "sie");
+    n_bad_vals += bad_val(spvol_v(i), "spvol");
+
+    if (n_bad_vals > 0) {
+      using PortsOfCall::printf;
+      printf("### Bad Value Output state:\n");
+      printf("  - Bulk state -\n");
+      printf("    Pressure        : % #24.15g\n", press_v(i));
+      printf("    Specific IE     : % #24.15g\n", sie_v(i));
+      printf("    Specific Volume : % #24.15g\n", spvol_v(i));
[…]
