### 2024-06-26 05377f4
Provide PTE success/failure to final functor
