### 2024-06-26 e042a6a
Add checks for negative and zero values (where appropriate)
