### 2024-06-26 9b69e76
Change 'and' to '&&'
