### 2024-06-26 91b0544
Change volume fraction guess to be equal to mass fraction instead of equipartitioning
