### 2024-06-27 db35049
Add error utilities for checking conditions and printing error information
