### 2024-06-27 996db13
Add optional argument to get_sg_eos_f() to specify the mass fraction cutoff for PTE participation
