### 2024-06-27 2f10112
Use error_utils.hpp functions and add mass fraction cutoff to init constructor
