+++ singularity-eos/eos/get_sg_eos_functors.hpp
@@ -17,6 +17,7 @@
 
 #include <ports-of-call/portability.hpp>
 #include <singularity-eos/eos/get_sg_eos.hpp>
+#include <singularity-eos/base/error_utils.hpp>
 
 #include <cmath>
 
@@ -42,6 +43,7 @@ struct init_functor {
   dev_v press_v;
   dev_v sie_v;
   int nmat;
+  double mass_frac_cutoff;
 
  public:
   init_functor() = default;
@@ -51,12 +53,14 @@ struct init_functor {
                ScratchV<double> &vfrac_pte_, ScratchV<double> &sie_pte_,
                ScratchV<double> &temp_pte_, ScratchV<double> &press_pte_,
                ScratchV<double> &rho_pte_, dev_v &spvol_v_, dev_v &temp_v_,
-               dev_v &press_v_, dev_v &sie_v_, int &nmat)
+               dev_v &press_v_, dev_v &sie_v_, int &nmat,
+               double &mass_frac_cutoff_)
       : frac_mass_v{frac_mass_v_}, pte_idxs{pte_idxs_}, eos_offsets_v{eos_offsets_v_},
         frac_vol_v{frac_vol_v_}, frac_ie_v{frac_ie_v_}, pte_mats{pte_mats_},
         vfrac_pte{vfrac_pte_}, sie_pte{sie_pte_}, temp_pte{temp_pte_},
         press_pte{press_pte_}, rho_pte{rho_pte_}, spvol_v{spvol_v_}, temp_v{temp_v_},
-        press_v{press_v_}, sie_v{sie_v_}, nmat{nmat} {}
+        press_v{press_v_}, sie_v{sie_v_}, nmat{nmat}, mass_frac_cutoff{mass_frac_cutoff_}
+        {}
 
   PORTABLE_INLINE_FUNCTION
   void operator()(const int i, const int tid, double &mass_sum, int &npte,
@@ -77,7 +81,7 @@ struct init_functor {
     // count the number of participating materials and zero the inputs
     npte = 0;
     for (int m = 0; m < nmat; ++m) {
-      if (frac_mass_v(i, m) > 1.e-12) {
+      if (frac_mass_v(i, m) > mass_frac_cutoff) {
         // participating materials are those with non-negligible mass fractions
         pte_idxs(tid, npte) = eos_offsets_v(m) - 1;
         pte_mats(tid, npte) = m;
@@ -109,33 +113,20 @@ struct init_functor {
   }
 
  private:
-  // Check to make sure returned values are normal or zero
-  template <typename valT>
-  PORTABLE_FORCEINLINE_FUNCTION int bad_val(valT const value,
-                                            const char *const var_name) const {
-    const bool good =
-        (value == valT{0}) || (std::isnormal(1e10 * value) && std::isnormal(value));
-    if (not good) {
-      using PortsOfCall::printf;
-      printf("### ERROR: Bad singularity-eos input\n  Var: %s\n  Value: %e\n", var_name,
-             value);
-    }
-    return not good;
-  }
   // Only run these checks when built with debug flags
   PORTABLE_FORCEINLINE_FUNCTION
   void check_all_vals(int const i) const {
 #ifndef NDEBUG
-    int n_bad_vals = 0;
-
+    bool any_bad_vals = false;
     for (int m = 0; m < nmat; ++m) {
-      n_bad_vals += bad_val(frac_mass_v(i, m), "frac_mass");
+      any_bad_vals = any_bad_vals || error_utils::bad_value(frac_mass_v(i, m),
+                                                            "frac_mass");
     }
-    n_bad_vals += bad_val(press_v(i), "pres");
-    n_bad_vals += bad_val(sie_v(i), "sie");
-    n_bad_vals += bad_val(spvol_v(i), "spvol");
[…]
