### 2024-06-27 ca97060
Make index variables auto type
