### 2024-06-27 08d7828
Change to appropriate Fortran input type for get_sg_eos_f()
