### 2024-06-27 2f78b47
Change get_sg_eos function prototype to match function
