### 2024-06-27 #6805 @Yurlungur
> > Hi @kyleniemeyer @parikshitbajpai @snikolov3 I just wanted to reach out and check on the status of review. Do you have a sense when you might be able to take a deeper look? Thanks!
>
> Hi @Yurlungur, Sorry about the delay on this. I am half way through the review and should be able to finish it beginning of next week.

:+1: thanks @parikshitbajpai .
