### 2024-06-27 02ddf39
Update changelog to reflect new mass frac cutoff
