### 2024-06-27 b9000ff
Adjust formatting and add copyright
