+++ singularity-eos/base/error_utils.hpp
@@ -1,3 +1,17 @@
+//------------------------------------------------------------------------------
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
+// program was produced under U.S. Government contract 89233218CNA000001
+// for Los Alamos National Laboratory (LANL), which is operated by Triad
+// National Security, LLC for the U.S.  Department of Energy/National
+// Nuclear Security Administration. All rights in the program are
+// reserved by Triad National Security, LLC, and the U.S. Department of
+// Energy/National Nuclear Security Administration. The Government is
+// granted for itself and others acting on its behalf a nonexclusive,
+// paid-up, irrevocable worldwide license in this material to reproduce,
+// prepare derivative works, distribute copies to the public, perform
+// publicly and display publicly, and to permit others to do so.
+//------------------------------------------------------------------------------
+
 #ifndef SINGULARITY_EOS_BASE_ERROR_UTILS_HPP_
 #define SINGULARITY_EOS_BASE_ERROR_UTILS_HPP_
 
@@ -45,8 +59,8 @@ PORTABLE_FORCEINLINE_FUNCTION bool violates_condition(valT &&value, condT &&cond
                                                       nameT &&var_name) {
   const bool good = condition(std::forward<valT>(value));
   if (!good) {
-    printf("### ERROR: Bad singularity-eos value\n  Var: %s\n Value: %.15e\n", var_name,
-           value);
+    printf("### ERROR: Bad singularity-eos value\n  Var:   %s\n  Value: %.15e\n",
+           var_name, value);
   }
   return !good;
 }
