+++ singularity-eos/eos/singularity_eos.f90
@@ -1,5 +1,5 @@
 !------------------------------------------------------------------------------
-! © 2021-2023. Triad National Security, LLC. All rights reserved.  This
+! © 2021-2024. Triad National Security, LLC. All rights reserved.  This
 ! program was produced under U.S. Government contract 89233218CNA000001
 ! for Los Alamos National Laboratory (LANL), which is operated by Triad
 ! National Security, LLC for the U.S.  Department of Energy/National
