### 2024-06-27 ef88154
Simplify line search and add comments
