+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -1426,16 +1426,16 @@ PORTABLE_INLINE_FUNCTION bool PTESolver(System &s) {
     err = s.TestUpdate(scale);
     if (err > err_old + line_search_alpha * gradfdx) {
       // backtrack
-      Real err_mid = s.TestUpdate(0.5);
+      scale = 0.5;
+      Real err_mid = s.TestUpdate(scale);
       if (err_mid < err && err_mid < err_old) {
+        // try a larger step since a half step reduces the error
         scale = 0.75 + 0.5 * robust::ratio(err_mid - err, err - 2.0 * err_mid + err_old);
-      } else {
-        scale = line_search_fac;
       }
-
       for (int line_iter = 0; line_iter < line_search_max_iter; line_iter++) {
         err = s.TestUpdate(scale);
         if (err < err_old + line_search_alpha * scale * gradfdx) break;
+        // shrink the step if the error isn't reduced
         scale *= line_search_fac;
       }
     }
