### 2024-06-27 8b99c75
Make sure the scale that comes out is the minimum over all checks
