### 2024-06-27 97bc6dd
Add spvol to final functor
