### 2024-06-27 f60e4b4
Remove commented variable and add clarifying comment
