+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -1420,7 +1420,7 @@ PORTABLE_INLINE_FUNCTION bool PTESolver(System &s) {
 
     // Line search
     Real gradfdx = -2.0 * scale * err;
-    scale = 1.0;  // New scale for line search
+    scale = 1.0; // New scale for line search
     Real err_old = err;
     err = s.TestUpdate(scale);
     if (err > err_old + line_search_alpha * gradfdx) {
