### 2024-06-27 503750f
Add missing initializer
