### 2024-06-28 ca3d16a
Change behavior to assume small mass fraction materials are at the cell sie and density
