+++ singularity-eos/eos/get_sg_eos_rho_t.cpp
@@ -56,9 +56,9 @@ void get_sg_eos_rho_t(const char *name, int ncell, indirection_v &offsets_v,
           // eos accessor
           singularity::EOSAccessor_ eos_inx(eos_v, &pte_idxs(tid, 0));
           PTESolverFixedT<singularity::EOSAccessor_, Real *, Real **> method(
-              npte, eos_inx, vfrac_sum, temp_pte(tid, 0), &rho_pte(tid, 0), &vfrac_pte(tid, 0),
-              &sie_pte(tid, 0), &temp_pte(tid, 0), &press_pte(tid, 0), cache,
-              &solver_scratch(tid, 0));
+              npte, eos_inx, vfrac_sum, temp_pte(tid, 0), &rho_pte(tid, 0),
+              &vfrac_pte(tid, 0), &sie_pte(tid, 0), &temp_pte(tid, 0), &press_pte(tid, 0),
+              cache, &solver_scratch(tid, 0));
           pte_converged = PTESolver(method);
           // calculate total internal energy
           for (int mp = 0; mp < npte; ++mp) {
