+++ singularity-eos/eos/get_sg_eos_rho_p.cpp
@@ -55,9 +55,9 @@ void get_sg_eos_rho_p(const char *name, int ncell, indirection_v &offsets_v,
           // eos accessor
           singularity::EOSAccessor_ eos_inx(eos_v, &pte_idxs(tid, 0));
           PTESolverFixedP<singularity::EOSAccessor_, Real *, Real *> method(
-              npte, eos_inx, vfrac_sum, press_pte(tid, 0), &rho_pte(tid, 0), &vfrac_pte(tid, 0),
-              &sie_pte(tid, 0), &temp_pte(tid, 0), &press_pte(tid, 0), cache[0],
-              &solver_scratch(tid, 0));
+              npte, eos_inx, vfrac_sum, press_pte(tid, 0), &rho_pte(tid, 0),
+              &vfrac_pte(tid, 0), &sie_pte(tid, 0), &temp_pte(tid, 0), &press_pte(tid, 0),
+              cache[0], &solver_scratch(tid, 0));
           pte_converged = PTESolver(method);
           // calculate total sie
           for (int mp = 0; mp < npte; ++mp) {
