### 2024-07-02 e885113
min not needed since scales are checked sequentially _after_ being modified
