+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -675,7 +675,7 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     // control how big of a step toward vfrac = 0 is allowed
     for (int m = 0; m < nmat; ++m) {
       if (scale * dx[m] < -vfrac_safety_fac * vfrac[m]) {
-        scale = std::min(scale, -vfrac_safety_fac * robust::ratio(vfrac[m], dx[m]));
+        scale = -vfrac_safety_fac * robust::ratio(vfrac[m], dx[m]);
       }
     }
     const Real Tnew = Tequil + scale * dx[nmat];
@@ -685,12 +685,12 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
           std::max(eos[m].RhoPmin(Tnorm * Tequil), eos[m].RhoPmin(Tnorm * Tnew));
       const Real alpha_max = robust::ratio(rhobar[m], rho_min);
       if (scale * dx[m] > 0.5 * (alpha_max - vfrac[m])) {
-        scale = std::min(scale, robust::ratio(0.5 * alpha_max - vfrac[m], dx[m]));
+        scale = robust::ratio(0.5 * alpha_max - vfrac[m], dx[m]);
       }
     }
     // control how big of a step toward T = 0 is allowed
     if (scale * dx[nmat] < -0.95 * Tequil) {
-      scale = std::min(scale, robust::ratio(-0.95 * Tequil, dx[nmat]));
+      scale = robust::ratio(-0.95 * Tequil, dx[nmat]);
     }
     // Now apply the overall scaling
     for (int i = 0; i < neq; ++i)
@@ -883,7 +883,7 @@ class PTESolverFixedT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
     // control how big of a step toward vfrac = 0 is allowed
     for (int m = 0; m < nmat; ++m) {
       if (scale * dx[m] < -vfrac_safety_fac * vfrac[m]) {
-        scale = std::min(scale, robust::ratio(-vfrac_safety_fac * vfrac[m], dx[m]));
+        scale = robust::ratio(-vfrac_safety_fac * vfrac[m], dx[m]);
       }
     }
     // control how big of a step toward rho = rho(Pmin) is allowed
@@ -891,7 +891,7 @@ class PTESolverFixedT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
       const Real rho_min = eos[m].RhoPmin(Tequil);
       const Real alpha_max = robust::ratio(rhobar[m], rho_min);
       if (scale * dx[m] > 0.5 * (alpha_max - vfrac[m])) {
-        scale = std::min(scale, robust::ratio(0.5 * (alpha_max - vfrac[m]), dx[m]));
+        scale = robust::ratio(0.5 * (alpha_max - vfrac[m]), dx[m]);
       }
     }
     // Now apply the overall scaling
@@ -1097,7 +1097,7 @@ class PTESolverFixedP : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
     // control how big of a step toward vfrac = 0 is allowed
     for (int m = 0; m < nmat; ++m) {
       if (scale * dx[m] < -vfrac_safety_fac * vfrac[m]) {
-        scale = std::min(scale, -vfrac_safety_fac * robust::ratio(vfrac[m], dx[m]));
+        scale = -vfrac_safety_fac * robust::ratio(vfrac[m], dx[m]);
       }
     }
     const Real Tnew = Tequil + scale * dx[nmat];
[…]
