### 2024-07-02 71250a5
Add sanity check that initial volume fractions add up to 1 or less
