### 2024-07-02 a4e1f0c
Add more asserts and label what is going on if an assert is tripped
