+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -713,7 +713,8 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
         vtemp[m] = vfrac[m];
     }
     Tequil = Ttemp + scale * dx[nmat];
-    PORTABLE_REQUIRE(Tequil >= 0., "Negative temperature update in rho-T PTE solver iteration");
+    PORTABLE_REQUIRE(Tequil >= 0.,
+                     "Negative temperature update in rho-T PTE solver iteration");
     for (int m = 0; m < nmat; ++m) {
       vfrac[m] = vtemp[m] + scale * dx[m];
       rho[m] = robust::ratio(rhobar[m], vfrac[m]);
@@ -1136,7 +1137,8 @@ class PTESolverFixedP : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
         vtemp[m] = vfrac[m];
     }
     Tequil = Ttemp + scale * dx[nmat];
-    PORTABLE_REQUIRE(Tequil >= 0., "Negative temperature in Fixed P PTE solver iteration");
+    PORTABLE_REQUIRE(Tequil >= 0.,
+                     "Negative temperature in Fixed P PTE solver iteration");
     for (int m = 0; m < nmat; ++m) {
       vfrac[m] = vtemp[m] + scale * dx[m];
       rho[m] = robust::ratio(rhobar[m], vfrac[m]);
