+++ aemattssing/_sources/src/using-eos.rst.txt
@@ -1,0 +1,699 @@
+.. _using-eos:
+
+The Equation of State API
+=========================
+
+This page describes the equation of state API in detail. For just the
+information needed to get started, check out the :ref:`getting started
+<getting-started>` page.
+
+The ``Real`` Type
+------------------
+
+``singularity-eos`` defines the ``singularity::Real`` type as a proxy
+for the ``float`` and ``double`` types. We currently resolve ``Real``
+to a double precision number, however we plan to have the option to
+select different precisions at compile time in the future.
+
+The parallelism model
+----------------------
+
+For the most part, ``singularity-eos`` tries to be agnostic to how you
+parallelize your code on-node. (It knows nothing at all about
+distributed memory parallelism.) An ``EOS`` object can be copied into
+any parallel code block by value (see below) and scalar calls do not
+attempt any internal multi-threading, meaning ``EOS`` objects are not
+thread-safe, but are compatible with thread safety, assuming the user
+calls them appropriately. The main complication is ``lambda`` arrays,
+which are discussed below.
+
+The vector ``EOS`` method overloads are a bit different. These are
+thread-parallel operations launched by ``singularity-EOS``. They run
+in parallel, and ordering between indices of vectors cannot be
+assumed. Therefore care must taken in memory layout to avoid race
+conditions. The type of parallelism used depends on how
+``singularity-eos`` is compiled. If the ``Kokkos`` backend is used,
+any parallel dispatch supported by ``Kokkos`` is supported.
+
+A more generic version of the vector calls exists in the ``Evaluate``
+method, which allows the user to specify arbitrary parallel dispatch
+models by writing their own loops. See the relevant section below.
+
+.. _variant section:
+
+Variants
+---------
+
+The equation of state library is object oriented, and uses a kind of
+type erasure called a `Variant`_. (Technically we use a backport of
+this C++ feture to C++11, see: `mpark variant`_.) The salient detail
+is that a variant is a kind of compile-time polymorphism.
+
+.. _Variant: https://en.cppreference.com/w/cpp/utility/variant
+
+.. _mpark variant: https://en.cppreference.com/w/cpp/utility/variant
+
+The ``singularity::EOS`` class is generic and can be initialized as
+any equation of state model listed in :ref:`the models section
+<models>`. Unlike with standard polymorphism, you don't need to
+initialize your equation of state as a pointer. Rather, just use the
+assignment operator. For example:
+
+.. code-block:: cpp
+
+   singularity::EOS my_eos = singularity::IdealGas(gm1, Cv);
+
+To make this machinery work, there's an underlying variatic class,
+``singularity::Variant``, defined in
+``singularity-eos/eos/eos_variant.hpp``. Only methods defined for the
+``singularity::Variant`` class are available for the equation of state
+models. Moreover, any new equation of state model must define all
+methods defined in the ``singularity::Variant`` class that call the ``visit``
+function, or compile errors may occur.
+
+If you wish to extract an underlying EOS model as an independent type,
+undoing the type erasure, you can do so with the ``get``
+method. ``get`` is templated and type deduction is not possible. You
+must specify the type of the class you're pulling out of the
+variant. For example:
+
+.. code-block::
+
+   auto my_ideal_gas = my_eos.get<singularity::IdealGas>();
+
+This will give you access to methods and fields which may be unique to
+a class but not shared by the ``Variant``.
+
+The EOS model also allows some host-side introspection. The method
+
+.. cpp:function:: static std::string EosType();
+
[…]
