+++ aemattssing/_sources/src/using-closures.rst.txt
@@ -1,0 +1,539 @@
+.. _using-closures:
+
+Mixed Cell Closures
+====================
+
+In the single-material Euler equations, the mass and energy are typically
+evolved and the EOS is called to provide a pressure for the momentum equations.
+When transitioning to a multi-material approach, a single velocity is typically
+used and the Euler equations are solved with respect to the bulk fluid motion.
+In this case, the pressure contribution to momentum isn't well-defined and in
+principle each material could have its own pressure contribution to material
+motion. Furthermore, the paritioning of volume and energy between the materials
+in the flow is not well-defined either.
+
+As a result, a multi-material closure rule is needed to determine both how to
+compute the pressure response of the flow and how to partition the volume and
+energy between the individual materials. In this situation, one must decide how
+to compute thermodynamic quantities in that cell for each material.
+
+Governing Equations and Assumptions
+------------------------------------
+
+In a general sense then the mixed
+material closure rule takes the form
+
+.. math::
+
+  P_i = F(\rho, \epsilon, \mu_1, ..., \mu_i, ..., \mu_{N-1}) \\
+  \rho_i = G(\rho, \epsilon, \mu_1, ..., \mu_i, ..., \mu_{N-1}) \\
+  \epsilon_i = H(\rho, \epsilon, \mu_1, ..., \mu_i, ..., \mu_{N-1}) \\
+  T_i = J(\rho, \epsilon, \mu_1, ..., \mu_i, ..., \mu_{N-1})
+
+where each material has its own density, :math:`\rho_i`, and specific internal
+energy, :math:`e_i`, as well as in principle its own pressure and temperature.
+Each will be a function of the bulk density, :math:`\rho`, the bulk specific
+internal energy, :math:`\epsilon`, and the mass fractions, :math:`\mu_j`,
+of all other materials present.
+
+For convenience of minimizing input arguments though, the solvers in
+``singularity-eos`` are typically posed in a different way that can sometimes be
+unintuitive.
+
+For some finite total volume :math:`V`, each material occupies some fraction of
+that volume given by the *volume fraction*
+:math:`f_i` such that
+
+.. math::
+
+  \sum_{i=0}^{N - 1} f_i = f_\mathrm{tot},
+
+where :math:`f_\mathrm{tot}` is the total fraction of the total volume being
+considered (in principle, different closure models can be used for different
+sets of materials). To consider the entire volume, :math:`f_\mathrm{tot}` can
+simply be set to one.
+
+The average density, :math:`\overline{\rho}_i`, (i.e. mass per *total* volume)
+for a material in the total volume is
+
+.. math::
+
+  \overline{\rho}_i = \rho_i f_i,
+
+where :math:`\rho_i` is the physical density (i.e. material mass per *material*
+volume). It is important to note here that while the densities, :math:`\rho_i`,
+and the volume fractions, :math:`f_i`, will vary as the closure model is
+applied, the average densities, :math:`\overline{\rho}_i`, will all remain
+constant, motiviating their internal use in the closure solvers. The total
+density (mass of *participating* materials per total volume) is then
+
+.. math::
+
+  \rho = \sum_{i=0}^{N - 1} \overline{\rho}_i = \sum_{i=0}^{N-1} \rho_i f_i
+
+Similarly the energy can be summed in a similar way so that
+
+.. math::
+
+  u = \rho \epsilon = \sum_{i = 0}^{N - 1} \rho_i \epsilon_i 
+  = \sum_{i = 0}^{N - 1} u_i
+
+where :math:`u` is the total internal energy density (internal energy per unit
[…]
