### 2024-07-04 bb61c27
Add closure unit test
