### 2024-07-04 1e2952b
Change TestUpdate() to take an argument to reset the cached variables and temporarily comment out vfrac checks
