### 2024-07-04 52de0c4
Move #endif to be in proper place
