### 2024-07-04 ac21c03
Fix another misplaced #endif
