+++ test/test_closure_pte.cpp
@@ -20,8 +20,8 @@
 #endif
 
 #include <ports-of-call/portability.hpp>
-#include <singularity-eos/eos/eos.hpp>
 #include <singularity-eos/closure/mixed_cell_models.hpp>
+#include <singularity-eos/eos/eos.hpp>
 
 using namespace singularity;
 
@@ -40,10 +40,10 @@ SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
     constexpr int num_eos = 3;
     std::array<EOS, num_eos> eos_arr;
     // Reference state
-    constexpr Real P0 = 1.0e6;  // 1 bar
-    constexpr Real T0 = 296;  // K
+    constexpr Real P0 = 1.0e6; // 1 bar
+    constexpr Real T0 = 296;   // K
     // Ideal gas air
-    constexpr Real rho0_air = 1e-03;  // g/cc
+    constexpr Real rho0_air = 1e-03; // g/cc
     constexpr Real Gruneisen_air = 0.4;
     constexpr Real CV_air = P0 / rho0_air / (Gruneisen_air * T0);
     eos_arr[0] = IdealGas(Gruneisen_air, CV_air);
@@ -56,8 +56,8 @@ SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
     constexpr Real b = 0.58;
     constexpr Real k = 1.35;
     constexpr Real n = 2.66182;
-    constexpr Real vc = 0.75419; // cm^3 / g
-    constexpr Real pc = 3.2 * GPa; // microbar
+    constexpr Real vc = 0.75419;              // cm^3 / g
+    constexpr Real pc = 3.2 * GPa;            // microbar
     constexpr Real Cv = 0.001072 * MJ_per_kg; // erg / g / K
     constexpr Real Q = 4.115 * MJ_per_kg;
     eos_arr[2] = DavisProducts(a, b, k, n, vc, pc, Cv).Modify<ShiftedEOS>(-Q);
@@ -67,15 +67,15 @@ SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
       eos.GetOnDevice();
     }
     constexpr size_t EOS_bytes = num_eos * sizeof(EOS);
-    EOS* v_EOS = (EOS*)PORTABLE_MALLOC(EOS_bytes);
+    EOS *v_EOS = (EOS *)PORTABLE_MALLOC(EOS_bytes);
     portableCopyToDevice(v_EOS, eos_arr.data(), EOS_bytes);
 
     GIVEN("A difficult state for the density-temperature PTE solver") {
       // Define the state
       constexpr Real spvol_bulk = 6.256037280402093e-01;
       constexpr Real sie_bulk = -1.441692060406520e+10;
-      const std::array<const Real, num_eos> mass_frac =
-          {1.10382442033331e-10, 0.124935312146569, 0.875064687743048};
+      const std::array<const Real, num_eos> mass_frac = {
+          1.10382442033331e-10, 0.124935312146569, 0.875064687743048};
       // Initial guess: all materials at the cell density
       const std::array<const Real, num_eos> vol_frac = mass_frac;
       // Calculate material densities (and corresponding energies) and the total
@@ -93,26 +93,25 @@ SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
       std::array<Real, num_eos> pressures = {0., 0., 0.};
       // Copy values to device (when available)
       constexpr size_t bytes = num_eos * sizeof(Real);
-      Real* v_densities = (Real*)PORTABLE_MALLOC(bytes);
+      Real *v_densities = (Real *)PORTABLE_MALLOC(bytes);
       portableCopyToDevice(v_densities, densities.data(), bytes);
[…]
