### 2024-07-11 f017752
Fix typo in scale update and skip update if appropriate
