### 2024-07-11 d834c84
Add PORTABLE_FREE and change check condition temporarily
