### 2024-07-11 ae40f6a
Fix bug in other PTE solvers
