### 2024-07-11 b994fdf
Remove requirements on volume fractions within iteration since there's better information afterwards
