### 2024-07-11 af9e22a
Add check to ensure step size is always less than or equal to Newton step
