### 2024-07-11 1c5d5b7
Change singularity version for updated POC
