### 2024-07-11 65788e5
Remove debug checks from inside the PTE solver iteration
