### 2024-07-11 49fbd27
Add PORTABLE_FREE and finalize... this should work but doesn't
