### 2024-07-15 bf879f0
Minor changes / typos in JOSS paper
