### 2024-07-16 1a048ad
Break out parts of tests into functions
