+++ test/test_closure_pte.cpp
@@ -11,8 +11,10 @@
 // prepare derivative works, distribute copies to the public, perform
 // publicly and display publicly, and to permit others to do so.
 
-#include <array>
+#ifdef SINGULARITY_BUILD_CLOSURE
+
 #include <string>
+#include <vector>
 
 #ifndef CATCH_CONFIG_FAST_COMPILE
 #define CATCH_CONFIG_FAST_COMPILE
@@ -29,17 +31,111 @@ using namespace singularity;
 constexpr Real GPa = 1.0e10;
 constexpr Real MJ_per_kg = 1.0e10;
 
-#ifdef SINGULARITY_BUILD_CLOSURE
 #ifdef SINGULARITY_TEST_SESAME
 #ifdef SINGULARITY_USE_SPINER
 #ifdef SPINER_USE_HDF
 
+using myEOS = singularity::Variant<IdealGas, ShiftedEOS<DavisProducts>, DavisProducts,
+                                   SpinerEOSDependsRhoT>;
+
+template <typename EOSArrT>
+myEOS *copy_eos_arr_to_device(const int num_eos, EOSArrT eos_arr) {
+  // Move EOS array from host to device
+  for (auto i = 0; i < num_eos; i++) {
+    eos_arr[i] = eos_arr[i].GetOnDevice();
+  }
+  const size_t EOS_bytes = num_eos * sizeof(myEOS);
+  myEOS *v_EOS = (myEOS *)PORTABLE_MALLOC(EOS_bytes);
+  portableCopyToDevice(v_EOS, eos_arr.data(), EOS_bytes);
+
+  return v_EOS;
+}
+
+void finalize_eos_arr(const int num_eos, myEOS *v_EOS) {
+  // Call Finalize on each EOS
+  for (auto i = 0; i < num_eos; i++) {
+    v_EOS[i].Finalize();
+  }
+  // Free EOS memory
+  PORTABLE_FREE(v_EOS);
+}
+
+template <typename ArrT>
+bool run_PTE_from_state(const int num_eos, myEOS *v_EOS, const Real spvol_bulk,
+                        const Real sie_bulk, ArrT mass_frac) {
+  // Calculate material densities (and corresponding energies) and the total
+  // volume fraction
+  std::vector<Real> vol_frac(num_eos);
+  std::vector<Real> densities(num_eos);
+  std::vector<Real> sies(num_eos);
+  std::vector<Real> temperatures(num_eos);
+  std::vector<Real> pressures(num_eos);
+  Real vfrac_sum = 0;
+  for (auto i = 0; i < num_eos; ++i) {
+    // Initial guess: all materials at the cell density
+    vol_frac[i] = mass_frac[i];
+    densities[i] = mass_frac[i] / spvol_bulk / vol_frac[i];
+    sies[i] = sie_bulk * mass_frac[i];
+    vfrac_sum += vol_frac[i];
+    // Initialize pressure and temperature arrays to zero
+    temperatures[i] = 0;
+    pressures[i] = 0;
+  }
+
+  // Copy values to device (when available)
+  const size_t bytes = num_eos * sizeof(Real);
+  Real *v_densities = (Real *)PORTABLE_MALLOC(bytes);
+  portableCopyToDevice(v_densities, densities.data(), bytes);
+  Real *v_vol_frac = (Real *)PORTABLE_MALLOC(bytes);
+  portableCopyToDevice(v_vol_frac, vol_frac.data(), bytes);
+  Real *v_sies = (Real *)PORTABLE_MALLOC(bytes);
+  portableCopyToDevice(v_sies, sies.data(), bytes);
+  Real *v_temperatures = (Real *)PORTABLE_MALLOC(bytes);
[…]
