### 2024-07-16 b03cb44
Remove portable copy and instead just output GetOnDevice() to device array
