### 2024-07-16 ae688b8
Consolidate ifdef statements
