### 2024-07-16 50edd22
Get on device and _then_ copy EOS array to device. Also free each lambda separately
