### 2024-07-16 10fb97a
Remove using namespace singularity and instead explicitly bring things in one by one
