### 2024-07-16 a27eecb
Clang format
