+++ test/test_closure_pte.cpp
@@ -118,12 +118,13 @@ bool run_PTE_from_state(const int num_eos, EOS *v_EOS, const Real spvol_bulk,
   bool pte_converged;
   constexpr size_t bool_bytes = 1 * sizeof(bool);
   bool *pte_converged_d = (bool *)PORTABLE_MALLOC(bool_bytes);
-  portableFor("Device execution of PTE Test", 0, 1, PORTABLE_LAMBDA(auto i) {
-    PTESolverRhoT<decltype(v_EOS), Real *, Real **> method(
-    num_eos, v_EOS, vfrac_sum, sie_bulk, v_densities, v_vol_frac, v_sies,
-    v_temperatures, v_pressures, lambdas, scratch);
-    pte_converged_d[0] = PTESolver(method);
-  });
+  portableFor(
+      "Device execution of PTE Test", 0, 1, PORTABLE_LAMBDA(auto i) {
+        PTESolverRhoT<decltype(v_EOS), Real *, Real **> method(
+            num_eos, v_EOS, vfrac_sum, sie_bulk, v_densities, v_vol_frac, v_sies,
+            v_temperatures, v_pressures, lambdas, scratch);
+        pte_converged_d[0] = PTESolver(method);
+      });
   portableCopyToHost(&pte_converged, pte_converged_d, bool_bytes);
 
   // Free temp memory
