### 2024-07-16 a825f1e
Change type of iteration variable
