### 2024-07-17 14a5303
Allocate lambda memory all at once and then use accessor to index into it
