### 2024-07-17 592c2d3
Free lambda memory, not the struct itself
