### 2024-07-17 #6805 @parikshitbajpai
> Just checking in, @snikolov3 @parikshitbajpai are there any questions regarding the package that we can answer?

Hi @junghans, sorry I was on travel and couldn't submit my review over the weekend. I think everything has been quite clear for me. I had a couple of very minor suggestions / typo fixes on the paper and documentation which I'll open an issue for once I'm back from travel tomorrow morning.
