### 2024-07-17 b68f524
Ensure Finalize() is called from the host and combine tests
