+++ test/test_closure_pte.cpp
@@ -43,7 +43,7 @@ using singularity::SpinerEOSDependsRhoT;
 using singularity::mix_impl::CacheAccessor;
 
 using EOS = singularity::Variant<IdealGas, ShiftedEOS<DavisProducts>, DavisProducts,
-                                 SpinerEOSDependsRhoT>;
+                                 DavisReactants, SpinerEOSDependsRhoT>;
 
 template <typename EOSArrT>
 EOS *copy_eos_arr_to_device(const int num_eos, EOSArrT eos_arr) {
@@ -58,13 +58,12 @@ EOS *copy_eos_arr_to_device(const int num_eos, EOSArrT eos_arr) {
   return v_EOS;
 }
 
-void finalize_eos_arr(const int num_eos, EOS *v_EOS) {
-  // Call Finalize on each EOS
-  for (auto i = 0; i < num_eos; i++) {
-    v_EOS[i].Finalize();
+template <typename EOSArrT>
+void finalize_eos_arr(EOSArrT eos_arr) {
+  // Call Finalize on each EOS on the host
+  for (auto eos : eos_arr) {
+    eos.Finalize();
   }
-  // Free EOS memory
-  PORTABLE_FREE(v_EOS);
 }
 
 template <typename ArrT>
@@ -141,22 +140,35 @@ bool run_PTE_from_state(const int num_eos, EOS *v_EOS, const Real spvol_bulk,
 
 SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
 
-  GIVEN("Three equations of state") {
-    // Set up the three EOS
-    constexpr int num_eos = 3;
-    std::array<EOS, num_eos> eos_arr;
+  GIVEN("Four equations of state") {
+    // A series of tests using some subset of these four EOS
+    constexpr int num_eos = 4;
+
     // Reference state
     constexpr Real P0 = 1.0e6; // 1 bar
     constexpr Real T0 = 296;   // K
     // Ideal gas air
     constexpr Real rho0_air = 1e-03; // g/cc
     constexpr Real Gruneisen_air = 0.4;
-    constexpr Real CV_air = P0 / rho0_air / (Gruneisen_air * T0);
-    eos_arr[0] = IdealGas(Gruneisen_air, CV_air);
+    constexpr Real Cv_air = P0 / rho0_air / (Gruneisen_air * T0);
+    air_eos = IdealGas(Gruneisen_air, Cv_air);
     // Spiner copper EOS
     constexpr int Cu_matid = 3337;
     const std::string eos_file = "../materials.sp5";
-    eos_arr[1] = SpinerEOSDependsRhoT(eos_file, Cu_matid);
+    copper_eos = SpinerEOSDependsRhoT(eos_file, Cu_matid);
+    // Davis Reactants EOS
+    constexpr rho0_DP = 1.890;
+    constexpr e0_DP = 0.;  // erg / g
+    constexpr P0_DP = 0.;  // microbar
+    constexpr T0_DP = 297; // K
+    constexpr A = 1.8 * std::sqrt(GPa); // sqrt(microbar)
+    constexpr B = 4.6;
+    constexpr C = 0.34;
+    constexpr G0 = 0.56;
+    constexpr Z = 0.0;
+    constexpr alpha = 0.4265;
+    constexpr Cv_DP = 0.001074 * MJ_per_kg; // erg / g / K
+    davis_r_eos = DavisReactants(rho0_DP, e0_DP, P0_DP, T0_DP, A, B, C, G0, Z, alpha, Cv_DP);
     // Davis Products EOS
     constexpr Real a = 0.798311;
     constexpr Real b = 0.58;
[…]
