### 2024-07-17 e59140f
Clang format
