### 2024-07-18 e9dbf6e
Make parameter constexpr
