+++ test/test_closure_pte.cpp
@@ -16,6 +16,7 @@
 #include <array>
 #include <string>
 #include <vector>
+#include <cmath>
 
 #ifndef CATCH_CONFIG_FAST_COMPILE
 #define CATCH_CONFIG_FAST_COMPILE
@@ -162,7 +163,7 @@ SCENARIO("Density-Temperature PTE Solver", "[PTE]") {
     constexpr Real e0_DP = 0.;               // erg / g
     constexpr Real P0_DP = 0.;               // microbar
     constexpr Real T0_DP = 297;              // K
-    constexpr Real A = 1.8 * std::sqrt(GPa); // sqrt(microbar)
+    constexpr Real A = 1.8 * 1.0e5; // 1.8 * sqrt(GPa) -> sqrt(microbar)
     constexpr Real B = 4.6;
     constexpr Real C = 0.34;
     constexpr Real G0 = 0.56;
