### 2024-07-18 f32ca90
Enforce convention that '*_TEST_*' variables are defined in the test CMakeLists and other '*_USE_*' or '*_BUILD_*' variables are in the main project
