### 2024-07-19 lanl/singularity-eos#391 @parikshitbajpai
JOSS-Review
Reviewing the Singularity-EOS paper for JOSS (https://github.com/openjournals/joss-reviews/issues/6805), I came across a couple of minor typos and a couple of suggestions

- [ ] I didn't find any example usage on the documentation website. While the `example` directory contains the examples, it would be nice to have them on the documentation site as well including instructions on running them. This is something that can be done in a future release and doesn't necessarily have to be done as part of the review.
- [ ] In the paper, would you consider reorganizing the *Statement of need* and *State of the field* sections? I think the first paragraph of the state of the field section very clearly states why Singularity-EOS is needed, i.e. lack of unified APIs and GUI-support in other models. I'm not sure if moving that line to the statement of need section is the best idea but maybe you can figure out the best way of doing it.
