### 2024-07-19 #6805 @parikshitbajpai
Hi @Yurlungur @junghans,

I opened an issue with my very minor comments (lanl/singularity-eos#391). You can ignore them if you feel that they are not required or don't add value.

The library is a great tool in my opinion and I appreciate the modularity and data portability features. I am wrapping up some last bullet points I have to verify and should have those done this weekend.
