### 2024-07-19 lanl/singularity-eos#391 @Yurlungur
> Also, as I was reviewing the paper, I noticed there's a few typos in the paper and documents, e.g. `descriptions multi-phase descriptions`:
>
> https://github.com/lanl/singularity-eos/blob/91ee6924a607899a65700aacc94e7a580b2bd5e1/joss-paper/paper.md?plain=1#L113
>
> I have made these changes locally as I was reviewing so please let me know if you prefer I open a PR with those changes, send you a patch or list them out here so you can fix those?

Thanks so much, @parikshitbajpai ! A PR would be very welcome. :) However, I'm happy to use another method if you prefer.
