### 2024-07-19 lanl/singularity-eos#391 @parikshitbajpai
> Thanks so much, @parikshitbajpai ! A PR would be very welcome. :) However, I'm happy to use another method if you prefer.

For sure, I can submit the PR tonight or tomorrow.
