### 2024-07-19 7d35c43
Don't call GetOnDevice() twice
