### 2024-07-20 fd07bdd
Fix array out of bounds issues when confusing num_pte and num_eos
