### 2024-07-20 6de9259
Remove unused num_eos
