### 2024-07-22 db8b50b
Fix typos in contributing doco
