### 2024-07-22 lanl/singularity-eos#391 @Yurlungur
@parikshitbajpai

> In the paper, would you consider reorganizing the Statement of need and State of the field sections? I think the first paragraph of the state of the field section very clearly states why Singularity-EOS is needed, i.e. lack of unified APIs and GUI-support in other models. I'm not sure if moving that line to the statement of need section is the best idea but maybe you can figure out the best way of doing it.

Looking at the paper again with fresh eyes I almost wonder if switching the statement of need and state of the field sections might work. This leaves the definition of an equation of state until later in the paper, but it frontloads the reason to use `singularity-EOS`. What do you think?
