### 2024-07-22 lanl/singularity-eos#391 @Yurlungur
Alternatively, looking at the JOSS documentation, it seems that I could simply fuse the statement of need and state of the field sections into a single statement of need section.
