### 2024-07-22 454039a
Remove asserts in Ideal calculation in favor of a positive temperature check to make sure the guess is valid
