### 2024-07-23 lanl/singularity-eos#391 @parikshitbajpai
> @parikshitbajpai please take a look at #394 and let me know what you think.

Thanks for adding the doco. It looks great to me and addresses both the points I raised.
