### 2024-07-23 fca9fdc
typo in build matrix
