### 2024-07-23 #6805 @Yurlungur
> Hi @Yurlungur @junghans,
>
> I opened an issue with my very minor comments ([lanl/singularity-eos#391](https://github.com/lanl/singularity-eos/issues/391)). You can ignore them if you feel that they are not required or don't add value.
>
> The library is a great tool in my opinion and I appreciate the modularity and data portability features. I am wrapping up some last bullet points I have to verify and should have those done this weekend.

Thanks for the review @parikshitbajpai ! It seems like all issues are now addressed on your end?

@snikolov3 @kyleniemeyer just to check in. Do we have a time line for the second review and acceptance? Thanks!
