### 2024-07-24 a51e9a3
update cmakelists for release.
