### 2024-07-25 e97aec7
Update new dependencies for the 1.9.0 release
