### 2024-07-25 ed95ee0
Make plugins depend on 1.9.0 and above
