### 2024-07-29 d3b5ad1
move bounds into singularity-eos. rewrite it for piecewise grids and begin testing.
