+++ singularity-eos/base/table_bounds.hpp
@@ -1,0 +1,138 @@
+//------------------------------------------------------------------------------
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
+// program was produced under U.S. Government contract 89233218CNA000001
+// for Los Alamos National Laboratory (LANL), which is operated by Triad
+// National Security, LLC for the U.S.  Department of Energy/National
+// Nuclear Security Administration. All rights in the program are
+// reserved by Triad National Security, LLC, and the U.S. Department of
+// Energy/National Nuclear Security Administration. The Government is
+// granted for itself and others acting on its behalf a nonexclusive,
+// paid-up, irrevocable worldwide license in this material to reproduce,
+// prepare derivative works, distribute copies to the public, perform
+// publicly and display publicly, and to permit others to do so.
+//------------------------------------------------------------------------------
+
+#ifndef SINGULARITY_EOS_BASE_TABLE_BOUNDS_HPP_
+#define SINGULARITY_EOS_BASE_TABLE_BOUNDS_HPP_
+#ifdef SINGULARITY_USE_SPINER
+
+#include <cmath>
+#include <iostream>
+#include <limits>
+#include <string>
+#include <vector>
+
+#include <ports-of-call/portability.hpp>
+#include <singularity-eos/base/fast-math/logs.hpp>
+#include <spiner/databox.hpp>
+#include <spiner/interpolation.hpp>
+
+namespace singularity {
+
+// For logarithmic interpolation, quantities may be negative.
+// If they are, use offset to ensure negative values make sense.
+template<int NGRIDS = 3>
+class Bounds {
+ public:
+  using RegularGrid1D = Spiner::RegularGrid1D<Real>;
+  using Grid_t = Spiner::PiecewiseGrid1D<Real, NGRIDS>;
+
+  Bounds() {}
+
+  Bounds(Real min, Real max, int N, Real offset)
+      : grid(Grid_t(std::vector<RegularGrid1D>{RegularGrid1D(min, max, N)})),
+        offset(offset), piecewise(false) {}
+
+  Bounds(Real min, Real max, int N, bool convertToLog = false, Real shrinkRange = 0,
+         Real anchor_point = std::numeric_limits<Real>::signaling_NaN())
+      : offset(0), piecewise(true) {
+    if (convertToLog) {
+      convertBoundsToLog_(min, max, shrinkRange);
+      if (!(std::isnan(anchor_point))) {
+        anchor_point += offset;
+        anchor_point = singularity::FastMath::log10(std::abs(anchor_point));
+      }
+    }
+    if (!(std::isnan(anchor_point))) {
+      adjustForAnchor_(min, max, N, anchor_point);
+    }
+    grid = Grid_t(std::vector<RegularGrid1D>{RegularGrid1D(min, max, N)});
+  }
+
+  Bounds(Real global_min, Real global_max, Real anchor_point, Real log_fine_diameter,
+         int N_fine, Real N_factor, Real shrinkRange = 0)
+      : offset(0), piecewise(true) {
+    convertBoundsToLog_(global_min, global_max, shrinkRange);
+    anchor_point += offset;
+    anchor_point = singularity::FastMath::log10(std::abs(anchor_point));
+
+    int N_coarse = ((N_factor >= 1) && (N_fine > N_factor))
+                       ? static_cast<int>(std::ceil(N_fine / N_factor))
+                       : N_fine;
+
+    Real mid_min = anchor_point - 0.5 * log_fine_diameter;
[…]
