### 2024-07-30 916f450
Initial commit for StiffGas python interface
