### 2024-07-30 f2502d3
add table bounds tests
