### 2024-07-31 342837e
Add NobleAble to module.cpp for python interfacing.
