### 2024-07-31 39391be
modified bounds to support 2 and 3 grid modes.
