### 2024-07-31 b6c2589
fix type issues in compare_to_eospac
