+++ test/compare_to_eospac.cpp
@@ -64,7 +64,7 @@ class Bounds {
   Bounds() {}
 
   Bounds(Real min, Real max, int N, Real offset)
-      : grid(RegularGrid1D<Real>(min, max, N)), offset(offset) {}
+      : grid(RegularGrid1D(min, max, N)), offset(offset) {}
 
   Bounds(Real min, Real max, int N) : offset(0) {
     constexpr Real epsilon = std::numeric_limits<float>::epsilon();
@@ -74,7 +74,7 @@ class Bounds {
     max += offset;
     min = std::log10(std::abs(min));
     max = std::log10(std::abs(max));
-    grid = RegularGrid1D<Real>(min, max, N);
+    grid = RegularGrid1D(min, max, N);
   }
 
   PORTABLE_INLINE_FUNCTION Real log2lin(Real xl) const { return pow(10., xl) - offset; }
@@ -89,7 +89,7 @@ class Bounds {
   }
 
  public:
-  RegularGrid1D<Real> grid;
+  RegularGrid1D grid;
   Real offset;
 };
 
@@ -149,59 +149,59 @@ int main(int argc, char *argv[]) {
 
     constexpr int NT = 3;
     EOS_INTEGER nXYPairs = (nFineRho) * (nFineT);
-    DataBox<Real> xVals(nFineRho, nFineT);
-    DataBox<Real> yVals(nFineRho, nFineT);
-    DataBox<Real> vars(nFineRho, nFineT);
-    DataBox<Real> dx(nFineRho, nFineT);
-    DataBox<Real> dy(nFineRho, nFineT);
-    DataBox<Real> pressEOSPAC(nFineRho, nFineT);
-    DataBox<Real> pressDiff_h(nFineRho, nFineT);
-    DataBox<Real> pressDiff_d(nFineRho, nFineT);
+    DataBox xVals(nFineRho, nFineT);
+    DataBox yVals(nFineRho, nFineT);
+    DataBox vars(nFineRho, nFineT);
+    DataBox dx(nFineRho, nFineT);
+    DataBox dy(nFineRho, nFineT);
+    DataBox pressEOSPAC(nFineRho, nFineT);
+    DataBox pressDiff_h(nFineRho, nFineT);
+    DataBox pressDiff_d(nFineRho, nFineT);
 
 #ifdef PORTABILITY_STRATEGY_KOKKOS
     using RView = Kokkos::View<Real *>;
     RView pressSpinerDView("pressSpinerDevice", nFineRho * nFineT);
     typename RView::HostMirror pressSpinerHView =
         Kokkos::create_mirror_view(pressSpinerDView);
-    DataBox<Real> pressSpiner_d(pressSpinerDView.data(), nFineRho, nFineT);
-    DataBox<Real> pressSpiner_hm(pressSpinerHView.data(), nFineRho, nFineT);
+    DataBox pressSpiner_d(pressSpinerDView.data(), nFineRho, nFineT);
+    DataBox pressSpiner_hm(pressSpinerHView.data(), nFineRho, nFineT);
 #else
-    DataBox<Real> pressSpiner_d(nFineRho, nFineT);
-    DataBox<Real> pressSpiner_hm = pressSpiner_d.slice(2, 0, nFineRho);
+    DataBox pressSpiner_d(nFineRho, nFineT);
+    DataBox pressSpiner_hm = pressSpiner_d.slice(2, 0, nFineRho);
 #endif
 
-    DataBox<Real> pressSpiner_h(nFineRho, nFineT);
-    DataBox<Real> tempSpiner_h(nFineRho, nFineT);
-    DataBox<Real> tempSpinerE_h(nFineRho, nFineT);
-    DataBox<Real> tempEOSPAC(nFineRho, nFineT);
[…]
