### 2024-07-31 f80bf9f
num points from PPD took a second log
