+++ sesame2spiner/generate_files.cpp
@@ -242,7 +242,7 @@ void getMatBounds(int i, int matid, const SesameMetadata &metadata, const Params
   Real rhoMin = params.Get("rhomin", TinyShift(metadata.rhoMin, 1));
   Real rhoMax = params.Get("rhomax", metadata.rhoMax);
   Real TMin = params.Get("Tmin", TinyShift(metadata.TMin, 1));
-  Real TMax = params.Get("Tmax", metadata.TMax);
+  Real TMax = params.Get("Tmax", metadata.TMax); 
   Real sieMin = params.Get("siemin", TinyShift(metadata.sieMin, 1));
   Real sieMax = params.Get("siemax", metadata.sieMax);
 
@@ -253,9 +253,9 @@ void getMatBounds(int i, int matid, const SesameMetadata &metadata, const Params
   checkValInMatBounds(matid, "sieMin", sieMin, metadata.sieMin, metadata.sieMax);
   checkValInMatBounds(matid, "sieMax", sieMax, metadata.sieMin, metadata.sieMax);
 
-  Real shrinklRhoBounds = params.Get("shrinklRhoBounds", 0.0);
-  Real shrinklTBounds = params.Get("shrinklTBounds", 0.0);
-  Real shrinkleBounds = params.Get("shrinkleBounds", 0.0);
+  Real shrinklRhoBounds = params.Get("shrinklRhoBounds", 0.);
+  Real shrinklTBounds = params.Get("shrinklTBounds", 0.);
+  Real shrinkleBounds = params.Get("shrinkleBounds", 0.);
 
   shrinklRhoBounds = std::min(1., std::max(shrinklRhoBounds, 0.));
   shrinklTBounds = std::min(1., std::max(shrinklTBounds, 0.));
@@ -287,16 +287,20 @@ void getMatBounds(int i, int matid, const SesameMetadata &metadata, const Params
   int numT = params.Get("numT", numTDefault);
   int numSie = params.Get("numsie", numSieDefault);
 
+  const Real TAnchor = 298.15;
   const Real rhoAnchor = metadata.normalDensity;
-  constexpr Real TAnchor = 298.15;
 
   // Piecewise grids stuff
-  bool do_piecewise_grids = params.Get("piecewise", true);
-  Real ppd_factor_rho = params.Get("rhoCoarseFactor", COARSE_FACTOR_DEFAULT);
-  Real ppd_factor_T = params.Get("TCoarseFactor", COARSE_FACTOR_DEFAULT);
-  Real ppd_factor_sie = params.Get("sieCoarseFactor", COARSE_FACTOR_DEFAULT);
-  Real rho_fine_diameter = params.Get("rhoFineDiameterDecades", 0.5);
-  Real TSplitPoint = params.Get("TSplitPoint", 1e4);
+  bool piecewiseRho = params.Get("piecewiseRho", true);
+  bool piecewiseT = params.Get("piecewiseT", true);
+  bool piecewiseSie = params.Get("piecewiseSie", true);
+
+  Real ppd_factor_rho_lo = params.Get("rhoCoarseFactorLo", COARSE_FACTOR_DEFAULT_RHO_LO);
+  Real ppd_factor_rho_hi = params.Get("rhoCoarseFactorHi", COARSE_FACTOR_DEFAULT_RHO_HI);
+  Real ppd_factor_T = params.Get("TCoarseFactor", COARSE_FACTOR_DEFAULT_T);
+  Real ppd_factor_sie = params.Get("sieCoarseFactor", COARSE_FACTOR_DEFAULT_T);
+  Real rho_fine_diameter = params.Get("rhoFineDiameterDecades", RHO_FINE_DIAMETER_DEFAULT);
[…]
