### 2024-08-01 9825064
Fixed an shortened argumentlist. Implemented Carl Greeff's rate equation for mass fraction update and added a test case. This is the first ingredient of the Kinetic PhaseTransiton (KPT) scheme.
