+++ test/test_kpt_models.cpp
@@ -1,0 +1,141 @@
+//------------------------------------------------------------------------------
+// © 2024. Triad National Security, LLC. All rights reserved.  This
+// program was produced under U.S. Government contract 89233218CNA000001
+// for Los Alamos National Laboratory (LANL), which is operated by Triad
+// National Security, LLC for the U.S.  Department of Energy/National
+// Nuclear Security Administration. All rights in the program are
+// reserved by Triad National Security, LLC, and the U.S. Department of
+// Energy/National Nuclear Security Administration. The Government is
+// granted for itself and others acting on its behalf a nonexclusive,
+// paid-up, irrevocable worldwide license in this material to reproduce,
+// prepare derivative works, distribute copies to the public, perform
+// publicly and display publicly, and to permit others to do so.
+//------------------------------------------------------------------------------
+
+#include <array>
+#include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#ifndef CATCH_CONFIG_FAST_COMPILE
+#define CATCH_CONFIG_FAST_COMPILE
+#include <catch2/catch_test_macros.hpp>
+#endif
+
+#include <ports-of-call/portability.hpp>
+#include <ports-of-call/portable_arrays.hpp>
+#include <ports-of-call/portable_errors.hpp>
+#include <singularity-eos/base/constants.hpp>
+#include <singularity-eos/closure/kinetic_phasetransition_models.hpp>
+#include <test/eos_unit_test_helpers.hpp>
+
+using singularity::LogRatesCGModel;
+
+SCENARIO("First log rate test") {
+  GIVEN("Parameters for a 5 phase system") {
+    constexpr Real w[25] = {0.,  0.8510E+01, 20., 20., 20., 0.8510E+01, 0.,  20., 20.,
+                            20., 20.,        20., 0.,  20., 20.,        20., 20., 20.,
+                            0.,  20.,        20., 20., 20., 20.,        0.};
+
+    constexpr Real b[25] = {0.,         0.3060E-04, 0.1000E-05, 0.1000E-05, 0.1000E-05,
+                            0.3060E-04, 0.,         0.1000E-05, 0.1000E-05, 0.1000E-05,
+                            0.1000E-05, 0.1000E-05, 0.,         0.1000E-05, 0.1000E-05,
+                            0.1000E-05, 0.1000E-05, 0.1000E-05, 0.,         0.1000E-05,
+                            0.1000E-05, 0.1000E-05, 0.1000E-05, 0.1000E-05, 0.};
+
+    GIVEN("Gibbs free energy and order") {
+      constexpr int num = 5;
+      constexpr int mnum = num * (num - 1) / 2;
+      constexpr int knum = 5;
+#ifdef PORTABILITY_STRATEGY_KOKKOS
+      // Create Kokkos views on device for the input arrays
+      Kokkos::View<Real[num]> v_gibbs("gibbs");
+      // order phases from high to low gibbs
+      Kokkos::View<int[num]> v_order("order");
+#else
+      // Otherwise just create arrays to contain values and create pointers to
+      // be passed to the functions in place of the Kokkos views
+      std::array<Real, num> gibbs;
+      std::array<int, num> order;
+      auto v_gibbs = gibbs.data();
+      auto v_order = order.data();
+#endif // PORTABILITY_STRATEGY_KOKKOS
+
+      // Populate the input arrays
+      portableFor(
[…]
