+++ singularity-eos/closure/kinetic_phasetransition_models.hpp
@@ -33,28 +33,31 @@
 
 namespace singularity {
 
-
 PORTABLE_INLINE_FUNCTION void LogRatesCGModel(const Real *w, const Real *b,
-    const int num_phases, const Real *gibbs, const int *gibbsorder, Real *logRjk, int *fromto) {
+                                              const int num_phases, const Real *gibbs,
+                                              const int *gibbsorder, Real *logRjk,
+                                              int *fromto) {
 
-// begin the calculation of logRik
-    int jk = 0;
-    for (int j = 0; j < num_phases-1; j++){       
-        // from high Gibb phase to low
-        for (int k = num_phases-1; k > j; k--){ 
-	  // from low Gibb phase to high, k>j always, gibbs_j > gibbs_k always
-          // always mass transport from j->k
-          Real hgj=gibbs[gibbsorder[j]];
-          Real hgk=gibbs[gibbsorder[k]];
-// calc phase dx from hgi to hgk
-          Real dgdb=(hgj-hgk)/(b[gibbsorder[j]*num_phases + gibbsorder[k]]);
-	  // First is highest level to levels below, largest gibbs energy diff first. Then follows 2nd highest to levels below.
-          logRjk[jk] = log(w[gibbsorder[j]*num_phases + gibbsorder[k]]*dgdb)+dgdb*dgdb;
-	  fromto[jk] = gibbsorder[j]*10+gibbsorder[k];
-	  jk++;
-        }
+  // begin the calculation of logRik
+  int jk = 0;
+  for (int j = 0; j < num_phases - 1; j++) {
+    // from high Gibb phase to low
+    for (int k = num_phases - 1; k > j; k--) {
+      // from low Gibb phase to high, k>j always, gibbs_j > gibbs_k always
+      // always mass transport from j->k
+      Real hgj = gibbs[gibbsorder[j]];
+      Real hgk = gibbs[gibbsorder[k]];
+      // calc phase dx from hgi to hgk
+      Real dgdb = (hgj - hgk) / (b[gibbsorder[j] * num_phases + gibbsorder[k]]);
+      // First is highest level to levels below, largest gibbs energy diff first. Then
+      // follows 2nd highest to levels below.
+      logRjk[jk] =
+          log(w[gibbsorder[j] * num_phases + gibbsorder[k]] * dgdb) + dgdb * dgdb;
+      fromto[jk] = gibbsorder[j] * 10 + gibbsorder[k];
+      jk++;
     }
-    return; 
+  }
+  return;
 }
 
 } // namespace singularity
