+++ python/module.cpp
@@ -34,11 +34,11 @@ PYBIND11_MODULE(singularity_eos, m) {
 
   eos_class<StiffGas>(m, "StiffGas")
   .def(py::init())
-  .def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("q"), py:;arg("Pinf"));
+  .def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("Pinf"), py:arg("q"));
 
-  eos_class<NobleAble>(m, "NobleAble")
-  .def(py::init())
-  .def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("b"), py::arg("q"))
+  //eos_class<NobleAble>(m, "NobleAble")
+  //.def(py::init())
+  //.def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("b"), py::arg("q"))
 
   eos_class<Gruneisen>(m, "Gruneisen")
     .def(py::init())
