### 2024-08-06 9c35a24
Update changelog for internal energy scaling fix
