### 2024-08-06 79ce8c9
add robustness for refined region out of table bounds or normal density ill defined
