### 2024-08-06 1dbb8fe
chad comments part 1
