### 2024-08-06 1374d7f
add ability to set rhoFineMin, rhoFineMax
