+++ singularity-eos/base/table_bounds.hpp
@@ -71,36 +71,18 @@ class Bounds {
       : Bounds(min, max, N, convertToLog, shrinkRange, anchor_point) {}
 
   Bounds(TwoGrids, Real global_min, Real global_max, Real anchor_point, Real splitPoint,
-         Real ppd_fine, Real ppd_factor, Real shrinkRange = 0)
+         Real ppd_fine, Real ppd_factor, Real shrinkRange = 0, bool convertToLog = true)
       : offset(0), piecewise(true), linmin_(global_min), linmax_(global_max) {
     const Real ppd_coarse = (ppd_factor > 0) ? ppd_fine / ppd_factor : ppd_fine;
 
-    convertBoundsToLog_(global_min, global_max, shrinkRange);
-    anchor_point += offset;
-    anchor_point = singularity::FastMath::log10(std::abs(anchor_point));
-    splitPoint += offset;
-    splitPoint = singularity::FastMath::log10(std::abs(splitPoint));
+    if (convertToLog) {
+      convertBoundsToLog_(global_min, global_max, shrinkRange);
+      toLog_(anchor_point, offset);
+      toLog_(splitPoint, offset);
+    }
 
-    if (splitPoint <= global_min) {
-      PORTABLE_ALWAYS_WARN("Split point less than global minimum. Adjusting.");
-      Real eps = 0.1 * std::abs(global_min);
-      splitPoint = global_min + eps;
-    }
-    if (splitPoint >= global_max) {
-      PORTABLE_ALWAYS_WARN("Split point greater than global maximum. Adjusting.");
-      Real eps = 0.1 * std::abs(global_max);
-      splitPoint = global_max - eps;
-    }
-    if (anchor_point <= global_min) {
-      PORTABLE_ALWAYS_WARN("Anchor point less than global minimum. Adjusting.");
-      Real eps = 0.1 * std::abs(global_min);
-      anchor_point = global_min + eps;
-    }
-    if (anchor_point >= splitPoint) {
-      PORTABLE_ALWAYS_WARN("Anchor point greater than split point. Adjusting.");
-      Real eps = 0.1 * std::abs(splitPoint);
-      anchor_point = splitPoint - eps;
-    }
+    checkInterval_(splitPoint, global_min, global_max, "Split point");
+    checkInterval_(anchor_point, global_min, splitPoint, "Anchor point");
 
     // add a point just to make sure we have enough points after adjusting for anchor
     int N_fine = getNumPointsFromDensity(global_min, splitPoint, ppd_fine) + 1;
@@ -113,56 +95,39 @@ class Bounds {
     grid = Grid_t(std::vector<RegularGrid1D>{grid_lower, grid_upper});
   }
 
+  Bounds(ThreeGrids, Real global_min, Real global_max, Real anchor_point, Real fine_min,
+         Real fine_max, Real ppd_fine, Real ppd_factor_lo, Real ppd_factor_hi,
+         Real shrinkRange = 0, bool convertToLog = true)
+      : offset(0), piecewise(true), linmin_(global_min), linmax_(global_max) {
+
+    if (convertToLog) {
+      convertBoundsToLog_(global_min, global_max, shrinkRange);
+      toLog_(anchor_point, offset);
+      toLog_(fine_min, offset);
+      toLog_(fine_max, offset);
+    }
+    checkInterval_(anchor_point, global_min, global_max, "Anchor point");
+
+    grid = gridFromIntervals_(ThreeGrids, global_min, global_max, anchor_point, fine_min,
+                              fine_max, ppd_fine, pdd_factor_lo, ppd_factor_hi);
+  }
+
[…]
