### 2024-08-06 ac0db6e
add diagram for mesh. fix some build issues with the docs.
