+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -310,7 +310,7 @@ class PTESolverBase {
 
     // intialize rhobar array and final density
     InitRhoBarandRho();
-    uscale = rho_total * abs(sie_total);
+    uscale = rho_total * sie_total;
 
     // guess some non-zero temperature to start
     const Real Tguess = GetTguess();
@@ -328,11 +328,8 @@ class PTESolverBase {
     }
 
     // note the scaling of the material internal energy densities
-    uscale_total = 0.0;
-    for (int m = 0; m < nmat; ++m) {
+    for (int m = 0; m < nmat; ++m)
       u[m] = sie[m] * robust::ratio(rhobar[m], uscale);
-      uscale_total += u[m];
-    }
   }
 
   PORTABLE_INLINE_FUNCTION
@@ -455,7 +452,7 @@ class PTESolverBase {
   const RealIndexer &press;
   Real *jacobian, *dx, *sol_scratch, *residual, *u, *rhobar;
   CacheAccessor Cache;
-  Real rho_total, uscale, uscale_total, Tnorm;
+  Real rho_total, uscale, Tnorm;
 };
 
 } // namespace mix_impl
@@ -542,7 +539,6 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::press;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::rho_total;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::uscale;
-  using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::uscale_total;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::MatIndex;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::TryIdealPTE;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::jacobian;
@@ -594,7 +590,8 @@ class PTESolverRhoT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
       esum += u[m];
     }
     residual[0] = vfrac_total - vsum;
-    residual[1] = uscale_total - esum;
+    // the 1 here is the scaled total internal energy density
+    residual[1] = 1.0 - esum;
     for (int m = 0; m < nmat - 1; ++m) {
       residual[2 + m] = press[m + 1] - press[m];
     }
@@ -769,7 +766,6 @@ class PTESolverFixedT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::press;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::rho_total;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::uscale;
-  using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::uscale_total;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::MatIndex;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::jacobian;
   using mix_impl::PTESolverBase<EOSIndexer, RealIndexer>::residual;
@@ -821,11 +817,9 @@ class PTESolverFixedT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> 
       // note the scaling of pressure
       press[m] = eos[m].PressureFromDensityTemperature(rho[m], Tequil, Cache[m]);
     }
-    uscale_total = 0.0;
     for (int m = 0; m < nmat; ++m) {
[…]
