### 2024-08-08 69a7949
Added sorting routine for Gibbs free energy and included it in test.
