+++ test/test_kpt_models.cpp
@@ -25,10 +25,12 @@
 #include <ports-of-call/portable_arrays.hpp>
 #include <ports-of-call/portable_errors.hpp>
 #include <singularity-eos/base/constants.hpp>
+#include <singularity-eos/closure/kinetic_phasetransition_methods.hpp>
 #include <singularity-eos/closure/kinetic_phasetransition_models.hpp>
 #include <test/eos_unit_test_helpers.hpp>
 
 using singularity::LogRatesCGModel;
+using singularity::SortGibbs;
 
 SCENARIO("First log rate test") {
   GIVEN("Parameters for a 5 phase system") {
@@ -45,7 +47,6 @@ SCENARIO("First log rate test") {
     GIVEN("Gibbs free energy and order") {
       constexpr int num = 5;
       constexpr int mnum = num * (num - 1) / 2;
-      constexpr int knum = 5;
 #ifdef PORTABILITY_STRATEGY_KOKKOS
       // Create Kokkos views on device for the input arrays
       Kokkos::View<Real[num]> v_gibbs("gibbs");
@@ -68,11 +69,6 @@ SCENARIO("First log rate test") {
             v_gibbs[2] = 0.0287409;
             v_gibbs[3] = 1.0e99;
             v_gibbs[4] = 0.02940076;
-            v_order[0] = 3;
-            v_order[1] = 4;
-            v_order[2] = 0;
-            v_order[3] = 2;
-            v_order[4] = 1;
           });
 #ifdef PORTABILITY_STRATEGY_KOKKOS
       // Create host-side mirrors of the inputs and copy the inputs. These are
@@ -89,6 +85,27 @@ SCENARIO("First log rate test") {
           1.00000000000000007e+210, 4.99943400447804946e+05,  4.35424707359967171e+05,
           1.07530324485867186e+04,  3.93959978909504514e+02,  3.09367756860215100e+05,
           2.23469012616621285e+03};
+      constexpr std::array<int, num> phaseorder{0, 1, 2, 3, 4};
+      constexpr std::array<int, num> order_true{3, 4, 0, 2, 1};
+
+      WHEN("Using SortGibbs(num,gibbs,order) to give phases in order of largest to "
+           "smallest gibbs") {
+        SortGibbs(num, v_gibbs, v_order);
+
+        std::cout << "Order obtained with SortGibbs, largest to smallest Gibbs: "
+                  << std::endl;
+        for (int l = 0; l < num; l++) {
+          std::cout << "Phase " << v_order[l] << " Gibbs: " << v_gibbs[v_order[l]]
+                    << std::endl;
+        }
+#ifdef PORTABILITY_STRATEGY_KOKKOS
+        Kokkos::fence();
+        Kokkos::deep_copy(order, v_order);
+#endif // PORTABILITY_STRATEGY_KOKKOS
+        THEN("The returned order should be equal to the true order") {
+          array_compare(num, gibbs, phaseorder, order, order_true, "Gibbs", "phase");
+        }
+      }
       // create deltagibbs and fromphasetophase
 #ifdef PORTABILITY_STRATEGY_KOKKOS
       Kokkos::View<Real[mnum]> v_dgibbs("dgibbs");
@@ -127,6 +144,12 @@ SCENARIO("First log rate test") {
 
       WHEN("A LogRij(w,b,num,gibbs,order,fromto) lookup is performed") {
         LogRatesCGModel(w, b, num, v_gibbs, v_order, v_logrates, v_fromto);
+
+        std::cout << "LogRates obtained with LogRatesCGModel: " << std::endl;
+        for (int l = 0; l < mnum; l++) {
[…]
