### 2024-08-08 03a180a
Add test for negative internal energy fix
