### 2024-08-08 86232de
Use AND_THEN clause in new closure test
