### 2024-08-08 e2e167b
Make tol depend on Real type and add test for DensityEnergyFromPressureTemperature
