+++ singularity-eos/eos/eos_gruneisen.hpp
@@ -402,9 +402,9 @@ PORTABLE_INLINE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
     const Real press, const Real temp, Indexer_t &&lambda, Real &rho, Real &sie) const {
   // Energy is not a function of density
   sie = InternalEnergyFromDensityTemperature(rho0, temp)
-  // We have a branch at rho0, so we need to decide, based on our pressure, whether we
-  // should be above or below rho0
-  Real Pref = PressureFromDensityInternalEnergy(_rho0, sie);
+      // We have a branch at rho0, so we need to decide, based on our pressure, whether we
+      // should be above or below rho0
+      Real Pref = PressureFromDensityInternalEnergy(_rho0, sie);
   if (press < Pref) {
     rho = (press - _P0 + _C0 * _C0 * _rho0) / (_C0 * _C0 + _G0 * sie);
   } else { // We are in compression; iterate
@@ -413,7 +413,8 @@ PORTABLE_INLINE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
     }
     using RootFinding1D::regula_falsi;
     using RootFinding1D::Status;
-    auto status = regula_falsi(P_residual, press, _rho0, 1.0e-5, 1.0e3, 1.0e-8, 1.0e-8, rho);
+    auto status =
+        regula_falsi(P_residual, press, _rho0, 1.0e-5, 1.0e3, 1.0e-8, 1.0e-8, rho);
     if (status != Status::SUCCESS) {
       // Root finder failed even though the solution was bracketed... this is an error
       EOS_ERROR("Gruneisen::DensityEnergyFromPressureTemperature: "
