### 2024-08-08 d15d8b0
debugged stiff gas in python module
