### 2024-08-08 8bb528f
Change bounds on root find to be actually physical
