### 2024-08-08 7587914
Removed NobleAble comments from module.cpp and added initialization testing for StiffGas in test/python_bindings.py
