+++ test/python_bindings.py
@@ -67,6 +67,9 @@ class EOS(unittest.TestCase):
     def testIdealGas(self):
         eos = singularity_eos.IdealGas(1,1)
 
+    def testStiffGase(self):
+        eos = singularity_eos.StiffGas(1,1,1)
+
     def testShiftedIdealGas(self):
         eos = singularity_eos.Shifted(singularity_eos.IdealGas(1,1),1)
 
