+++ python/module.cpp
@@ -36,10 +36,6 @@ PYBIND11_MODULE(singularity_eos, m) {
   .def(py::init())
   .def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("Pinf"), py::arg("q"));
 
-  //eos_class<NobleAble>(m, "NobleAble")
-  //.def(py::init())
-  //.def(py::init<Real, Real, Real, Real>(), py::arg("gm1"), py::arg("Cv"), py::arg("b"), py::arg("q"))
-
   eos_class<Gruneisen>(m, "Gruneisen")
     .def(py::init())
     .def(
