### 2024-08-08 16e4bc4
Removed wrong changes to eos_variant.hpp
