### 2024-08-08 64eb628
Fixed spelling mistake in StiffGas for python_bindings.py in test/
