### 2024-08-08 f67ff35
Update test/python_bindings.py
Co-authored-by: Jeff Peterson <83598606+jhp-lanl@users.noreply.github.com>
