### 2024-08-08 b4caf67
Update python_bindings.py
Update copyright
