### 2024-08-08 390a7ed
Make REfromPT iterate in expansion too
