+++ singularity-eos/eos/eos_gruneisen.hpp
@@ -403,24 +403,29 @@ PORTABLE_INLINE_FUNCTION void Gruneisen::DensityEnergyFromPressureTemperature(
   // We have a branch at rho0, so we need to decide, based on our pressure, whether we
   // should be above or below rho0
   Real Pref = PressureFromDensityTemperature(_rho0, temp);
+  Real rho_lower;
+  Real rho_upper;
+  // Pick bounds appropriate depending on whether in compression or expansion
   if (press < Pref) {
-    sie = InternalEnergyFromDensityTemperature(_rho0, temp);
-    rho = (press - _P0 + _C0 * _C0 * _rho0) / (_C0 * _C0 + _G0 * sie);
-  } else { // We are in compression; iterate
-    auto PofRatT = PORTABLE_LAMBDA(const Real r) {
-      return PressureFromDensityTemperature(r, temp);
-    };
-    using RootFinding1D::regula_falsi;
-    using RootFinding1D::Status;
-    auto status = regula_falsi(PofRatT, press, _rho0, 0.9 * _rho0,
-                               std::min(_rho_max, 1.0e4), 1.0e-8, 1.0e-8, rho);
-    if (status != Status::SUCCESS) {
-      // Root finder failed even though the solution was bracketed... this is an error
-      EOS_ERROR("Gruneisen::DensityEnergyFromPressureTemperature: "
-                "Root find failed to find a solution given P, T\n");
-    }
-    sie = InternalEnergyFromDensityTemperature(_rho0, temp);
+    rho_lower = 0.;
+    rho_upper = 1.1 * _rho0;
+  } else {
+    rho_lower = 0.9 * _rho0;
+    rho_upper = std::min(_rho_max, 1.0e4);
   }
+  auto PofRatT = PORTABLE_LAMBDA(const Real r) {
+    return PressureFromDensityTemperature(r, temp);
+  };
+  using RootFinding1D::regula_falsi;
+  using RootFinding1D::Status;
+  auto status = regula_falsi(PofRatT, press, _rho0, 0.9 * _rho0,
+                             std::min(_rho_max, 1.0e4), 1.0e-8, 1.0e-8, rho);
+  if (status != Status::SUCCESS) {
+    // Root finder failed even though the solution was bracketed... this is an error
+    EOS_ERROR("Gruneisen::DensityEnergyFromPressureTemperature: "
+              "Root find failed to find a solution given P, T\n");
+  }
+  sie = InternalEnergyFromDensityTemperature(rho, temp);
 }
 template <typename Indexer_t>
 PORTABLE_INLINE_FUNCTION void
