### 2024-08-09 3b1c6ff
Add ability to compute maximum stable density
