### 2024-08-09 5663445
Update aluminum parameters to trigger failure and adjust tests
