### 2024-08-11 dac88bc
Update doc/sphinx/src/models.rst
Co-authored-by: Jeff Peterson <83598606+jhp-lanl@users.noreply.github.com>
