### 2024-08-11 c103b51
Update sesame2spiner/generate_files.cpp
Co-authored-by: Jeff Peterson <83598606+jhp-lanl@users.noreply.github.com>
