+++ sesame2spiner/generate_files.cpp
@@ -287,7 +287,7 @@ void getMatBounds(int i, int matid, const SesameMetadata &metadata, const Params
   int numT = params.Get("numT", numTDefault);
   int numSie = params.Get("numsie", numSieDefault);
 
-  const Real TAnchor = 298.15;
+  constexpr Real TAnchor = 298.15;
   Real rhoAnchor = params.Get("rho_fine_center", metadata.normalDensity);
   if (std::isnan(rhoAnchor) || rhoAnchor <= 0 || rhoAnchor > 1e8) {
     std::cerr << "WARNING [" << matid << "] "
