### 2024-08-11 e0425fe
add const to piecewise vars
as suggested by @jhp-lanl
