### 2024-08-11 c1d2040
add a bunch of const calls
