+++ singularity-eos/base/table_bounds.hpp
@@ -26,11 +26,10 @@
 #include <ports-of-call/portability.hpp>
 #include <ports-of-call/portable_errors.hpp>
 #include <singularity-eos/base/fast-math/logs.hpp>
-#include <spiner/databox.hpp>
 #include <spiner/interpolation.hpp>
 
 namespace singularity {
-
+namespace table_utils {
 // For logarithmic interpolation, quantities may be negative.
 // If they are, use offset to ensure negative values make sense.
 template <int NGRIDS = 3>
@@ -45,12 +44,14 @@ class Bounds {
 
   Bounds() : offset(0), piecewise(false), linmin_(0), linmax_(0) {}
 
-  Bounds(Real min, Real max, int N, Real offset)
+  Bounds(const Real min, const Real max, const int N, const Real offset)
       : grid(Grid_t(std::vector<RegularGrid1D>{RegularGrid1D(min, max, N)})),
         offset(offset), piecewise(false), linmin_(min), linmax_(max) {}
-  Bounds(OneGrid, Real min, Real max, int N, Real offset) : Bounds(min, max, N, offset) {}
+  Bounds(OneGrid, const Real min, const Real max, const int N, const Real offset)
+      : Bounds(min, max, N, offset) {}
 
-  Bounds(Real min, Real max, int N, bool convertToLog = false, Real shrinkRange = 0,
+  Bounds(Real min, Real max, int N, const bool convertToLog = false,
+         const Real shrinkRange = 0,
          Real anchor_point = std::numeric_limits<Real>::signaling_NaN())
       : offset(0), piecewise(true), linmin_(min), linmax_(max) {
     if (convertToLog) {
@@ -65,13 +66,14 @@ class Bounds {
     }
     grid = Grid_t(std::vector<RegularGrid1D>{RegularGrid1D(min, max, N)});
   }
-  Bounds(OneGrid, Real min, Real max, int N, bool convertToLog = false,
-         Real shrinkRange = 0,
+  Bounds(OneGrid, Real min, Real max, int N, const bool convertToLog = false,
+         const Real shrinkRange = 0,
          Real anchor_point = std::numeric_limits<Real>::signaling_NaN())
       : Bounds(min, max, N, convertToLog, shrinkRange, anchor_point) {}
 
   Bounds(TwoGrids, Real global_min, Real global_max, Real anchor_point, Real splitPoint,
-         Real ppd_fine, Real ppd_factor, bool convertToLog, Real shrinkRange = 0)
+         const Real ppd_fine, const Real ppd_factor, const bool convertToLog,
+         const Real shrinkRange = 0)
       : offset(0), piecewise(true), linmin_(global_min), linmax_(global_max) {
     const Real ppd_coarse = (ppd_factor > 0) ? ppd_fine / ppd_factor : ppd_fine;
 
@@ -96,8 +98,8 @@ class Bounds {
   }
 
   Bounds(ThreeGrids, Real global_min, Real global_max, Real anchor_point, Real fine_min,
-         Real fine_max, Real ppd_fine, Real ppd_factor_lo, Real ppd_factor_hi,
-         bool convertToLog, Real shrinkRange = 0)
+         Real fine_max, const Real ppd_fine, const Real ppd_factor_lo,
+         const Real ppd_factor_hi, const bool convertToLog, const Real shrinkRange = 0)
       : offset(0), piecewise(true), linmin_(global_min), linmax_(global_max) {
 
     if (convertToLog) {
@@ -113,8 +115,8 @@ class Bounds {
   }
 
   Bounds(ThreeGrids, Real global_min, Real global_max, Real anchor_point,
-         Real log_fine_diameter, Real ppd_fine, Real ppd_factor_lo, Real ppd_factor_hi,
-         bool convertToLog, Real shrinkRange = 0)
[…]
