+++ test/test_bounds.cpp
@@ -39,10 +39,9 @@ constexpr Real T_split = 1e4;
 constexpr int N_per_decade_fine = 200;
 constexpr Real N_factor = 5;
 
-constexpr Real REAL_TOL =
-    std::numeric_limits<Real>::epsilon * 1e3
+constexpr Real REAL_TOL = std::numeric_limits<Real>::epsilon() * 1e3;
 
-    SCENARIO("Bounds can compute number of points from points per decade", "[Bounds]") {
+SCENARIO("Bounds can compute number of points from points per decade", "[Bounds]") {
   WHEN("We compute the number of points from points per decade") {
     int np = Bounds::getNumPointsFromPPD(T_min, T_max, N_per_decade_fine);
     constexpr int NDECADES = 9;
