+++ test/compare_to_eospac.cpp
@@ -47,7 +47,6 @@
 
 #include <singularity-eos/eos/eos.hpp>
 
-// using namespace Spiner;
 using namespace singularity;
 using namespace EospacWrapper;
 
