+++ test/test_bounds.cpp
@@ -17,7 +17,7 @@
 
 #include <ports-of-call/portability.hpp>
 #include <singularity-eos/base/fast-math/logs.hpp>
-#include <singularity-eos/base/table_bounds.hpp>
+#include <singularity-eos/base/spiner_table_bounds.hpp>
 
 #ifndef CATCH_CONFIG_FAST_COMPILE
 #define CATCH_CONFIG_FAST_COMPILE
