### 2024-08-13 a721e04
Adjust initial conditions
