### 2024-08-13 cc84d3c
start with example CI
