+++ .gitlab-ci.yml
@@ -6,329 +6,193 @@ stages:
   - prepare
   - build_n_test
 
-variables:
-  GIT_SUBMODULE_STRATEGY: normal
-  XCAP_PROJECT_DIR: "/usr/projects/xcap/oss"
-  TESTING_CLUSTER: darwin
-  SINGULARITY_GOLDFILES_VERSION: "goldfiles-1.8.0"
-  SINGULARITY_EOS_GCC_VERSION: "10.3.0"
-  SINGULARITY_EOS_OPENMPI_VERSION: "4.1.1"
-  SINGULARITY_EOS_CUDA_VERSION: "11.6.0"
-  SINGULARITY_EOS_BUILD_TYPE: "RelWithDebInfo"
-  SINGULARITY_EOS_CXX_STANDARD: 14
-  SINGULARITY_EOS_SPACK_SPEC: "singularity-eos@main+mpi+spiner+hdf5+python+tests%gcc@=${SINGULARITY_EOS_GCC_VERSION} ^openmpi@${SINGULARITY_EOS_OPENMPI_VERSION}"
-  COLOR_CYAN: "\e[1;36m"
-  COLOR_PLAIN: "\e[0m"
-  # uncomment to have the CI Spack installation for debugging
-  # PROJECT_TMP_CI_DIR: "${CI_PROJECT_DIR}"
-  # uncomment & set to different spack upstream for testing
-  # PROJECT_TMP_SPACK_DIR: "deployment/CI/spack-v0.20.1-8"
-  # uncomment for spack debug output
-  # PROJECT_SPACK_DEBUG_FLAG: "-d"
-
-workflow:
-  rules:
-    - if: $CI_PIPELINE_SOURCE == "merge_request_event"
-    - if: $CI_PIPELINE_SOURCE == "pipeline"
-    - if: $CI_PIPELINE_SOURCE == "schedule"
-    - if: $CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH
-    - if: $CI_COMMIT_TAG
-
 default:
-  tags:
-    - darwin-slurm-shared
+  interruptible: true
+  before_script:
+   - echo "Running on $(hostname)"
+   - section() { echo $'\e[0K'"section_$1:$(date +%s):$2"$'\r\e[0K'"${3+${COLOR_CYAN}$3${COLOR_PLAIN}}"; }
+
+variables:
+  ENABLED_CLUSTERS: "darwin"
+  PROJECT_SPACK_ENV: /usr/projects/xcap/spack-env/oss/2024-08-01
+  SINGULARITY_GOLDFILES_VERSION: "goldfiles-1.8.0"
+  GIT_SUBMODULE_STRATEGY: normal
+
+.ascgit_job:
   id_tokens:
     SITE_ID_TOKEN:
       aud: https://asc-git.lanl.gov
 
-##########################
-# Build Matrix Variables #
-##########################
-
-.buildtype_debug: &buildtype_debug
-  SINGULARITY_EOS_BUILD_TYPE: "Debug"
-
-.buildtype_release: &buildtype_release
-  SINGULARITY_EOS_BUILD_TYPE: "Release"
-  
-.buildtype_relDebugInfo: &buildtype_relDebugInfo
-  SINGULARITY_EOS_BUILD_TYPE: "RelWithDebInfo"
-
-.gpu: &gpu
-  SINGULARITY_USE_CUDA: "true"
[…]
