### 2024-08-14 69470b6
Error out when all pressures are unstable above the reference density
