### 2024-08-15 bc4b50a
It works for analytic EOSs and modifiers and the variant
