### 2024-08-20 073dbe5
add CheckParams as official part of API
