+++ singularity-eos/eos/modifiers/eos_unitsystem.hpp
@@ -24,6 +24,7 @@
 #include <utility>
 
 #include <ports-of-call/portability.hpp>
+#include <ports-of-call/portable_errors.hpp>
 #include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/eos_error.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
@@ -96,7 +97,9 @@ class UnitSystem : public EosBase<UnitSystem<T>> {
         inv_dpdr_unit_(rho_unit / press_unit_), inv_dtdr_unit_(rho_unit / temp_unit),
         inv_dtde_unit_(sie_unit / temp_unit) // obviously this is also Cv
         ,
-        inv_cv_unit_(temp_unit / sie_unit), inv_bmod_unit_(1 / press_unit_) {}
+        inv_cv_unit_(temp_unit / sie_unit), inv_bmod_unit_(1 / press_unit_) {
+    CheckParams();
+  }
   UnitSystem(T &&t, eos_units_init::LengthTimeUnitsInit, const Real time_unit,
              const Real mass_unit, const Real length_unit, const Real temp_unit)
       : UnitSystem(std::forward<T>(t), eos_units_init::thermal_units_init_tag,
@@ -106,6 +109,13 @@ class UnitSystem : public EosBase<UnitSystem<T>> {
       : UnitSystem(std::forward<T>(t), eos_units_init::thermal_units_init_tag, rho_unit,
                    sie_unit, temp_unit) {}
   UnitSystem() = default;
+
+  PORTABLE_INLINE_FUNCTION void CheckParams() const {
+    PORTABLE_ALWAYS_REQUIRE(rho_unit_ > 0, "Nonzero density unit");
+    PORTABLE_ALWAYS_REQUIRE(sie_unit_ > 0, "Nonzero energy unit");
+    PORTABLE_ALWAYS_REQUIRE(temp_unit_ > 0, "Nonzero temperature unit");
+    t_.CheckParams();
+  }
 
   auto GetOnDevice() {
     return UnitSystem<T>(t_.GetOnDevice(), eos_units_init::thermal_units_init_tag,
