### 2024-08-20 b11a8de
got a little paranoid about CRTP
