### 2024-08-20 8424317
serialization routines implemented in stellar collapse, with some nice machinery for looping over databoxes
