+++ singularity-eos/eos/eos_stellar_collapse.hpp
@@ -86,6 +86,12 @@ class StellarCollapse : public EosBase<StellarCollapse> {
   using EosBase<StellarCollapse>::GruneisenParamFromDensityTemperature;
   using EosBase<StellarCollapse>::GruneisenParamFromDensityInternalEnergy;
   using EosBase<StellarCollapse>::FillEos;
+  using EosBase<StellarCollapse>::SerializedSizeInBytes;
+  using EosBase<StellarCollapse>::Serialize;
+  using EosBase<StellarCollapse>::DeSerialize;
+  using EosBase<StellarCollapse>::IsModified;
+  using EosBase<StellarCollapse>::UnmodifyOnce;
+  using EosBase<StellarCollapse>::GetUnmodifiedObject;
 
   inline StellarCollapse(const std::string &filename, bool use_sp5 = false,
                          bool filter_bmod = true);
@@ -223,6 +229,10 @@ class StellarCollapse : public EosBase<StellarCollapse> {
   inline static void dataBoxToFastLogs(DataBox &db, DataBox &scratch,
                                        bool dependent_var_log);
 
+  std::size_t DynamicMemorySizeInBytes() const;
+  std::size_t DumpDynamicMemory(char *dst) const;
+  std::size_t SetDynamicMemory(char *src);
+
  private:
   class LogT {
    public:
@@ -356,6 +366,14 @@ class StellarCollapse : public EosBase<StellarCollapse> {
   // Bounds of dependent variables. Needed for root finding.
   DataBox eCold_, eHot_;
 
+  // TODO(JMM): Pointers here? or reference_wrapper? IMO the pointers are more clear
+#define DBLIST                                                                           \
+  &lP_, &lE_, &dPdRho_, &dPdE_, &dEdT_, &lBMod_, &entropy_, &Xa_, &Xh_, &Xn_, &Xp_,      \
+      &Abar_, &Zbar_, &mu_e_, &mu_n_, &mu_p_, &muhat_, &munu_, &eCold_, &eHot_
+  auto GetDataBoxPointers_() const { return std::vector<const DataBox *>{DBLIST}; }
+  auto GetDataBoxPointers_() { return std::vector<DataBox *>{DBLIST}; }
+#undef DBLIST
+
   // Independent variable bounds
   int numRho_, numT_, numYe_;
   Real lRhoMin_, lRhoMax_;
@@ -453,71 +471,52 @@ inline void StellarCollapse::Save(const std::string &filename) {
 }
 
 inline StellarCollapse StellarCollapse::GetOnDevice() {
-  StellarCollapse other;
-  other.lP_ = Spiner::getOnDeviceDataBox<Real>(lP_);
-  other.lE_ = Spiner::getOnDeviceDataBox<Real>(lE_);
-  other.dPdRho_ = Spiner::getOnDeviceDataBox<Real>(dPdRho_);
-  other.dPdE_ = Spiner::getOnDeviceDataBox<Real>(dPdE_);
-  other.dEdT_ = Spiner::getOnDeviceDataBox<Real>(dEdT_);
-  other.entropy_ = Spiner::getOnDeviceDataBox<Real>(entropy_);
-  other.Xa_ = Spiner::getOnDeviceDataBox<Real>(Xa_);
-  other.Xh_ = Spiner::getOnDeviceDataBox<Real>(Xh_);
-  other.Xn_ = Spiner::getOnDeviceDataBox<Real>(Xn_);
-  other.Xp_ = Spiner::getOnDeviceDataBox<Real>(Xp_);
[…]
