### 2024-08-20 f84ed1d
stellar collapse EOS has serialization now
