+++ test/test_eos_stellar_collapse.cpp
@@ -37,6 +37,75 @@
 
 #ifdef SPINER_USE_HDF
 #ifdef SINGULARITY_TEST_STELLAR_COLLAPSE
+template <typename EOS_t>
+void CompareStellarCollapse(EOS_t sc, EOS_t sc2) {
+  Real yemin = sc.YeMin();
+  Real yemax = sc.YeMax();
+  Real tmin = sc.TMin();
+  Real tmax = sc.TMax();
+  Real ltmin = std::log10(tmin);
+  Real ltmax = std::log10(tmax);
+  Real lrhomin = std::log10(sc.rhoMin());
+  Real lrhomax = std::log10(sc.rhoMax());
+  REQUIRE(yemin == sc2.YeMin());
+  REQUIRE(yemax == sc2.YeMax());
+  REQUIRE(sc.TMin() == sc2.TMin());
+  REQUIRE(sc.TMax() == sc2.TMax());
+  REQUIRE(isClose(lrhomin, std::log10(sc2.rhoMin()), 1e-12));
+  REQUIRE(isClose(lrhomax, std::log10(sc2.rhoMax()), 1e-12));
+
+  auto sc1_d = sc.GetOnDevice();
+  auto sc2_d = sc2.GetOnDevice();
+
+  int nwrong_h = 0;
+#ifdef PORTABILITY_STRATEGY_KOKKOS
+  using atomic_view = Kokkos::MemoryTraits<Kokkos::Atomic>;
+  Kokkos::View<int, atomic_view> nwrong_d("wrong");
+#else
+  PortableMDArray<int> nwrong_d(&nwrong_h, 1);
+#endif
+
+  const int N = 123;
+  constexpr Real gamma = 1.4;
+  const Real dY = (yemax - yemin) / (N + 1);
+  const Real dlT = (ltmax - ltmin) / (N + 1);
+  const Real dlR = (lrhomax - lrhomin) / (N + 1);
+  portableFor(
+      "fill eos", 0, N, 0, N, 0, N,
+      PORTABLE_LAMBDA(const int &k, const int &j, const int &i) {
+        Real lambda[2];
+        Real Ye = yemin + k * dY;
+        Real lT = ltmin + j * dlT;
+        Real lR = lrhomin + i * dlR;
+        Real T = std::pow(10., lT);
+        Real R = std::pow(10., lR);
+        Real e1, e2, p1, p2, cv1, cv2, b1, b2, s1, s2;
+        unsigned long output = (singularity::thermalqs::pressure |
+                                singularity::thermalqs::specific_internal_energy |
+                                singularity::thermalqs::specific_heat |
+                                singularity::thermalqs::bulk_modulus);
+        lambda[0] = Ye;
+
+        sc1_d.FillEos(R, T, e1, p1, cv1, b1, output, lambda);
+        sc2_d.FillEos(R, T, e2, p2, cv2, b2, output, lambda);
+        // Fill entropy. Will need to change later.
+        s1 = sc1_d.EntropyFromDensityTemperature(R, T, lambda);
+        s2 = p2 * std::pow(R, -gamma); // ideal
+        if (!isClose(e1, e2)) nwrong_d() += 1;
+        if (!isClose(p1, p2)) nwrong_d() += 1;
+        if (!isClose(cv1, cv2)) nwrong_d() += 1;
+        if (!isClose(b1, b2)) nwrong_d() += 1;
+        if (!isClose(s1, s2)) nwrong_d() += 1;
+      });
+#ifdef PORTABILITY_STRATEGY_KOKKOS
+  Kokkos::deep_copy(nwrong_h, nwrong_d);
+#endif
+  REQUIRE(nwrong_h == 0);
+
+  sc1_d.Finalize();
[…]
