### 2024-08-20 e6289d8
add GetDataBoxPointers_ to SpinerEOSDependsRhoT
