### 2024-08-20 d438ce2
spiner EOS compiles now
