### 2024-08-21 f7bb9fc
cleaned up base class and modifiers a little bit
