+++ singularity-eos/eos/modifiers/scaled_eos.hpp
@@ -36,32 +36,7 @@ using namespace eos_base;
 template <typename T>
 class ScaledEOS : public EosBase<ScaledEOS<T>> {
  public:
-  // Generic functions provided by the base class. These contain
-  // e.g. the vector overloads that use the scalar versions declared
-  // here We explicitly list, rather than using the macro because we
-  // overload some methods.
-
-  // TODO(JMM): The modifier EOS's should probably call the specific
-  // sub-functions of the class they modify so that they can leverage,
-  // e.g., an especially performant or special version of these
-  using EosBase<ScaledEOS<T>>::TemperatureFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::InternalEnergyFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::PressureFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::PressureFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::MinInternalEnergyFromDensity;
-  using EosBase<ScaledEOS<T>>::EntropyFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::EntropyFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::SpecificHeatFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::SpecificHeatFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::BulkModulusFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::BulkModulusFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::GruneisenParamFromDensityTemperature;
-  using EosBase<ScaledEOS<T>>::GruneisenParamFromDensityInternalEnergy;
-  using EosBase<ScaledEOS<T>>::FillEos;
-  using EosBase<ScaledEOS<T>>::SerializedSizeInBytes;
-  using EosBase<ScaledEOS<T>>::Serialize;
-  using EosBase<ScaledEOS<T>>::DeSerialize;
-
+  SG_ADD_BASE_CLASS_USINGS(ScaledEOS<T>);
   using BaseType = T;
 
   // give me std::format or fmt::format...
@@ -367,13 +342,8 @@ class ScaledEOS : public EosBase<ScaledEOS<T>> {
     return t_.MinimumTemperature();
   }
 
-  inline constexpr bool IsModified() const { return true; }
-
+  static inline constexpr bool IsModified() { return true; }
   inline constexpr T UnmodifyOnce() { return t_; }
-
-  inline constexpr decltype(auto) GetUnmodifiedObject() {
-    return t_.GetUnmodifiedObject();
-  }
 
   std::size_t DynamicMemorySizeInBytes() const { return t_.DynamicMemorySizeInBytes(); }
   std::size_t DumpDynamicMemory(char *dst) const { return t_.DumpDynamicMemory(dst); }
