### 2024-08-21 2abb8b8
thread SpinerTricks through a few models
