+++ singularity-eos/eos/eos_stellar_collapse.hpp
@@ -38,6 +38,7 @@
 #include <singularity-eos/base/robust_utils.hpp>
 #include <singularity-eos/base/root-finding-1d/root_finding.hpp>
 #include <singularity-eos/base/sp5/singularity_eos_sp5.hpp>
+#include <singularity-eos/base/spiner_table_utils.hpp>
 #include <singularity-eos/base/variadic_utils.hpp>
 #include <singularity-eos/eos/eos_base.hpp>
 
@@ -61,6 +62,7 @@ using namespace eos_base;
 // and introduce extrapolation as needed.
 class StellarCollapse : public EosBase<StellarCollapse> {
  public:
+  friend class table_utils::SpinerTricks<StellarCollapse>;
   using DataBox = Spiner::DataBox<Real>;
   using Grid_t = Spiner::RegularGrid1D<Real>;
 
@@ -467,55 +469,23 @@ inline void StellarCollapse::Save(const std::string &filename) {
 }
 
 inline StellarCollapse StellarCollapse::GetOnDevice() {
-  // trivially copy all but dynamic memory
-  StellarCollapse other = *this;
-  // set dynamic memory
-  auto other_dbs = other.GetDataBoxPointers_();
-  auto my_dbs = GetDataBoxPointers_();
-  int idb = 0;
-  for (DataBox *pother_db : other_dbs) {
-    DataBox *pmy_db = my_dbs[idb++];
-    *pother_db = Spiner::getOnDeviceDataBox<Real>(*pmy_db);
-  }
-  // set memory status
-  other.memoryStatus_ = DataStatus::OnDevice;
-  return other;
+  return table_utils::SpinerTricks<StellarCollapse>::GetOnDevice(this);
 }
 
 inline void StellarCollapse::Finalize() {
-  if (memoryStatus_ != DataStatus::UnManaged) {
-    for (DataBox *pdb : GetDataBoxPointers_()) {
-      pdb->finalize();
-    }
-  }
-  memoryStatus_ = DataStatus::Deallocated;
+  table_utils::SpinerTricks<StellarCollapse>::Finalize(this);
 }
 
 inline std::size_t StellarCollapse::DynamicMemorySizeInBytes() const {
-  std::size_t out = 0;
-  for (const DataBox *pdb : GetDataBoxPointers_()) {
-    out += pdb->sizeBytes();
-  }
-  return out;
+  return table_utils::SpinerTricks<StellarCollapse>::DynamicMemorySizeInBytes(this);
 }
 
 inline std::size_t StellarCollapse::DumpDynamicMemory(char *dst) const {
-  std::size_t offst = 0;
-  for (const DataBox *pdb : GetDataBoxPointers_()) {
-    std::size_t size = pdb->sizeBytes();
-    memcpy(dst + offst, pdb->data(), size);
-    offst += size;
-  }
-  return offst;
+  return table_utils::SpinerTricks<StellarCollapse>::DumpDynamicMemory(dst, this);
 }
 
 inline std::size_t StellarCollapse::SetDynamicMemory(char *src) {
-  std::size_t offst = 0;
-  for (DataBox *pdb : GetDataBoxPointers_()) {
-    offst += pdb->setPointer(src + offst);
-  }
-  memoryStatus_ = DataStatus::UnManaged;
-  return offst;
+  return table_utils::SpinerTricks<StellarCollapse>::SetDynamicMemory(src, this);
 }
 
 template <typename Indexer_t>
