### 2024-08-21 d8adfa3
add test for spiner tricks
