### 2024-08-21 3699b42
Add some sanity checks to base class. Fix bug in base class. Add tests that catch issues related to CRTP.
