### 2024-08-21 1505a33
add helmholtz tests
