### 2024-08-22 aa365d6
add remaining darwin jobs
