### 2024-08-22 0f0c5bb
eospac serialization. Still need to test. Time for Darwin.
