+++ singularity-eos/eos/modifiers/shifted_eos.hpp
@@ -357,7 +357,9 @@ class ShiftedEOS : public EosBase<ShiftedEOS<T>> {
 
   std::size_t DynamicMemorySizeInBytes() const { return t_.DynamicMemorySizeInBytes(); }
   std::size_t DumpDynamicMemory(char *dst) const { return t_.DumpDynamicMemory(dst); }
-  std::size_t SetDynamicMemory(char *src) { return t_.SetDynamicMemory(src); }
+  std::size_t SetDynamicMemory(char *src, bool node_root = true) {
+    return t_.SetDynamicMemory(src, node_root);
+  }
 
  private:
   T t_;
