### 2024-08-22 3baf842
spack: use hipcc for +rocm compilation
