### 2024-08-22 93f302c
Redesign with shared mem settings in plac
