+++ singularity-eos/eos/eos_stellar_collapse.hpp
@@ -228,7 +228,8 @@ class StellarCollapse : public EosBase<StellarCollapse> {
 
   std::size_t DynamicMemorySizeInBytes() const;
   std::size_t DumpDynamicMemory(char *dst) const;
-  std::size_t SetDynamicMemory(char *src, bool node_root = true);
+  std::size_t SetDynamicMemory(char *src,
+                               const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS);
 
  private:
   class LogT {
@@ -484,7 +485,9 @@ inline std::size_t StellarCollapse::DumpDynamicMemory(char *dst) const {
   return table_utils::SpinerTricks<StellarCollapse>::DumpDynamicMemory(dst, this);
 }
 
-inline std::size_t StellarCollapse::SetDynamicMemory(char *src, bool node_root) {
+inline std::size_t StellarCollapse::SetDynamicMemory(char *src,
+                                                     const SharedMemSettings &stngs) {
+  if (stngs.data != nullptr) src = stngs.data;
   return table_utils::SpinerTricks<StellarCollapse>::SetDynamicMemory(src, this);
 }
 
