+++ singularity-eos/eos/modifiers/ramps_eos.hpp
@@ -457,8 +457,9 @@ class BilinearRampEOS : public EosBase<BilinearRampEOS<T>> {
 
   std::size_t DynamicMemorySizeInBytes() const { return t_.DynamicMemorySizeInBytes(); }
   std::size_t DumpDynamicMemory(char *dst) const { return t_.DumpDynamicMemory(dst); }
-  std::size_t SetDynamicMemory(char *src, bool node_root = true) {
-    return t_.SetDynamicMemory(src, node_root);
+  std::size_t SetDynamicMemory(char *src,
+                               const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
+    return t_.SetDynamicMemory(src, stngs);
   }
 
  private:
