### 2024-08-22 d187609
All ingredients for a strawman kinetic phasetransition framework implemented.
