### 2024-08-22 1fbf448
completed testing shared memory routines for stellar collapse
