+++ singularity-eos/eos/modifiers/shifted_eos.hpp
@@ -361,6 +361,7 @@ class ShiftedEOS : public EosBase<ShiftedEOS<T>> {
                                const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
     return t_.SetDynamicMemory(src, stngs);
   }
+  constexpr bool StaticMemoryIsThis() const { return t_.StaticMemoryIsThis(); }
 
  private:
   T t_;
