+++ singularity-eos/eos/modifiers/ramps_eos.hpp
@@ -461,6 +461,7 @@ class BilinearRampEOS : public EosBase<BilinearRampEOS<T>> {
                                const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
     return t_.SetDynamicMemory(src, stngs);
   }
+  constexpr bool StaticMemoryIsThis() const { return t_.StaticMemoryIsThis(); }
 
  private:
   T t_;
