+++ singularity-eos/eos/modifiers/scaled_eos.hpp
@@ -351,6 +351,7 @@ class ScaledEOS : public EosBase<ScaledEOS<T>> {
                                const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
     return t_.SetDynamicMemory(src, stngs);
   }
+  constexpr bool StaticMemoryIsThis() const { return t_.StaticMemoryIsThis(); }
 
  private:
   T t_;
