+++ singularity-eos/eos/modifiers/eos_unitsystem.hpp
@@ -434,6 +434,7 @@ class UnitSystem : public EosBase<UnitSystem<T>> {
                                const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
     return t_.SetDynamicMemory(src, stngs);
   }
+  constexpr bool StaticMemoryIsThis() const { return t_.StaticMemoryIsThis(); }
 
  private:
   T t_;
