### 2024-08-23 0e302d8
change spinereosdependsrhosie to work like dependsrhot
