+++ singularity-eos/eos/eos_spiner.hpp
@@ -339,24 +339,15 @@ class SpinerEOSDependsRhoT : public EosBase<SpinerEOSDependsRhoT> {
   mitigated by Ye and (1-Ye) to control how important each term is.
  */
 class SpinerEOSDependsRhoSie : public EosBase<SpinerEOSDependsRhoSie> {
+  friend class table_utils::SpinerTricks<SpinerEOSDependsRhoSie>;
+
  public:
   using Grid_t = SpinerEOSDependsRhoT::Grid_t;
   using DataBox = SpinerEOSDependsRhoT::DataBox;
   struct SP5Tables {
     DataBox P, bMod, dPdRho, dPdE, dTdRho, dTdE, dEdRho;
-
-#define DBLIST &P, &bMod, &dPdRho, &dPdE, &dTdRho, &dTdE, &dEdRho
-    std::vector<const DataBox *> GetDataBoxPointers_() const {
-      return std::vector<const DataBox *>{DBLIST};
-    }
-    std::vector<DataBox *> GetDataBoxPointers_() {
-      return std::vector<DataBox *>{DBLIST};
-    }
-#undef DBLIST
-    DataStatus memoryStatus_ = DataStatus::Deallocated;
-    ;
   };
-  using STricks = table_utils::SpinerTricks<SP5Tables>;
+  using STricks = table_utils::SpinerTricks<SpinerEOSDependsRhoSie>;
 
   SG_ADD_BASE_CLASS_USINGS(SpinerEOSDependsRhoSie);
   PORTABLE_INLINE_FUNCTION SpinerEOSDependsRhoSie()
@@ -530,6 +521,18 @@ class SpinerEOSDependsRhoSie : public EosBase<SpinerEOSDependsRhoSie> {
   DataBox PlRhoMax_, dPdRhoMax_;
 
   Real lRhoOffset_, lTOffset_, lEOffset_; // offsets must be non-negative
+
+#define DBLIST                                                                           \
+  &sie_, &T_, &(dependsRhoT_.P), &(dependsRhoT_.bMod), &(dependsRhoT_.dPdRho),           \
+      &(dependsRhoT_.dPdE), &(dependsRhoT_.dTdRho), &(dependsRhoT_.dTdE),                \
+      &(dependsRhoT_.dEdRho), &(dependsRhoSie_.P), &(dependsRhoSie_.bMod),               \
+      &(dependsRhoSie_.dPdRho), &(dependsRhoSie_.dPdE), &(dependsRhoSie_.dTdRho),        \
+      &(dependsRhoSie_.dTdE), &(dependsRhoSie_.dEdRho), &PlRhoMax_, &dPdRhoMax_
+  std::vector<const DataBox *> GetDataBoxPointers_() const {
+    return std::vector<const DataBox *>{DBLIST};
+  }
+  std::vector<DataBox *> GetDataBoxPointers_() { return std::vector<DataBox *>{DBLIST}; }
+#undef DBLIST
 
   static constexpr unsigned long _preferred_input =
       thermalqs::density | thermalqs::temperature;
@@ -1623,75 +1626,23 @@ inline void SpinerEOSDependsRhoSie::calcBMod_(SP5Tables &tables) {
 }
 
 inline SpinerEOSDependsRhoSie SpinerEOSDependsRhoSie::GetOnDevice() {
-  SpinerEOSDependsRhoSie other = *this; // static memory
-  // dynamic memory
-  other.sie_ = sie_.getOnDevice();
-  other.T_ = T_.getOnDevice();
-  other.PlRhoMax_ = PlRhoMax_.getOnDevice();
-  other.dPdRhoMax_ = dPdRhoMax_.getOnDevice();
[…]
