### 2024-08-23 528294c
add test for tabulated EOS... now to make sure it works
