+++ test/test_eos_tabulated.cpp
@@ -21,6 +21,7 @@
 #include <ports-of-call/portability.hpp>
 #include <ports-of-call/portable_arrays.hpp>
 #include <ports-of-call/portable_errors.hpp>
+#include <singularity-eos/base/constants.hpp>
 #include <singularity-eos/base/variadic_utils.hpp>
 #include <singularity-eos/eos/eos.hpp>
 
@@ -32,6 +33,7 @@
 #include <test/eos_unit_test_helpers.hpp>
 
 #ifdef SPINER_USE_HDF
+#include <singularity-eos/base/spiner_table_utils.hpp>
 using singularity::SpinerEOSDependsRhoSie;
 using singularity::SpinerEOSDependsRhoT;
 #endif
@@ -62,7 +64,7 @@ constexpr Real ev2k = 1.160451812e4;
 #ifdef SINGULARITY_USE_EOSPAC
 using EOS = singularity::Variant<SpinerEOSDependsRhoSie, SpinerEOSDependsRhoT, EOSPAC>;
 
-SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS],[DependsRhoT][EOSPAC]") {
+SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS][DependsRhoT][EOSPAC]") {
 
   GIVEN("SpinerEOS and EOSPAC EOS for steel can be initialized with matid") {
     EOS steelEOS_host_polymorphic = SpinerEOSDependsRhoT(eosName, steelID);
@@ -124,6 +126,7 @@ SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS],[DependsRhoT][EOSPAC]") 
     steelEOS_host_polymorphic.Finalize(); // host and device must be
                                           // finalized separately.
     steelEOS.Finalize();                  // cleans up memory on device.
+    eospac.Finalize();
   }
 
   GIVEN("SpinerEOS and EOSPAC for air can be initialized with matid") {
@@ -179,6 +182,7 @@ SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS],[DependsRhoT][EOSPAC]") 
     }
     airEOS_host.Finalize();
     airEOS.Finalize();
+    eospac.Finalize();
   }
 
   GIVEN("EOS initialized with matid") {
@@ -199,6 +203,7 @@ SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS],[DependsRhoT][EOSPAC]") 
       REQUIRE(isClose(cv, cv_pac));
     }
     eos_spiner.Finalize();
+    eos_eospac.Finalize();
   }
 
   GIVEN("EOS initialized with matid") {
@@ -216,11 +221,11 @@ SCENARIO("SpinerEOS depends on Rho and T", "[SpinerEOS],[DependsRhoT][EOSPAC]") 
       REQUIRE(isClose(rho, rho_pac));
     }
     eos_spiner.Finalize();
+    eos_eospac.FInalize();
   }
 }
 
-// Disabling these tests for now as the DependsRhoSie code is not well-maintained
-SCENARIO("SpinerEOS depends on rho and sie", "[SpinerEOS],[DependsRhoSie]") {
+SCENARIO("SpinerEOS depends on rho and sie", "[SpinerEOS][DependsRhoSie]") {
 
   GIVEN("SpinerEOSes for steel can be initialised with matid") {
     SpinerEOSDependsRhoSie steelEOS_host(eosName, steelID);
@@ -257,6 +262,98 @@ SCENARIO("SpinerEOS depends on rho and sie", "[SpinerEOS],[DependsRhoSie]") {
     steelEOS.Finalize();      // cleans up device memory
   }
 }
+
+SCENARIO("SpinerEOS and EOSPA Serialization",
[…]
