+++ singularity-eos/eos/modifiers/shifted_eos.hpp
@@ -356,7 +356,7 @@ class ShiftedEOS : public EosBase<ShiftedEOS<T>> {
   inline constexpr T UnmodifyOnce() { return t_; }
 
   std::size_t DynamicMemorySizeInBytes() const { return t_.DynamicMemorySizeInBytes(); }
-  std::size_t DumpDynamicMemory(char *dst) const { return t_.DumpDynamicMemory(dst); }
+  std::size_t DumpDynamicMemory(char *dst) { return t_.DumpDynamicMemory(dst); }
   std::size_t SetDynamicMemory(char *src,
                                const SharedMemSettings &stngs = DEFAULT_SHMEM_STNGS) {
     return t_.SetDynamicMemory(src, stngs);
