### 2024-08-23 2159368
added hidden static size
