### 2024-08-23 bfc8ed5
trying to debug eospac
