### 2024-08-24 d89aec3
Give up and expose shared memory and dynamic memory as different concepts
