+++ singularity-eos/base/constants.hpp
@@ -39,11 +39,11 @@ constexpr Real ATMOSPHERIC_PRESSURE = 1e6;
 
 struct SharedMemSettings {
   SharedMemSettings() = default;
-  SharedMemSettings(char *data_, bool is_root_node_)
-      : data(data_), is_root_node(is_root_node_) {}
-  bool CopyNeeded() const { return (data != nullptr) && is_root_node; }
+  SharedMemSettings(char *data_, bool is_domain_root_)
+      : data(data_), is_domain_root(is_domain_root_) {}
+  bool CopyNeeded() const { return (data != nullptr) && is_domain_root; }
   char *data = nullptr;
-  bool is_root_node = false; // default true or false?
+  bool is_domain_root = false; // default true or false?
 };
 const SharedMemSettings DEFAULT_SHMEM_STNGS = SharedMemSettings();
 
