### 2024-08-24 263bea6
CC, CHANGELOG, and some minor typos
