### 2024-08-26 9d2b76c
Add BulkSerializer
