+++ singularity-eos/eos/eos_base.hpp
@@ -700,7 +700,7 @@ class EosBase {
       offst += pcrtp->DumpDynamicMemory(dst + sizeof(CRTP));
     }
     const std::size_t tot_size = pcrtp->SerializedSizeInBytes();
-    PORTABLE_REQUIRE(offst == tot_size, "Serialization succesful");
+    PORTABLE_REQUIRE(offst == tot_size, "Serialization successful");
     return offst;
   }
   auto Serialize() {
@@ -708,7 +708,7 @@ class EosBase {
     std::size_t size = pcrtp->SerializedSizeInBytes();
     char *dst = (char *)malloc(size);
     std::size_t size_new = Serialize(dst);
-    PORTABLE_REQUIRE(size_new == size, "Serialization succesful");
+    PORTABLE_REQUIRE(size_new == size, "Serialization successful");
     return std::make_pair(size, dst);
   }
   std::size_t DeSerialize(char *src,
@@ -725,7 +725,7 @@ class EosBase {
       offst += pcrtp->SetDynamicMemory(src + offst, stngs);
     }
     const std::size_t tot_size = pcrtp->SerializedSizeInBytes();
-    PORTABLE_REQUIRE(offst == tot_size, "Deserialization succesful");
+    PORTABLE_REQUIRE(offst == tot_size, "Deserialization successful");
     return offst;
   }
 
