### 2024-08-26 e4098c9
ci: add new roci jobs
