### 2024-08-27 58e676e
Set GOLDFILE_URL directly through spack package
