### 2024-08-27 cf37b15
ci: add new build_and_test and download_prereq scripts
