+++ .gitlab/download_prereq.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+SINGULARITY_GOLDFILES_VERSION="goldfiles-1.8.0"
+wget https://github.com/lanl/singularity-eos/releases/download/${SINGULARITY_GOLDFILES_VERSION}/goldfiles.tar.gz
