+++ .gitlab/build_and_test.sh
@@ -1,0 +1,221 @@
+#!/bin/bash
+PROJECT_NAME=singularity-eos
+PROJECT_DEFAULT_BRANCH=main
+PROJECT_GROUP=oss
+BUILD_DIR=${BUILD_DIR:-build}
+SOURCE_DIR=${CI_PROJECT_DIR:-$PWD}
+PROJECT_SPACK_ENV=${PROJECT_SPACK_ENV:-/usr/projects/xcap/spack-env/oss/2024-08-01}
+CI_SPACK_ENV=${CI_SPACK_ENV:-$TMPDIR/$USER-ci-env}
+UNTIL=${UNTIL:-install}
+
+# colors
+COLOR_BLUE='\033[1;34m'
+COLOR_MAGENTA='\033[1;35m'
+COLOR_CYAN='\033[1;36m'
+COLOR_PLAIN='\033[0m'
+
+section() {
+  if [ -z "${CI}" ]; then
+    echo -e "${3:+${COLOR_CYAN}$3${COLOR_PLAIN}}"
+  else
+    echo -e $'\e[0K'"section_$1:$(date +%s):$2"$'\r\e[0K'"${3:+${COLOR_CYAN}$3${COLOR_PLAIN}}"
+  fi
+}
+
+print_usage() {
+  echo "usage: build_and_test [-h] [-u {spack,env,cmake,build,test,install}] system environment"
+  echo
+  echo "After the 'env' phase has run, you will be in an active Spack environment."
+  echo "You can then use the 'activate_build_env' command to enter a sub-shell"
+  echo "with the Spack build environment for your project."
+  echo
+  echo "positional arguments:"
+  echo "  system         name of the system"
+  echo "  environment    name of the Spack environment"
+  echo
+  echo "options:"
+  echo "  -u {spack,env,cmake,build,test,install}, --until {spack,env,cmake,build,test,install}"
+  echo "                 run script until the given phase"
+  echo "  -h, --help     show this help message and exit"
+}
+
+VALID_CMD=false
+if [ $# -eq 2 ] && [[ ! "$1" =~ ^-.*  ]] && [[ ! "$2" =~ ^-.*  ]]; then
+  VALID_CMD=true
+elif [ $# -eq 4 ]; then
+  if [ "$1" = "-u" ] || [ "$1" = "--until" ]; then
+    if [[ "$2" =~ ^(spack|env|cmake|build|test|install)$ ]]; then
+      UNTIL=$2
+      VALID_CMD=true
+      shift 2
+    fi
+  fi
+fi
+
+if ! $VALID_CMD; then
+    print_usage
+    return
+fi
+
+SYSTEM_NAME=$1
+SPACK_ENV_NAME=$2
+
+###############################################################################
+# Project-specific steps
+###############################################################################
+
+#------------------------------------------------------------------------------
+# Configure dependencies that were cloned as submodules
+#------------------------------------------------------------------------------
+setup_submodule_dependencies() {
+  if [ "$(ls -A ${SOURCE_DIR}/utils/spiner)" ]; then
+    spack develop --no-clone -p ${SOURCE_DIR}/utils/spiner spiner@main
+  fi
+
+  if [ "$(ls -A ${SOURCE_DIR}/utils/ports-of-call)" ]; then
+    spack develop --no-clone -p ${SOURCE_DIR}/utils/ports-of-call ports-of-call@main
+  fi
+}
+
+#------------------------------------------------------------------------------
+# Hooks for before and after ctest
+#------------------------------------------------------------------------------
+pre_test_steps() {
+  (
+  source ${BUILD_ENV}
+  pushd ${BUILD_DIR}
[…]
