### 2024-08-27 9f24343
Update doc/sphinx/src/using-eos.rst
Co-authored-by: Jeff Peterson <83598606+jhp-lanl@users.noreply.github.com>
