+++ singularity-eos/eos/eos_gruneisen.hpp
@@ -69,7 +69,8 @@ class Gruneisen : public EosBase<Gruneisen> {
     PORTABLE_ALWAYS_REQUIRE(_rho0 >= 0, "Non-negative reference density required");
     PORTABLE_ALWAYS_REQUIRE(_C0 >= 0, "Non-negative Hugoniot intercept required");
     PORTABLE_ALWAYS_REQUIRE(_Cv >= 0, "Non-negative heat capacity required");
-    PORTABLE_ALWAYS_REQUIRE(_rho_max > _rho0, "Maximum density must be greater than reference");
+    PORTABLE_ALWAYS_REQUIRE(_rho_max > _rho0,
+                            "Maximum density must be greater than reference");
   }
   PORTABLE_INLINE_FUNCTION Real
   MaxStableDensityAtTemperature(const Real temperature) const;
