### 2024-08-27 ee7b62e
extended MPI example a bit
