### 2024-08-27 b7cd6cd
update docs with MPI info
