### 2024-08-27 5d7cb38
actually we can't use views they don't have built in iterators... only in experimental.
