+++ doc/sphinx/src/using-kpt.rst
@@ -1,0 +1,244 @@
+.. _using-kpt:
+
+.. _WillsKPT: https://www.osti.gov/biblio/2202604
+
+.. _WillsParameterstudy: https://www.osti.gov/biblio/1900445
+
+Kinetic Phase Transition Framework
+==================================
+
+A step in adapting to more realistic calculations is to 
+get away from that phase transitions are instantanious and the 
+system is in, at least, local thermal equilibrium (LTE) all the time.
+Allowing for superheated (the material is hotter than the equilibrium state) 
+and supercooled (the material is cooler than the equilibrium state) materials is
+a large step towards more realistic modelling. Gas bubbles "exploding" when a 
+spoon is put into a microwave heated cup of coffee is an example of superheating and
+raindrops turning directly to ice when hitting the asphalt is an example
+of supercooling.
+
+Phase transitions are governed by the Gibbs free energy, :math:`G(P,T)`. The equilibrium phase is the phase with lowest
+Gibbs free energy. If the material is heated or cooled from this equilibrium state and another
+phase emerges to have a lower Gibbs free energy, a phase transition occurs and the system eventually transform
+to this new phase. However, all phase transitions need some type of *nucleation event* to get started.
+For the liquid raindrop cooling down while falling to the ground through very cold air, the hit against the
+asphalt is the nucleation event, and the supercooled liquid in the drop instantaniously transform to the equilibrium ice phase. 
+For the very hot coffee, the spoon insertion is the "nucleation" event when the very hot liquid coffee 
+instantaneously transforms into its equilibrium phase, gas. 
+
+In a solid state phase transition, say between an face-centered cubic (fcc) phase and a body-centered
+cubic (bcc) phase, the mere rearrangement of atoms can take some time, also resulting in a phase transition
+that doesn't go directly from the previously lowest Gibbs free energy phase to the presently 
+lowest Gibbs free energy phase. 
+
+``singularity-eos`` provides the ingredients to build a kinetic phase transition framework in a host code.
+
+A comprehensive guide for implementing KPT into a hydro code is available `here <WillsKPT_>`_.
+
+Mass and volume fractions
+-------------------------
+
+The first ingredient needed in a KPT framework is a quantity describing what phase a material currently
+is in. *Mass fractions*, :math:`\mu_i`, gives a measure of how much of the total mass of the material in a cell, :math:`M`, 
+is in a specific phase
+
+.. math::
+
+ \mu_i = \frac{M_i}{M} \qquad \qquad \sum_i \mu_i = 1
+
+For completeness we also introduce *volume fractions*, :math:`f_i`, that give how much of the volume of 
+the material in a cell, :math:`V`, is in a specific phase
+
+.. math::
+
+ f_i = \frac{V_i}{V} \qquad \qquad \sum_i f_i = 1
+
+With these definitions it is clear that both :math:`0 \leq \mu_i \leq 1` and :math:`0 \leq f_i \leq 1`.
+Mass and volume fractions are discussed more closely in the :ref:`mixed cell closure <massandvolumefractions>` section.
+Note, however, that mass and volume fractions in the KPT framework give how much of each *phase* of *one* material is present,
+while in mixed cells they give how much of one *material* is present. Even though we can use some of the same routines
+for mixed material cells and KPT, implementing KPT into a mixed cell environment need to use separate mass and volume 
+fractions for the mixed cell and within each material in the cell.
+
+Homogenization
+--------------
+
[…]
