### 2024-08-28 d7bcd14
ci: disable rocm job for now
