### 2024-08-28 8fb6a87
ci: add example on how to customize build
