### 2024-08-28 08d68b9
ci: update rules
