### 2024-08-28 87ed790
ci: add more (disabled) jobs
