### 2024-08-28 fc132ab
remove a bunch of warnings
