### 2024-08-29 #6805 @kyleniemeyer
Hi @yes @arosen93 and @dominikusbrian, would any of you be able to review this submission for JOSS? We have one completed review now and just need one more. It is a library for equations of state, focused on portable performance.
