### 2024-08-29 5ca3c22
Minor changes, mostly comments and documentation.
