### 2024-08-29 4dc2390
only run sesame2spiner if it is built
