### 2024-09-04 ae48904
Add docs and update build_and_test.sh
