### 2024-09-05 a8c43bd
bug in deserialize based on when SharedMemorySizeInBytes is available
