### 2024-09-06 1a9f49f
aematts tests now build/run on device. Small issues with cout vs printf and issue that eospac does not run on device easily. Comment out eospac for device builds. Fix printf etc. Speed up builds by limiting what the variant is allowed to be for the tests.
