### 2024-09-09 4383de1
Gibbs Free Energy
