+++ singularity-eos/eos/eos_variant.hpp
@@ -179,6 +179,28 @@ class Variant {
   }
 
   template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION Real GibbsFreeEnergyFromDensityTemperature(
+      const Real rho, const Real T,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+    return mpark::visit(
+        [&rho, &T, &lambda](const auto &eos) {
+          return eos.GibbsFreeEnergyFromDensityTemperature(rho, T, lambda);
+        },
+        eos_);
+  }
+
+  template <typename Indexer_t = Real *>
+  PORTABLE_INLINE_FUNCTION Real GibbsFreeEnergyFromDensityInternalEnergy(
+      const Real rho, const Real sie,
+      Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
+    return mpark::visit(
+        [&rho, &sie, &lambda](const auto &eos) {
+          return eos.GibbsFreeEnergyFromDensityInternalEnergy(rho, sie, lambda);
+        },
+        eos_);
+  }
+
+  template <typename Indexer_t = Real *>
   PORTABLE_INLINE_FUNCTION Real SpecificHeatFromDensityTemperature(
       const Real rho, const Real temperature,
       Indexer_t &&lambda = static_cast<Real *>(nullptr)) const {
@@ -673,106 +695,107 @@ class Variant {
   }
 
   template <typename RealIndexer, typename ConstRealIndexer>
-  inline void
-  GibbsFreeEnergyFromDensityTemperature(ConstRealIndexer &&rhos, ConstRealIndexer &&temperatures,
-                                RealIndexer &&entropies, const int num) const {
+  inline void GibbsFreeEnergyFromDensityTemperature(ConstRealIndexer &&rhos,
+                                                    ConstRealIndexer &&temperatures,
+                                                    RealIndexer &&Gs,
+                                                    const int num) const {
     NullIndexer lambdas{}; // Returns null pointer for every index
-    return GibbsFreeEnergyFromDensityTemperature(std::forward<ConstRealIndexer>(rhos),
-                                         std::forward<ConstRealIndexer>(temperatures),
-                                         std::forward<RealIndexer>(entropies), num,
-                                         lambdas);
+    return GibbsFreeEnergyFromDensityTemperature(
+        std::forward<ConstRealIndexer>(rhos),
+        std::forward<ConstRealIndexer>(temperatures), std::forward<RealIndexer>(Gs), num,
+        lambdas);
   }
 
   template <typename RealIndexer, typename ConstRealIndexer, typename LambdaIndexer>
   inline void GibbsFreeEnergyFromDensityTemperature(ConstRealIndexer &&rhos,
-                                            ConstRealIndexer &&temperatures,
-                                            RealIndexer &&entropies, const int num,
-                                            LambdaIndexer &&lambdas) const {
+                                                    ConstRealIndexer &&temperatures,
+                                                    RealIndexer &&Gs, const int num,
+                                                    LambdaIndexer &&lambdas) const {
     return mpark::visit(
-        [&rhos, &temperatures, &entropies, &num, &lambdas](const auto &eos) {
+        [&rhos, &temperatures, &Gs, &num, &lambdas](const auto &eos) {
           return eos.GibbsFreeEnergyFromDensityTemperature(
               std::forward<ConstRealIndexer>(rhos),
-              std::forward<ConstRealIndexer>(temperatures),
-              std::forward<RealIndexer>(entropies), num,
-              std::forward<LambdaIndexer>(lambdas));
+              std::forward<ConstRealIndexer>(temperatures), std::forward<RealIndexer>(Gs),
[…]
