+++ singularity-eos/eos/eos_eospac.hpp
@@ -472,7 +472,7 @@ class EOSPAC : public EosBase<EOSPAC> {
                                                              lambdas);
   }
 
-  // TODO(JMM): Add performant entropy and Gibbs Free Energy 
+  // TODO(JMM): Add performant entropy and Gibbs Free Energy
   using EosBase<EOSPAC>::FillEos;
   using EosBase<EOSPAC>::EntropyIsNotEnabled;
 
