### 2024-09-10 #6805 @junghans
> @kyleniemeyer looking at the list of JOSS reviewers, harpolea might be appropriate.

Ping @kyleniemeyer
