### 2024-09-10 #6805 @kyleniemeyer
@zingale would you be able to contribute a review of this submission to JOSS? Thanks!
