+++ test/test_eos_ideal.cpp
@@ -32,7 +32,7 @@
 using singularity::IdealGas;
 using EOS = singularity::Variant<IdealGas>;
 
-SCENARIO("Ideal gas entropy", "[IdealGas][Entropy][FreeEnergy]") {
+SCENARIO("Ideal gas entropy", "[IdealGas][Entropy][GibbsFreeEnergy]") {
   GIVEN("Parameters for an ideal gas with entropy reference states") {
     // Create ideal gas EOS ojbect
     constexpr Real Cv = 5.0;
@@ -59,8 +59,10 @@ SCENARIO("Ideal gas entropy", "[IdealGas][Entropy][FreeEnergy]") {
           const Real sie = host_eos.InternalEnergyFromDensityTemperature(rho, T);
           const Real P = host_eos.PressureFromDensityTemperature(rho, T);
           const Real G_true = sie + (P / rho) - T * entropy;
-          const Real G = host_eos.GibbsFreeEnergyFromDensityTemperature(rho, T);
-          CHECK(isClose(G, G_true, 1e-12));
+          const Real GT = host_eos.GibbsFreeEnergyFromDensityTemperature(rho, T);
+          CHECK(isClose(GT, G_true, 1e-12));
+          const Real Gsie = host_eos.GibbsFreeEnergyFromDensityInternalEnergy(rho, sie);
+          CHECK(isClose(Gsie, G_true, 1e-12));
         }
       }
     }
