### 2024-09-11 45b2d31
changes needed to build against newest hdf5 in spack, add test for eospac with serialization. Fix minor errors related to that.
