### 2024-09-23 a03dadb
spack: add changes for ats4
