### 2024-09-24 fe714a0
ci: add ats4 jobs
