+++ .gitlab/build_and_test.sh
@@ -1,12 +1,33 @@
 #!/bin/bash
-PROJECT_NAME=singularity-eos
-PROJECT_DEFAULT_BRANCH=main
-PROJECT_GROUP=oss
-BUILD_DIR=${BUILD_DIR:-build}
-SOURCE_DIR=${CI_PROJECT_DIR:-$PWD}
-PROJECT_SPACK_ENV=${PROJECT_SPACK_ENV:-/usr/projects/xcap/spack-env/oss/current}
+###############################################################################
+export PROJECT_NAME=singularity-eos
+export PROJECT_DEFAULT_BRANCH=main
+export PROJECT_TYPE=oss
+export PROJECT_GROUP=oss
+###############################################################################
+
+export BUILD_DIR=${BUILD_DIR:-build}
+export SOURCE_DIR=${CI_PROJECT_DIR:-$PWD}
+TMPDIR=${TMPDIR:-/tmp/xcap/${PROJECT_GROUP}/${PROJECT_NAME}}
 CI_SPACK_ENV=${CI_SPACK_ENV:-$TMPDIR/$USER-ci-env}
-UNTIL=${UNTIL:-install}
+export CTEST_MODE=${CTEST_MODE:-Continuous}
+SUBMIT_TO_CDASH=${SUBMIT_TO_CDASH:-false}
+BUILD_WITH_CTEST=${BUILD_WITH_CTEST:-${SUBMIT_TO_CDASH}}
+SUBMIT_ON_ERROR=${SUBMIT_ON_ERROR:-${SUBMIT_TO_CDASH}}
+SHOW_HELP_MESSAGE=${SHOW_HELP_MESSAGE:-true}
+
+if ${SUBMIT_TO_CDASH}; then
+  UNTIL=${UNTIL:-submit}
+else
+  UNTIL=${UNTIL:-install}
+fi
+
+if ${SUBMIT_ON_ERROR}; then
+  REPORT_ERRORS="ReportErrors"
+else
+  REPORT_ERRORS=""
+fi
+PROJECT_SPACK_ENV_VERSION=${PROJECT_SPACK_ENV_VERSION:-current}
 
 # colors
 COLOR_BLUE='\033[1;34m'
@@ -47,7 +68,7 @@ print_usage() {
   echo "  environment    name of the Spack environment"
   echo
   echo "options:"
-  echo "  -u {spack,env,cmake,build,test,install}, --until {spack,env,cmake,build,test,install}"
+  echo "  -u {spack,env,cmake,build,test,install,submit}, --until {spack,env,cmake,build,test,install,submit}"
   echo "                 run script until the given phase"
   echo "  -h, --help     show this help message and exit"
 }
@@ -67,15 +88,34 @@ fi
 
 if ! $VALID_CMD; then
     print_usage
+    false
     return
 fi
 
-SYSTEM_NAME=$1
-SPACK_ENV_NAME=$2
+export SYSTEM_NAME=$1
+export SPACK_ENV_NAME=$2
 
-###############################################################################
-# Project-specific steps
-###############################################################################
+if [[ "$SYSTEM_NAME" == "darwin" || "$SYSTEM_NAME" == "rocinante" || "$SYSTEM_NAME" == "venado" ]]; then
+  PROJECT_SPACK_ENV=${PROJECT_SPACK_ENV:-/usr/projects/xcap/spack-env/${PROJECT_TYPE}/${PROJECT_SPACK_ENV_VERSION}}
+elif [[ "$SYSTEM_NAME" == "rzadams" || "${SYSTEM_NAME}" == "rzansel" || "$SYSTEM_NAME" == "rzvernal" ]]; then
[…]
