+++ .gitlab-ci.yml
@@ -1,62 +1,14 @@
-###################
-# Global Settings #
-###################
-
 stages:
   - prepare
   - build_n_test
 
-default:
-  interruptible: true
-
 variables:
-  ENABLED_CLUSTERS: "darwin,rocinante"
+  PROJECT_NAME: singularity-eos
+  ENABLE_CDASH: false
   GIT_SUBMODULE_STRATEGY: normal
 
-.ascgit_job:
-  id_tokens:
-    SITE_ID_TOKEN:
-      aud: https://asc-git.lanl.gov
-
-.darwin_job:
-  rules:
-    - if: $ENABLED_CLUSTERS =~ /darwin/ && $CI_PIPELINE_SOURCE == "merge_request_event"
-    - if: $ENABLED_CLUSTERS =~ /darwin/ && $CI_PIPELINE_SOURCE == "pipeline"
-    - if: $ENABLED_CLUSTERS =~ /darwin/ && $CI_PIPELINE_SOURCE == "schedule"
-    - if: $ENABLED_CLUSTERS =~ /darwin/ && $CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH
-    - if: $ENABLED_CLUSTERS =~ /darwin/ && $CI_COMMIT_TAG
-  variables:
-    CLUSTER: darwin
-    SCHEDULER_PARAMETERS: "-N 1 --qos=debug -p general,skylake-gold,skylake-platinum --constraint=\"(cpu_family:skylake)&ib:edr\""
-  tags:
-    - darwin-slurm-shared
-
-.rocinante_job:
-  rules:
-    - if: $ENABLED_CLUSTERS =~ /rocinante/ && $CI_PIPELINE_SOURCE == "merge_request_event"
-    - if: $ENABLED_CLUSTERS =~ /rocinante/ && $CI_PIPELINE_SOURCE == "pipeline"
-    - if: $ENABLED_CLUSTERS =~ /rocinante/ && $CI_PIPELINE_SOURCE == "schedule"
-    - if: $ENABLED_CLUSTERS =~ /rocinante/ && $CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH
-    - if: $ENABLED_CLUSTERS =~ /rocinante/ && $CI_COMMIT_TAG
-  variables:
-    CLUSTER: rocinante
-    SCHEDULER_PARAMETERS: "-N 1 -A asc-ci -p ci --reservation ci --time=00:45:00"
-  tags:
-    - rocinante
-    - batch
-
-.report_tests:
-  artifacts:
-    expire_in: 2 weeks
-    paths:
-      - ${CI_PROJECT_DIR}/build/tests.xml
-    reports:
-      junit: ${CI_PROJECT_DIR}/build/tests.xml
-
-.build_and_test:
-  stage: build_n_test
-  script:
-    - source .gitlab/build_and_test.sh ${CLUSTER} ${SPACK_ENV_NAME}
+include:
+ - .gitlab/common.yml
 
 
 ########
@@ -65,7 +17,13 @@ variables:
 
 download_prereq:
   stage: prepare
-  extends: [.ascgit_job, .darwin_job]
+  extends: [.ascgit_job]
+  rules:
+    - if: $CI_PIPELINE_SOURCE == "merge_request_event"
+    - if: $CI_PIPELINE_SOURCE == "schedule"
+    - if: $CI_PIPELINE_SOURCE == "web"
+    - if: $CI_PIPELINE_SOURCE == "push"
+  needs: []
   variables:
     GIT_SUBMODULE_STRATEGY: none
   script:
@@ -75,26 +33,26 @@ download_prereq:
       - goldfiles.tar.gz
 
[…]
