### 2024-09-25 a3bef97
Cleanup test PTE structure and remove variant from shared header
