+++ test/test_get_sg_eos.cpp
@@ -17,6 +17,7 @@
 #include <stdlib.h>
 
 #include <ports-of-call/portability.hpp>
+#include <pte_test_3mat_analytic.hpp>
 #include <pte_test_utils.hpp>
 #include <singularity-eos/closure/mixed_cell_models.hpp>
 #include <singularity-eos/eos/eos.hpp>
