### 2024-09-25 c493244
Add minimal tests with and without kokkos to biuld matrix on github
