### 2024-09-25 77a3f5a
ci: use cce envs for ats4
