+++ spack-repo/packages/singularity-eos/package.py
@@ -130,7 +130,7 @@ class SingularityEos(CMakePackage, CudaPackage, ROCmPackage):
     depends_on(
         "mpark-variant",
         patches=patch(
-            "https://raw.githubusercontent.com/lanl/singularity-eos/refs/heads/rberger/ats4/utils/gpu_compatibility.patch",
+            "https://raw.githubusercontent.com/lanl/singularity-eos/refs/heads/main/utils/gpu_compatibility.patch",
             sha256="c803670cbd95f9b97458fb4ef403de30229ec81566a5b8e5ccb75ad9d0b22541"
         ),
         when="+cuda",
@@ -138,7 +138,7 @@ class SingularityEos(CMakePackage, CudaPackage, ROCmPackage):
     depends_on(
         "mpark-variant",
         patches=patch(
-            "https://raw.githubusercontent.com/lanl/singularity-eos/refs/heads/rberger/ats4/utils/gpu_compatibility.patch",
+            "https://raw.githubusercontent.com/lanl/singularity-eos/refs/heads/main/utils/gpu_compatibility.patch",
             sha256="c803670cbd95f9b97458fb4ef403de30229ec81566a5b8e5ccb75ad9d0b22541",
         ),
         when="+rocm",
