### 2024-10-09 5ebda78
further down the road in terms of init and residual. Jacobian is next but need to make sure where to put derivatives
