+++ aemattssing/_sources/src/using-eos.rst.txt
@@ -1,0 +1,1044 @@
+.. _using-eos:
+
+The Equation of State API
+=========================
+
+This page describes the equation of state API in detail. For just the
+information needed to get started, check out the :ref:`getting started
+<getting-started>` page.
+
+The ``Real`` Type
+------------------
+
+``singularity-eos`` defines the ``singularity::Real`` type as a proxy
+for the ``float`` and ``double`` types. We currently resolve ``Real``
+to a double precision number, however we plan to have the option to
+select different precisions at compile time in the future.
+
+The parallelism model
+----------------------
+
+For the most part, ``singularity-eos`` tries to be agnostic to how you
+parallelize your code on-node. It knows nothing at all about
+distributed memory parallelism, with one exception, discussed
+below. An ``EOS`` object can be copied into any parallel code block by
+value (see below) and scalar calls do not attempt any internal
+multi-threading, meaning ``EOS`` objects are not thread-safe, but are
+compatible with thread safety, assuming the user calls them
+appropriately. The main complication is ``lambda`` arrays, which are
+discussed below.
+
+The vector ``EOS`` method overloads are a bit different. These are
+thread-parallel operations launched by ``singularity-EOS``. They run
+in parallel, and ordering between indices of vectors cannot be
+assumed. Therefore care must taken in memory layout to avoid race
+conditions. The type of parallelism used depends on how
+``singularity-eos`` is compiled. If the ``Kokkos`` backend is used,
+any parallel dispatch supported by ``Kokkos`` is supported.
+
+A more generic version of the vector calls exists in the ``Evaluate``
+method, which allows the user to specify arbitrary parallel dispatch
+models by writing their own loops. See the relevant section below.
+
+Serialization and shared memory
+--------------------------------
+
+While ``singularity-eos`` makes a best effort to be agnostic to
+parallelism, it exposes several methods that are useful in a
+distributed memory environment. In particular, there are two use-cases
+the library seeks to support:
+
+#. To avoid stressing a filesystem, it may desirable to load a table from one thread (e.g., MPI rank) and broadcast this data to all other ranks.
+#. To save memory it may be desirable to place tabulated data, which is read-only after it has been loaded from file, into shared memory on a given node, even if all other data is thread local in a distributed-memory environment. This is possible via, e.g., `MPI Windows`_.
+
+Therefore ``singularity-eos`` exposes several methods that can be used
+in this context. The function
+
+.. cpp:function:: std::size_t EOS::SerializedSizeInBytes() const;
+
+returns the amount of memory required in bytes to hold a serialized
+EOS object. The return value will depend on the underlying equation of
+state model currently contained in the object. The function
+
+.. cpp:function:: std::size_t EOS::SharedMemorySizeInBytes() const;
+
+returns the amount of data (in bytes) that a given object can place into shared memory. Again, the return value depends on the model the object currently represents.
+
+.. note::
+
+  Many models may not be able to utilize shared memory at all. This
+  holds for most analytic models, for example. The ``EOSPAC`` backend
+  will only utilize shared memory if the ``EOSPAC`` version is sufficiently recent
+  to support it and if ``singularity-eos`` is built with serialization
+  support for ``EOSPAC`` (enabled with
[…]
