### 2024-10-10 8237aa3
Some progress on jacobian calc. Not there yet, probably doesnt compile
