### 2024-10-15 8dceffc
update PTE solver to report instrumentation and expose parameters
