+++ test/test_eos_modifiers.cpp
@@ -242,116 +242,3 @@ SCENARIO("EOS Builder and Modifiers", "[EOSBuilder][Modifiers][IdealGas]") {
 #endif // SINGULARITY_BUILD_CLOSURE
   }
 }
-
-SCENARIO("Relativistic EOS", "[EOSBuilder][RelativisticEOS][IdealGas]") {
-  GIVEN("Parameters for an ideal gas") {
-    constexpr Real Cv = 2.0;
-    constexpr Real gm1 = 0.5;
-    WHEN("We construct a relativistic IdealGas with EOSBuilder") {
-      constexpr Real cl = 1;
-      EOS eos = IdealGas(gm1, Cv);
-      eos = Modify<RelativisticEOS>(eos, cl);
-      THEN("The EOS has finite sound speeds") {
-        constexpr Real rho = 1e3;
-        constexpr Real sie = 1e3;
-        Real bmod = eos.BulkModulusFromDensityInternalEnergy(rho, sie);
-        Real cs2 = bmod / rho;
-        REQUIRE(cs2 < 1);
-      }
-    }
-  }
-}
-
-SCENARIO("EOS Unit System", "[EOSBuilder][UnitSystem][IdealGas]") {
-  GIVEN("Parameters for an ideal gas") {
-    constexpr Real Cv = 2.0;
-    constexpr Real gm1 = 0.5;
-    GIVEN("Units with a thermal unit system") {
-      constexpr Real rho_unit = 1e1;
-      constexpr Real sie_unit = 1e-1;
-      constexpr Real temp_unit = 123;
-      WHEN("We construct an IdealGas with EOSBuilder") {
-        EOS eos = IdealGas(gm1, Cv);
-        eos = Modify<UnitSystem>(eos, rho_unit, sie_unit, temp_unit);
-        THEN("Units cancel out for an ideal gas") {
-          Real rho = 1e3;
-          Real sie = 1e3;
-          Real P = eos.PressureFromDensityInternalEnergy(rho, sie);
-          Real Ptrue = gm1 * rho * sie;
-          REQUIRE(std::abs(P - Ptrue) / Ptrue < 1e-3);
-        }
-      }
-    }
-    GIVEN("Units with length and time units") {
-      constexpr Real time_unit = 456;
-      constexpr Real length_unit = 1e2;
-      constexpr Real mass_unit = 1e6;
-      constexpr Real temp_unit = 789;
-      WHEN("We construct an IdealGas with EOSBuilder") {
-        EOS eos = IdealGas(gm1, Cv);
-        eos = Modify<UnitSystem>(eos, eos_units_init::length_time_units_init_tag,
-                                 time_unit, mass_unit, length_unit, temp_unit);
-        THEN("Units cancel out for an ideal gas") {
-          constexpr Real rho = 1e3;
-          constexpr Real sie = 1e3;
-          constexpr Real Ptrue = gm1 * rho * sie;
-          Real P = eos.PressureFromDensityInternalEnergy(rho, sie);
-          REQUIRE(std::abs(P - Ptrue) / Ptrue < 1e-3);
-        }
-      }
-    }
-  }
-}
-
-SCENARIO("Serialization of modified EOSs preserves their properties",
-         "[ScaledEOS][IdealGas][Serialization]") {
-  GIVEN("A scaled ideal gas object") {
-    constexpr Real Cv = 2.0;
-    constexpr Real gm1 = 0.5;
-    constexpr Real scale = 2.0;
-
-    constexpr Real rho_test = 1.0;
-    constexpr Real sie_test = 1.0;
-    constexpr Real temp_trivial = sie_test / (Cv);      // = 1./2.
-    constexpr Real temp_test = sie_test / (Cv * scale); // = 1./4.
[…]
