+++ doc/sphinx/src/using-closures.rst
@@ -450,9 +450,11 @@ The constructor for the ``PTESolverRhoT`` is of the form
 .. code-block:: cpp
 
   template <typename EOS_t, typename Real_t, typename Lambda_t>
-  PTESolverRhoT(const int nmat, EOS_t &&eos, const Real vfrac_tot, const Real sie_tot,
+  PORTABLE_INLINE_FUNCTION
+  PTESolverRhoT(const std::size_t nmat, EOS_t &&eos, const Real vfrac_tot, const Real sie_tot,
                 Real_t &&rho, Real_t &&vfrac, Real_t &&sie, Real_t &&temp, Real_t &&press,
-                Lambda_t &&lambda, Real *scratch, const Real Tguess = 0);
+                Lambda_t &&lambda, Real *scratch, const Real Tnorm = 0.0,
+                const MixParams &params = MixParams())
 
 where ``nmat`` is the number of materials, ``eos`` is an indexer over
 equation of state objects, one per material, and ``vfrac_tot`` is a
@@ -466,9 +468,64 @@ one per material. ``press`` is an indexer over pressures, one per
 material. ``lambda`` is an indexer over lambda arrays, one per
 material. ``scratch`` is a pointer to pre-allocated scratch memory, as
 described above. It is assumed enough scratch has been allocated.
-Finally, the optional argument ``Tguess`` allows for host codes to
-pass in an initial temperature guess for the solver.  For more
-information on initial guesses, see the section below.
+The optional argument ``Tnorm`` allows for host codes to pass in a
+normalization for the temperature scale. Initial guesses for density
+and temperature may be passed in through the ``rho`` and ``sie`` input
+parameters.
+
+The optional ``MixParams`` input contains a struct of runtime
+parameters that may be used by the various PTE solvers. This struct
+contains the following member fields, with default values:
+
+.. code-block:: cpp
+
+  struct MixParams {
+    bool verbose = false; // verbosity
+    Real derivative_eps = 3.0e-6;
+    Real pte_rel_tolerance_p = 1.e-6;
+    Real pte_rel_tolerance_e = 1.e-6;
+    Real pte_rel_tolerance_t = 1.e-4;
+    Real pte_abs_tolerance_p = 0.0;
+    Real pte_abs_tolerance_e = 1.e-4;
+    Real pte_abs_tolerance_t = 0.0;
+    Real pte_residual_tolerance = 1.e-8;
+    std::size_t pte_max_iter_per_mat = 128;
+    Real line_search_alpha = 1.e-2;
+    std::size_t line_search_max_iter = 6;
+    Real line_search_fac = 0.5;
+    Real vfrac_safety_fac = 0.95;
+    Real temperature_limit = 1.0e15;
+    Real default_tguess = 300.;
+    Real min_dtde = 1.0e-16;
+  };
+
+where here ``verbose`` specifies how erbose PTE output is,
+``derivative_eps`` is the spacing used for finite differences
+evaluations of equations of state when building a jacobian. The
+``pte_rel_tolerance_p``, ``pte_rel_tolerance_e``, and
+``pte_rel_tolerance_t`` variables are relative tolerances for the
+error in the pressure, energy, temperature respectively. The
+``pte_abs_tolerance_*`` variables are the same but are absolute,
+rather than relative tolerances. ``pte_residual_tolerance`` is the
+absolute tolerance for the residual.
+
+The maximum number of iterations the solver is allowed to take before
+giving up is ``pte_max_iter_per_mat`` multiplied by the number of
+materials used. ``line_search_alpha`` is used as a safety factor in
+the line search. ``line_search_max_iter`` is the maximum number of
+iterations the solver is allowed to take in the line
[…]
