### 2024-10-15 56ea7b0
fix calls to PTESolver in get_sg_eos
