+++ singularity-eos/eos/get_sg_eos_rho_t.cpp
@@ -59,7 +59,8 @@ void get_sg_eos_rho_t(const char *name, int ncell, indirection_v &offsets_v,
               npte, eos_inx, vfrac_sum, temp_pte(tid, 0), &rho_pte(tid, 0),
               &vfrac_pte(tid, 0), &sie_pte(tid, 0), &temp_pte(tid, 0), &press_pte(tid, 0),
               cache, &solver_scratch(tid, 0));
-          pte_converged = PTESolver(method);
+          auto status = PTESolver(method);
+          pte_converged = status.converged;
           // calculate total internal energy
           for (int mp = 0; mp < npte; ++mp) {
             const int m = pte_mats(tid, mp);
