### 2024-10-15 15ae016
Update doc/sphinx/src/using-closures.rst
Co-authored-by: Jeff Peterson <83598606+jhp-lanl@users.noreply.github.com>
