### 2024-10-25 efc1033
Fix residual and number of equations.
