### 2024-10-25 0e32674
More work on check pte, etc. Not quite there yet, should probably create a test soon.
