+++ singularity-eos/closure/mixed_cell_models.hpp
@@ -746,7 +746,7 @@ inline int PTESolverPTRequiredScratch(const int nmat) {
   int neq = 2;
   return neq * neq                 // jacobian
          + 4 * neq                 // dx, residual, and sol_scratch
-         + 3 * nmat                // all the nmat sized arrays
+         + 2 * nmat                // all the nmat sized arrays
          + MAX_NUM_LAMBDAS * nmat; // the cache
 }
 inline size_t PTESolverPTRequiredScratchInBytes(const int nmat) {
@@ -791,7 +791,6 @@ class PTESolverPT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
       : mix_impl::PTESolverBase<EOSIndexer, RealIndexer>(nmat, 2, eos, vfrac_tot,
                                                          sie_tot, rho, vfrac, sie, temp,
                                                          press, scratch, Tguess) {
-    vtemp = AssignIncrement(scratch, nmat);
     // TODO(JCD): use whatever lambdas are passed in
     /*for (int m = 0; m < nmat; m++) {
       if (!variadic_utils::is_nullptr(lambda[m])) Cache[m] = lambda[m];
@@ -836,19 +835,13 @@ class PTESolverPT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
   PORTABLE_INLINE_FUNCTION
   bool CheckPTE() const {
     using namespace mix_params;
-    Real mean_p = vfrac[0] * press[0];
-    Real error_p = 0.0;
-    for (int m = 1; m < nmat; ++m) {
-      mean_p += vfrac[m] * press[m];
-      error_p += residual[m + 1] * residual[m + 1];
-    }
-    error_p = std::sqrt(error_p);
+    Real error_v = std::abs(residual[0]);
     Real error_u = std::abs(residual[1]);
+    // this may not be quite right as we may need an energy scaling factor
     // Check for convergence
-    bool converged_p = (error_p < pte_rel_tolerance_p * std::abs(mean_p) ||
-                        error_p < pte_abs_tolerance_p);
     bool converged_u = (error_u < pte_rel_tolerance_e || error_u < pte_abs_tolerance_e);
-    return converged_p && converged_u;
+    bool converged_v = (error_v < pte_rel_tolerance_e || error_v < pte_abs_tolerance_e);
+    return converged_v && converged_u;
   }
 
   PORTABLE_INLINE_FUNCTION
@@ -896,8 +889,6 @@ class PTESolverPT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     }
 
     // Fill in the Jacobian
-    for (int i = 0; i < neq * neq; ++i)
-      jacobian[i] = 0.0;
     jacobian[0] = rtor2_dr_dT_p_sum;
     jacobian[1] = rtor2_dr_dP_T_sum;
     jacobian[2] = dedT_P_sum;
@@ -909,32 +900,9 @@ class PTESolverPT : public mix_impl::PTESolverBase<EOSIndexer, RealIndexer> {
     using namespace mix_params;
     // Each check reduces the scale further if necessary
     Real scale = 1.0;
-    // control how big of a step toward vfrac = 0 is allowed
-    for (int m = 0; m < nmat; ++m) {
-      if (scale * dx[m] < -vfrac_safety_fac * vfrac[m]) {
-        scale = -vfrac_safety_fac * robust::ratio(vfrac[m], dx[m]);
-      }
-    }
-    const Real Tnew = Tequil + scale * dx[nmat];
-    // control how big of a step toward rho = rho(Pmin) is allowed
[…]
