### 2024-10-28 lanl/singularity-eos#427 @guadabsb15
JOSS review comments
Hello, these are some starting comments for review https://github.com/openjournals/joss-reviews/issues/6805

For what I have seen so far, singularity-eos looks very helpful and comprehensive.

Looking at your Getting Started site in the documentation, I think it could be useful to have also `-DSINGULARITY_BUILD_EXAMPLES=ON` and then say something like "as a quick check of your installation go to /build/example and run `./get_sound_speed_press` you should be able to see the following output: "
```
The final values are:
rho = 0.987467
uu  = 0.00987467
P   = 0.0059248
cs  = 0.0979796
```
I personally like having a sanity check when installing/using a new tool. And as a reviewer, is this the result I should be getting? I must admit I didn't compute the EOS values myself. I believe it is valuable to have the expected output of the example.

I'm not an expert in CMake, but I think it might look more consistent if the code snipets for both Getting Started and  Examples used `mkdir -p build && cd build`,  and not `build` for Example and `bin` for the Getting started?

For the Python Bindings documentation, it might be worth it to mention to add the `singularity-eos\build\python` to the `$PYTHONPATH` (or the path were users install singularity-eos /lib/python.../site-packages)

The EOS Models page doesn't show several of the Figures, is that on purpose?

I still need to check the written paper and a few more things, but these are my initial thoughts.
