### 2024-10-30 0066727
Changes from round of JOSS review: - Add discussion of sanity checks to getting started - Make build/bin more consistent in documentation - Add comment about python path - Make sure figures display in all browsers by making png versions
