### 2024-10-30 lanl/singularity-eos#427 @Yurlungur
> I personally like having a sanity check when installing/using a new tool. And as a reviewer, is this the result I should be getting? I must admit I didn't compute the EOS values myself. I believe it is valuable to have the expected output of the example.

I added the examples as a sanity check. But I think a more quantitative sanity check is the unit tests, so I added those to getting started as well.

> The EOS Models page doesn't show several of the Figures, is that on purpose?

Not on purpose. It appears that whether or not the pdfs display properly is browser dependent. I changed the figures to pngs which should be supported by all browsers.

> I'm not an expert in CMake, but I think it might look more consistent if the code snipets for both Getting Started and Examples used mkdir -p build && cd build, and not build for Example and bin for the Getting started?

Good point. I tried to make this more consistent.

> For the Python Bindings documentation, it might be worth it to mention to add the singularity-eos\build\python to the $PYTHONPATH (or the path were users install singularity-eos /lib/python.../site-packages)

Good point. I added some verbage to this effect.
