### 2024-10-30 80c596f
Update doc/sphinx/src/getting-started.rst
Co-authored-by: Christoph Junghans <christoph.junghans@gmail.com>
