### 2024-10-30 lanl/singularity-eos#427 @guadabsb15
Thank you for addressing my feedback, I will be taking a look at your PR hopefully tomorrow.

I read the software paper and it reads really well, I only have a question about one reference. In the performance portable polymorphism section, the citation for OpenMP Target Offload is from 2001. I could be wrong, but I thought OpenMP 4.0 was the first version that started supporting accelerators offload, and that was released in 2013? Or are you citing OpenMP in general?
