### 2024-10-30 lanl/singularity-eos#427 @Yurlungur
> Thank you for addressing my feedback, I will be taking a look at your PR hopefully tomorrow.

Thanks @guadabsb15 ! Sounds good.

> I read the software paper and it reads really well, I only have a question about one reference. In the performance portable polymorphism section, the citation for OpenMP Target Offload is from 2001. I could be wrong, but I thought OpenMP 4.0 was the first version that started supporting accelerators offload, and that was released in 2013? Or are you citing OpenMP in general?

Yes---the 2001 reference is the recommended citation for OpenMP as a whole. I will modify the citation location to clarify this. I'll also stick in a reference for OpenMP 4.0. I'll just sneak these into MR #428.
