+++ joss-paper/paper.md
@@ -217,7 +217,7 @@ knowledge to compute them as performantly as possible.
 
 Accelerators present new challenges to standard object-oriented
 programming. In particular, not all compiler stacks (such as Sycl
-[@SYCL] or OpenMP Target Offload [@chandra2001parallel])
+[@SYCL] or OpenMP [@chandra2001parallel] Target Offload [@openmp4])
 support relocatable device code, which is required for standard C++
 polymorphism. Even in programming models, such as CUDA [@cuda],
 which do support relocatable device code, polymorphism can be slower
