### 2024-10-31 lanl/singularity-eos#427 @guadabsb15
Hi! I took a look at PR #428  and it looks good. However, for the quickstart I think the unit tests are a bit too involved/in depth. Just the example, to check the installation worked, like you put it is great.
