### 2024-10-31 1e88aaa
build -> builddir as easier to distinguish names. Remove the unit tests from getting started.
