### 2024-10-31 lanl/singularity-eos#427 @Yurlungur
> for the quickstart I think the unit tests are a bit too involved/in depth. Just the example, to check the installation worked, like you put it is great.

Fair enough! I removed the reference to the unit tests. Thanks @guadabsb15 . Should I go ahead and merge that MR and close this issue?
