### 2024-11-04 lanl/singularity-eos#427 @guadabsb15
@Yurlungur sounds good! The MR looks great and takes care of the comments in this issue.
