### 2024-11-04 #6805 @kyleniemeyer
@Yurlungur In the paper, I think you just need to add a `# References` section header at the end of the paper.
