### 2024-11-05 #6805 @Yurlungur
> @Yurlungur In the paper, I think you just need to add a `# References` section header at the end of the paper.

Ah I see. Thanks @kyleniemeyer ! Should be fixed now.
