### 2024-11-07 #6805 @kyleniemeyer
@Yurlungur please replace the URL with the correct DOI for the O'Connor and Ott reference (10.1088/0264-9381/27/11/114103), and it appears the EOSPAC User's Manual also has a DOI (10.2172/1765849)
