+++ joss-paper/refs.bib
@@ -50,7 +50,7 @@ year = "1992"}
   volume={27},
   number={11},
   pages={114103},
-  url={http://stacks.iop.org/0264-9381/27/i=11/a=114103},
+  doi={10.1088/0264-9381/27/11/114103},
   year={2010},
 }
 
@@ -69,6 +69,7 @@ language = {eng},
 address = {United States},
 author = {Pimentel, David A},
 keywords = {EOSPAC SESAME Equation of State Interpolation Thermodynamics ; MATHEMATICS AND COMPUTING},
+doi = {10.2172/1765849},
 organization = {Los Alamos National Lab. (LANL), Los Alamos, NM (United States)},
 }
 
